### 2025-09-04 #8930 @editorialbot
[PRE REVIEW]: tqec: A Python package for topological quantum error correction
**Submitting author:** @BSchelpe (<a href="http://orcid.org/0009-0003-9046-3104">Arabella Schelpe</a>)
**Repository:** https://github.com/tqec/tqec
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.1
