### 2025-09-04 #8930 @danielskatz
👋 @BSchelpe - thanks for your submission.

However, I see what I think is a major problem. While your software has an Apache-2 license, it also has a legal notice that contradicts that license. My point is not about either one of these documents or making any statement about the values the project holds, but that these documents contradict each other, and JOSS requires an OSI-approved license, which in my opinion, is not true for this project when the legal notice that you have given is also included.

Also, as a very minor point in comparison, the text in LICENSE that follows "END OF TERMS AND CONDITIONS" should be removed, as it's not part of the license, but rather instructions about how to use the license.

Finally, if we can get the license issue resolved, you should add countries to all affiliations in your .md file, and also, work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to make a new PDF. editorialbot commands need to be the first entry in a new comment. However, there's no need to do this until we resolve the license issue.
