### 2025-09-04 #8930 @BSchelpe
Hi! Thank you very much for looking at the paper so quickly. I will speak to the others about the license questions and hopefully come back to you soon.
