### 2025-09-04 2d589b6
Change url to doi and add countries
