+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{polian:2015,
-  	doi = {https://doi.org/10.1145/2744769.2747921},
+  	doi = {10.1145/2744769.2747921},
   	Author = {{Polian}, Ilia and {Fowler}, Austin G.},
   	Journal = {52nd ACM/EDAC/IEEE Design Automation Conference (DAC), 1–6},
   	Title = {{Design automation challenges for scalable quantum architectures}},
