+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{polian:2015,
-  	doi = {https://doi.org/10.1145/2744769.2747921},
+  	doi = {10.1145/2744769.2747921},
   	Author = {{Polian}, Ilia and {Fowler}, Austin G.},
   	Journal = {52nd ACM/EDAC/IEEE Design Automation Conference (DAC), 1–6},
   	Title = {{Design automation challenges for scalable quantum architectures}},
@@ -14,7 +14,7 @@
     Volume = 86,
     Number = 3,
     Pages = 32324,
-    doi = {https://doi.org/10.1103/PhysRevA.86.032324},
+    doi = {10.1103/PhysRevA.86.032324},
     month = September
     }
 
@@ -26,7 +26,7 @@
      Volume = 7,
      Month = November,
      Number = 1172,
-     doi = {https://doi.org/10.22331/q-2023-11-07-1172}
+     doi = {10.22331/q-2023-11-07-1172}
      }
 
 @article{gidney:2025,
@@ -36,7 +36,7 @@
      Volume = 16,
      Number = 4498,
      Year = 2025,
-     doi = {https://doi.org/10.1038/s41467-025-59714-1}
+     doi = {10.1038/s41467-025-59714-1}
      }
 
 @inproceedings{kissinger:2020,
@@ -47,7 +47,7 @@
      Address = {Chapman University, Orange, CA, USA},
      Series = {Electronic Proceedings in Theoretical Computer Science},
      Publisher = {Open Publishing Association},
-     doi = {https://doi.org/10.4204/EPTCS.318.14}
+     doi = {10.4204/EPTCS.318.14}
      }
 
 @article{gidney:2012,
@@ -58,7 +58,7 @@
      Volume = 5,
      Month = July,
      Number = 497,
-     doi = {https://doi.org/10.22331/q-2021-07-06-497}
+     doi = {10.22331/q-2021-07-06-497}
      }
 
 @inproceedings{kan:2025,
