+++ .github/workflows/ci.yml
@@ -21,9 +21,9 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: "3.13"
-          cache: "pip"
+          cache: "uv"
       - name: Install dependencies
-        run: python -m pip install --upgrade pip pre-commit
+        run: uv pip install --upgrade pip pre-commit
       - name: Run pre-commit hooks
         run: pre-commit run --all-files
 
@@ -40,25 +40,25 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-          cache: "pip"
+          cache: "uv"
       - name: Install dependencies
         run: |
-          python -m pip install --upgrade pip
-          python -m pip install -r requirements.txt -r requirements-dev.txt
-          python -m pip install pytest mypy
+          uv pip install --upgrade pip
+          uv pip install ".[all]"
+          uv pip install pytest mypy
       - name: Build
-        run: python -m pip install -e .
+        run: uv pip install -e .
       # Type check the code base.
       - name: Mypy type checking
         run: mypy src/tqec/
       # Perform the full tests (including slow ones) when pushing on main.
       - name: Complete tests
         if: github.ref == 'refs/heads/main'
-        run: python -m pytest $(git ls-files '*_test.py')
+        run: uv pytest $(git ls-files '*_test.py')
       # Only perform quick tests when pushing on a PR.
       - name: Partial tests
         if: github.ref != 'refs/heads/main'
-        run: python -m pytest $(git ls-files '*_test.py') -m "not slow"
+        run: uv pytest $(git ls-files '*_test.py') -m "not slow"
 
   coverage-python:
     runs-on: ubuntu-latest
@@ -72,12 +72,12 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: "3.13"
-          cache: "pip"
+          cache: "uv"
       - name: Install dependencies and tqec package
         run: |
-          python -m pip install --upgrade pip
-          python -m pip install -e '.[test]'
-          python -m pip install pytest pytest-cov
+          uv pip install --upgrade pip
+          uv pip install -e '.[test]'
+          uv pip install pytest pytest-cov
       # Only perform quick tests when pushing on a PR
       - name: Code health - format
         uses: astral-sh/ruff-action@v3
@@ -88,7 +88,7 @@ jobs:
       - name: Code coverage for partial tests
         if: github.ref != 'refs/heads/main'
         run: |
-          python -m pytest --cov=tqec --cov-report=xml:coverage.xml --cov-report term-missing:skip-covered $(git ls-files '*_test.py') -m "not slow"
+          uv pytest --cov=tqec --cov-report=xml:coverage.xml --cov-report term-missing:skip-covered $(git ls-files '*_test.py') -m "not slow"
       # Print code coverage if in PR
       - name: Code Coverage Report
         uses: irongut/CodeCoverageSummary@v1.3.0
