### 2025-09-08 88ada39
make sure uv is installed
