+++ .github/workflows/ci.yml
@@ -19,6 +19,8 @@ jobs:
       - uses: actions/checkout@v4
       - name: Set up Python
         uses: actions/setup-python@v5
+      - name: Install uv
+        uses: astral-sh/setup-uv@v6
       - name: Install dependencies
         run: uv pip install --upgrade pip pre-commit
       - name: Run pre-commit hooks
@@ -35,6 +37,8 @@ jobs:
       - uses: actions/checkout@v4
       - name: Set up Python
         uses: actions/setup-python@v5
+      - name: Install uv
+        uses: astral-sh/setup-uv@v6
       - name: Install dependencies
         run: |
           uv pip install --upgrade pip
@@ -64,6 +68,8 @@ jobs:
       - uses: actions/checkout@v4
       - name: Set up Python
         uses: actions/setup-python@v5
+      - name: Install uv
+        uses: astral-sh/setup-uv@v6
       - name: Install dependencies and tqec package
         run: |
           uv pip install --upgrade pip
