+++ .github/workflows/ci.yml
@@ -22,7 +22,7 @@ jobs:
       - name: Install uv
         uses: astral-sh/setup-uv@v6
       - name: Install dependencies
-        run: uv pip install --system --upgrade pip pre-commit
+        run: uv pip install --upgrade pip pre-commit
       - name: Run pre-commit hooks
         run: pre-commit run --all-files
 
