### 2025-09-08 d9e54e9
try adding uv venv to ci.yml
