### 2025-09-08 564b55b
cleanup installation commands in ci.yml
