+++ .github/workflows/gh-pages.yml
@@ -20,18 +20,17 @@ jobs:
       group: ${{ github.workflow }}-${{ github.ref }}
     steps:
       - uses: actions/checkout@v4
+      - name: Install uv
+        uses: astral-sh/setup-uv@v6
       - name: Set up Python
         uses: actions/setup-python@v5
-        with:
-          python-version: "3.13"
-          cache: "pip"
       - name: Install dependencies
         run: |
+          uv venv
           sudo apt-get update
           sudo apt-get install -y --no-install-recommends pandoc
-          python -m pip install --upgrade pip
-          python -m pip install -r requirements.txt -r requirements-dev.txt
-          python -m pip install '.[all]'
+          uv pip install --upgrade pip
+          uv pip install -e '.[all]'
       - uses: actions/cache@v4
         id: cache-docs
         with:
@@ -42,7 +41,7 @@ jobs:
           restore-keys: |
             docs_database_and_stim_results_cache
       - name: Generate the documentation
-        run: make html
+        run: uv run make html
         working-directory: docs/
         # working-directory ensures the build relies on cached statistics and database in docs/_examples_database/
         # this is used to reduce the docs build time in the CI.
