### 2025-09-08 b318d50
make sphinx-copybutton installation explicit
