+++ .github/workflows/gh-pages.yml
@@ -31,6 +31,7 @@ jobs:
           sudo apt-get install -y --no-install-recommends pandoc
           uv pip install --upgrade pip
           uv pip install -e '.[all]'
+          uv pip install sphinx-copybutton
       - uses: actions/cache@v4
         id: cache-docs
         with:
