### 2025-09-09 11d5ea4
try: remove explicit installation of sphinx-copybutton and force make html in virtual environment
