+++ .github/workflows/gh-pages.yml
@@ -30,6 +30,7 @@ jobs:
           sudo apt-get update
           sudo apt-get install -y --no-install-recommends pandoc
           uv pip install --upgrade pip
+          uv lock
           uv pip install -e '.[all]'
       - uses: actions/cache@v4
         id: cache-docs
