### 2025-09-09 c719f64
activate virtual environment in main directory before make html
