+++ .github/workflows/gh-pages.yml
@@ -43,7 +43,8 @@ jobs:
             docs_database_and_stim_results_cache
       - name: Generate the documentation
         run: |
-          uv run --frozen make html
+          source ../.venv/bin/activate
+          make html
         working-directory: docs/
         # working-directory ensures the build relies on cached statistics and database in docs/_examples_database/
         # this is used to reduce the docs build time in the CI.
