### 2025-09-09 d382424
try: use uv run without virtual environment
