+++ .github/workflows/gh-pages.yml
@@ -30,7 +30,6 @@ jobs:
           sudo apt-get update
           sudo apt-get install -y --no-install-recommends pandoc
           uv pip install --upgrade pip
-          uv lock
           uv pip install -e '.[all]'
       - uses: actions/cache@v4
         id: cache-docs
@@ -43,8 +42,7 @@ jobs:
             docs_database_and_stim_results_cache
       - name: Generate the documentation
         run: |
-          source .venv/bin/activate
-          make html
+          uv run --extra all make html
         working-directory: docs/
         # working-directory ensures the build relies on cached statistics and database in docs/_examples_database/
         # this is used to reduce the docs build time in the CI.
