### 2025-09-09 1f13a73
update readme and contributor guide
