+++ README.md
@@ -35,10 +35,10 @@ Documentation is available at <https://tqec.github.io/tqec/index.html>
 
 ## Installation
 
-Currently, you need to install `tqec` from source:
+Currently, you need to install `tqec` from source with [`uv`](https://docs.astral.sh/uv/):
 
 ```sh
-python -m pip install git+https://github.com/tqec/tqec.git
+uv pip install git+https://github.com/tqec/tqec.git
 ```
 
 For a more detailed installation guide and common troubleshooting tips, see the [installation page](https://tqec.github.io/tqec/user_guide/installation.html) in the documentation.
