### 2025-09-09 f0f82b5
add back docs caching
