+++ pull/682/_sources/legal_notice.rst.txt
@@ -1,0 +1,5 @@
+Legal Notice
+============
+
+.. include:: ../LEGAL_NOTICE.md
+   :parser: myst_parser.sphinx_
