+++ .github/workflows/gh-pages.yml
@@ -24,6 +24,8 @@ jobs:
         uses: astral-sh/setup-uv@v6
       - name: Set up Python
         uses: actions/setup-python@v5
+        with:
+          python-version: "3.13"
       - name: Install dependencies
         run: |
           uv venv
