### 2025-09-09 cb1adb4
update editable installation command
