+++ .github/workflows/ci.yml
@@ -44,9 +44,9 @@ jobs:
       - name: Install dependencies after creating a virtual environment
         run: |
           uv venv
-          uv pip install ".[all]"
+          uv pip install --group all
       - name: Build
-        run: uv pip install -e .
+        run: uv pip install --editable .
       # Type check the code base.
       - name: Mypy type checking
         run: uv run mypy src/tqec/
