+++ .github/workflows/gh-pages.yml
@@ -32,7 +32,8 @@ jobs:
           sudo apt-get update
           sudo apt-get install -y --no-install-recommends pandoc
           uv pip install --upgrade pip
-          uv pip install -e '.[all]'
+          uv pip install --group all
+          uv pip install --editable .
       - uses: actions/cache@v4
         id: cache-docs
         with:
