### 2025-09-09 b4c6301
update warning in contributor guide
