+++ docs/contributor_guide.rst
@@ -38,14 +38,14 @@ Currently, ``tqec`` is compatible with Python 3.10, 3.11, 3.12 and 3.13.
     See https://stackoverflow.com/a/71585691 for more info.
 
 .. warning::
-    The ``-e`` option to the ``python -m pip install`` call is **important** as it installs an editable version
+    The ``--editable`` option to the ``uv pip install`` call is **important** as it installs an editable version
     of ``tqec``. Without that option, changes made in the folder ``tqec`` will **not** be reflected on the
     ``tqec`` package installed.
 
-    Without the ``-e`` option, ``pip`` copies all the files it needs (mainly, the code) to the current Python
+    Without the ``--editable`` option, ``uv`` copies all the files it needs (mainly, the code) to the current Python
     package folder. Any modification to the original ``tqec`` folder you installed the package from
     will not be reflected automatically on the copied files, which will limit your ability to test new
-    changes on the code base. The ``-e`` option tells ``pip`` to create a link instead of copying, which means
+    changes on the code base. The ``--editable`` option tells ``uv`` to create a link instead of copying, which means
     that the code in the ``tqec`` folder will be the code used when importing ``tqec``.
 
 If you encounter any issue during the installation, please refer to :ref:`installation` for more information.
