### 2025-09-09 dc301db
temporarily run exotic-oses on PR
