+++ .github/workflows/exotic-oses.yml
@@ -1,16 +1,13 @@
 name: exotic-oses
 
 on:
-  push:
+  workflow_run:
+    workflows:
+      - CI
+    types:
+      - completed
     branches:
       - main
-  pull_request:
-    branches:
-      - main
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.ref }}
-  cancel-in-progress: true
 
 jobs:
   exotic-os-python-tests:
