### 2025-09-09 4741504
add uv, pip tabs
