### 2025-09-10 7fbf222
update installation info in user guide
