### 2025-09-10 758290e
us uv sync instead of uv pip install
