### 2025-09-10 0a68965
test: no need for uv sync --editable .
