+++ .github/workflows/exotic-oses.yml
@@ -30,8 +30,6 @@ jobs:
         run: |
           uv venv
           uv sync --group all
-      - name: Build
-        run: uv sync --editable .
       # Perform type checking before testing because that is quicker.
       - name: Mypy type checking
         run: uv run mypy src/tqec/
