### 2025-09-10 8e95cfa
add python version to ci.yml where there is no matrix of python versions
