+++ .github/workflows/ci.yml
@@ -19,6 +19,8 @@ jobs:
       - uses: actions/checkout@v4
       - name: Set up Python
         uses: actions/setup-python@v5
+        with:
+          python-version: "3.13"
       - name: Install uv
         uses: astral-sh/setup-uv@v6
       - name: Create virtual environment and install dependencies
@@ -67,6 +69,8 @@ jobs:
       - uses: actions/checkout@v4
       - name: Set up Python
         uses: actions/setup-python@v5
+        with:
+          python-version: "3.13"
       - name: Install uv
         uses: astral-sh/setup-uv@v6
       - name: Install dependencies and tqec package after creating virtual environment
