+++ src/tqec/compile/specs/library/generators/fixed_boundary.py
@@ -9,6 +9,8 @@ from __future__ import annotations
 import inspect
 from typing import Final, Literal
 
+from pyzx.hrules import is_hadamard
+
 from tqec.compile.specs.base import CubeSpec
 from tqec.compile.specs.enums import SpatialArms
 from tqec.compile.specs.library.generators.constants import (
@@ -1080,6 +1082,7 @@ class FixedBoundaryConventionGenerator:
         is_reversed: bool,
         reset: Basis | None = None,
         measurement: Basis | None = None,
+        is_hadamard
     ) -> Plaquettes:
         """Returns the plaquettes needed to implement **one** pipe connecting to
         a spatial cube.
@@ -1138,26 +1141,45 @@ class FixedBoundaryConventionGenerator:
             SpatialArms.RIGHT,
             SpatialArms.LEFT | SpatialArms.RIGHT,
         ]:
-            return self._get_left_right_spatial_cube_arm_plaquettes(
-                spatial_boundary_basis,
-                arms,
-                linked_cubes,
-                is_reversed,
-                reset,
-                measurement,
-            )
+            if is_hadamard:
+                return self._get_left_right_spatial_hadamard_cube_arm_plaquettes(
+                    spatial_boundary_basis,
+                    arms,
+                    linked_cubes,
+                    is_reversed,
+                    reset,
+                    measurement,
+                )
+            else:
+                return self._get_left_right_spatial_cube_arm_plaquettes(
+                    spatial_boundary_basis,
+                    arms,
+                    linked_cubes,
+                    is_reversed,
+                    reset,
+                    measurement,
+                )
         if arms in [
             SpatialArms.UP,
             SpatialArms.DOWN,
             SpatialArms.UP | SpatialArms.DOWN,
         ]:
-            return self._get_up_down_spatial_cube_arm_plaquettes(
-                spatial_boundary_basis,
-                linked_cubes,
-                is_reversed,
-                reset,
-                measurement,
-            )
+            if is_hadamard:
+                return self._get_up_down_spatial_hadamard_cube_arm_plaquettes(
+                    spatial_boundary_basis,
+                    linked_cubes,
+                    is_reversed,
+                    reset,
+                    measurement,
+                )
+            else:
+                return self._get_up_down_spatial_cube_arm_plaquettes(
+                    spatial_boundary_basis,
+                    linked_cubes,
+                    is_reversed,
+                    reset,
+                    measurement,
+                )
         raise TQECError(f"Got an invalid arm: {arms}.")
 
     def _get_left_right_spatial_cube_arm_plaquettes(
@@ -1170,7 +1192,7 @@ class FixedBoundaryConventionGenerator:
         measurement: Basis | None = None,
     ) -> Plaquettes:
         # This is a regular memory arm, except that we should make sure that one
-        # of the boundary does not override the extended stabilizer.
+        # of the boundary does not override the extended stabilizer (in the 'corner' between the arms).
         z_orientation = (
             Orientation.VERTICAL if spatial_boundary_basis == Basis.Z else Orientation.HORIZONTAL
         )
@@ -1184,10 +1206,34 @@ class FixedBoundaryConventionGenerator:
             regular_memory = regular_memory.without_plaquettes([3])
         return regular_memory
 
+    def _get_left_right_spatial_hadamard_cube_arm_plaquettes(
+            self,
[…]
