+++ pull/682/_sources/_autosummary/tqec.gallery.rst.txt
@@ -1,35 +1,0 @@
-﻿tqec.gallery
-============
-
-.. currentmodule:: tqec.gallery
-
-.. automodule:: tqec.gallery
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-      :toctree:
-   
-      cnot
-      cz
-      memory
-      move_rotation
-      stability
-      steane_encoding
-      three_cnots
-   
-   
-
-   
-   
-   
-
-   
-   
-   
