+++ .github/workflows/ci.yml
@@ -41,6 +41,8 @@ jobs:
       - uses: actions/checkout@v4
       - name: Set up Python
         uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
       - name: Install uv
         uses: astral-sh/setup-uv@v6
       - name: Install dependencies after creating a virtual environment
