+++ .github/workflows/ci.yml
@@ -23,6 +23,8 @@ jobs:
           python-version: "3.13"
       - name: Install uv
         uses: astral-sh/setup-uv@v6
+        with:
+          enable-cache: true
       - name: Create virtual environment and install dependencies
         run: |
           uv venv
@@ -45,6 +47,8 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install uv
         uses: astral-sh/setup-uv@v6
+        with:
+          enable-cache: true
       - name: Install dependencies after creating a virtual environment
         run: |
           uv venv
@@ -75,6 +79,8 @@ jobs:
           python-version: "3.13"
       - name: Install uv
         uses: astral-sh/setup-uv@v6
+        with:
+          enable-cache: true
       - name: Install dependencies and tqec package after creating a virtual environment
         run: |
           uv venv
