### 2025-09-10 3abfbaa
Update docs/contributor_guide.rst
Co-authored-by: Yiming Zhang <61700160+inmzhang@users.noreply.github.com>
