### 2025-09-10 a23f1d0
Migrate to uv (#682)
Fixes #606

~~the first commit in this PR uses
https://mkniewallner.github.io/migrate-to-uv/~~

---------

Co-authored-by: Yiming Zhang <61700160+inmzhang@users.noreply.github.com>
