+++ docs/conf.py
@@ -63,6 +63,10 @@ extensions = [
     # make code blocks in the documentation executable
     # https://jupyter-sphinx.readthedocs.io/en/latest/
     "jupyter_sphinx",
+    # add code tabs; required by the pydata sphinx theme
+    # https://pydata-sphinx-theme.readthedocs.io/en/stable/user_guide/web-components.html#tabs
+    # https://sphinx-design.readthedocs.io/en/pydata-theme/tabs.html
+    "sphinx_design",
 ]
 
 templates_path = ["_templates"]
@@ -189,3 +193,14 @@ autosummary_imported_members = True
 bibtex_bibfiles = ["refs.bib"]
 bibtex_default_style = "unsrt"
 suppress_warnings = ["bibtex.duplicate_label", "bibtex.duplicate_citation"]
+
+from pygments.lexers import BashLexer
+from pygments import highlight
+from pygments.formatters import HtmlFormatter
+
+
+# Map 'uv' and 'pip' to use bash syntax highlighting in the readme
+# or any other markdown documents.
+def setup(app):
+    app.add_lexer("uv", BashLexer)
+    app.add_lexer("pip", BashLexer)
