+++ requirements.txt
@@ -1,15 +1,0 @@
-matplotlib<=3.10.3
-numpy<=2.2.6
-scipy<=1.15.3
-networkx<=3.5
-stim<=1.15.0
-sinter<=1.15.0
-pycollada<=0.9.1
-typing-extensions<=4.15.0
-pycryptosat<=5.11.21
-python-sat[cryptosat]<=1.8.dev20
-tqecd<=0.1.2
-pyzx<=0.9.0
-platformdirs<=4.4.0
-semver<=3.0.4
-svg.py<=1.8.0
