+++ .github/workflows/exotic-oses.yml
@@ -17,6 +17,10 @@ jobs:
     runs-on: ${{matrix.os}}
     steps:
       - uses: actions/checkout@v4
+      - name: Install uv
+        uses: astral-sh/setup-uv@v6
+        with:
+          enable-cache: true
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
@@ -24,17 +28,13 @@ jobs:
           # In theory, if everything succeed in 3.10, everything should succeed in
           # 3.11, 3.12, 3.13 and follow-up version.
           python-version: "3.10"
-          cache: "pip"
       - name: Install dependencies
         run: |
-          python -m pip install --upgrade pip
-          python -m pip install -r requirements.txt -r requirements-dev.txt
-          python -m pip install pytest mypy
-      - name: Build
-        run: python -m pip install '.[all]'
+          uv venv
+          uv sync --group all
       # Perform type checking before testing because that is quicker.
       - name: Mypy type checking
-        run: python -m mypy src/tqec/
+        run: uv run mypy src/tqec/
       # Only perform quick tests to avoid paying too much minutes on that.
       - name: Partial tests
-        run: python -m pytest $(git ls-files '*_test.py') -m "not slow"
+        run: uv run pytest $(git ls-files '*_test.py') -m "not slow"
