+++ docs/conf.py
@@ -5,11 +5,14 @@
 
 from __future__ import annotations
 
+import datetime
+
 # -- Updating sys.path to let autodoc find the tqec package ------------------
 import sys
 import typing as ty
 from pathlib import Path
-import datetime
+
+from pygments.lexers import BashLexer
 
 DOCUMENTATION_DIRECTORY = Path(__file__).parent
 PROJECT_DIRECTORY = DOCUMENTATION_DIRECTORY.parent
@@ -148,6 +151,11 @@ def setup(app):
     # Connect the autodoc-skip-member event from apidoc to the callback
     app.connect("autodoc-skip-member", autodoc_skip_member_handler)
 
+    # Map 'uv' and 'pip' to use bash syntax highlighting in the readme
+    # or any other markdown documents.
+    app.add_lexer("uv", BashLexer)
+    app.add_lexer("pip", BashLexer)
+
 
 autodoc_member_order = "groupwise"
 # See https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html#confval-autoclass_content
@@ -193,14 +201,3 @@ autosummary_imported_members = True
 bibtex_bibfiles = ["refs.bib"]
 bibtex_default_style = "unsrt"
 suppress_warnings = ["bibtex.duplicate_label", "bibtex.duplicate_citation"]
-
-from pygments.lexers import BashLexer
-from pygments import highlight
-from pygments.formatters import HtmlFormatter
-
-
-# Map 'uv' and 'pip' to use bash syntax highlighting in the readme
-# or any other markdown documents.
-def setup(app):
-    app.add_lexer("uv", BashLexer)
-    app.add_lexer("pip", BashLexer)
