+++ pyproject.toml
@@ -101,6 +101,7 @@ dev = [
     # cannot afford someone installing a different version of ruff and being forbidden to commit
     # anything because the different version sees some errors in the code.
     "ruff==0.12.5",
+    "ty>=0.0.1a20",
     "scipy-stubs<=1.16.0.0",
     "sphinx<=8.1.3",
     "sphinxcontrib-mermaid<=1.0.0",
