### 2025-09-11 317167e
remove mypy as dependency, add ty: expect CI failures
