### 2025-09-12 c61f8ac
make sure measure rules only apply apply to measurements
