### 2025-09-12 14f0bfa
add X basis reset and measurement rules for the si1000 noise model
