+++ src/tqec/compile/specs/library/generators/fixed_boundary.py
@@ -361,6 +361,56 @@ class FixedBoundaryConventionGenerator:
             RPNGDescription.from_string(f"---- -{b}{vs[1]}- ---- {r}{o}{vs[3]}{m}"),
         )
 
+    def get_3_body_horizontal_hadamard_rpng_descriptions(
+        self,
+        top_left_basis: Basis,
+        is_reversed: bool,
+        reset: Basis | None = None,
+        measurement: Basis | None = None,
+    ) -> tuple[RPNGDescription, RPNGDescription, RPNGDescription, RPNGDescription]:
+        """Return the four 3-body stabilizer plaquettes, with split basis horizontally,
+        as needed for a horizontal hadamard.
+
+        Args:
+            top_left_basis: basis of the top-left-most stabilizer.
+            is_reversed: flag indicating if the plaquette schedule should be
+                reversed or not. Useful to limit the loss of code distance when
+                hook errors are not correctly oriented by alternating regular
+                and reversed plaquettes.
+            reset: basis of the reset operation performed on data-qubits. Defaults
+                to ``None`` that translates to no reset being applied on data-qubits.
+            measurement: basis of the measurement operation performed on data-qubits.
+                Defaults to ``None`` that translates to no measurement being applied
+                on data-qubits.
+
+        Returns:
+            the four 3-body stabilizer measurement plaquettes, with split basis horizontally.
+            Their order follows the usual convention:
+            ``(top_left, top_right, bottom_left, bottom_right)``.
+
+        """
+        b = top_left_basis.value.lower()
+        o = top_left_basis.flipped().value.lower()
+        # r/m: reset/measurement basis applied to each data-qubit
+        r = reset.value.lower() if reset is not None else "-"
+        m = measurement.value.lower() if measurement is not None else "-"
+        # Note: the schedule of CNOT gates in corner plaquettes is less important
+        # because hook errors do not exist on 3-body stabilizers. We arbitrarily
+        # pick the vertical schedule.
+        s = VERTICAL_HOOK_SCHEDULES[is_reversed]
+        # Note that we include resets and measurements on all the used data-qubits.
+        # That should be fine because this plaquette only touches cubes and pipes
+        # that are related to the spatial junction being implemented, and it is not
+        # valid to have a temporal pipe coming from below a spatial junction, hence
+        # the data-qubits cannot be already initialised to a value we would like to
+        # keep and that would be destroyed by reset/measurement.
+        return (
+            RPNGDescription.from_string(f"---- {r}{b}{s[1]}{m} {r}{o}{s[2]}{m} {r}{o}{s[3]}{m}"),
+            RPNGDescription.from_string(f"{r}{b}{s[0]}{m} ---- {r}{o}{s[2]}{m} {r}{o}{s[3]}{m}"),
+            RPNGDescription.from_string(f"{r}{b}{s[0]}{m} {r}{b}{s[1]}{m} ---- {r}{o}{s[3]}{m}"),
+            RPNGDescription.from_string(f"{r}{b}{s[0]}{m} {r}{b}{s[1]}{m} {r}{o}{s[2]}{m} ----"),
+        )
+
     ############################################################
     #                          Memory                          #
     ############################################################
@@ -1068,6 +1118,7 @@ class FixedBoundaryConventionGenerator:
     def get_spatial_cube_arm_plaquettes(
         self,
         spatial_boundary_basis: Basis,
+        top_left_basis: Basis,
         arms: SpatialArms,
         linked_cubes: tuple[CubeSpec, CubeSpec],
         is_reversed: bool,
[…]
