### 2025-09-14 c01a9e3
document the addition of X basis reset and measurement rules in si1000
