### 2025-09-14 760b54b
Fix _check_correlation_surface_validity() (#693)
In
`tqec.compile.observables.abstract_observable_check_correlation_surface_validity()`,
the number of neighboring edges is incorrectly derived with
`len(edges)`, where `edges = correlation_surface.edges_at(v)`. When the
correlation surface on a neighboring edge has two bases, for example,
this causes duplicate counting. See the added test for an example that
falsely triggers the error in the if statement.
