### 2025-09-15 1ddd114
Modify BlockGraph._ports when relabeling ports (#699)
`BlockGraph.relabel_cubes()` should change `BlockGraph._ports` as well
when the relabelled cube is a port.
