### 2025-09-15 0205fdf
further mypy fixes
