### 2025-09-15 fbfe96b
further mypy fixes
