### 2025-09-16 94cc44f
Change the legal notice to an ethical notice, on behalf of angelaelisa
