+++ ETHICAL_NOTICE.md
@@ -1,0 +1,78 @@
+# ETHICAL NOTICE
+
+_Last updated: 16/09/2025_
+
+## 1. Purpose and Scope
+
+This document expresses the **ethical intent and community values** of the TQEC (Topological Quantum Error Correction) project.
+**Important**: This is **not a license**.
+The sole legal license governing this software is the [Apache License 2.0](https://www.apache.org/licenses/LICENSE-2.0).
+
+Nothing in this document alters or restricts the rights granted under that license.
+
+The TQEC (Topological Quantum Error Correction) library is an open-source software project made available for the advancement of **large-scale fault-tolerant quantum computing**, particularly based on **surface code** and  **lattice surgery**.
+
+This library is distributed to promote **academic, educational, and civilian research**. Its open accessibility reflects the contributors’ commitment to transparency, scientific progress, and global collaboration.
+
+## 2. Permitted Uses
+
+You may use, modify, and distribute this software under the terms of the [Apache License 2.0](https://www.apache.org/licenses/LICENSE-2.0). The following activities are **consistent with the intent of this project**:
+- Academic or scientific research
+- Educational instruction and student use
+- Civilian research and development
+- Open collaboration within the quantum computing community
+
+## 3. Prohibited Uses
+
+The following uses are **inconsistent with the intent of this project**:
+
+- Military, defense, or weapons-related development
+- Surveillance, intelligence gathering, or national security infrastructure
+- Dual-use technologies that may serve both civilian and military purposes
+- Applications that violate international human rights, academic integrity, or ethical standards
+
+While these are not legal prohibitions, the maintainers request that users and contributors **refrain from engaging in, supporting, or facilitating** such activities.
+
+## 4. Openness and Research Context
+
+This software:
+- Is **published**,
+- Results from **fundamental research**,
+- Is distributed **freely without access restrictions**,
+- It is not designed for use in any prohibited military or dual-use context,
+- Is distributed without access limitations or commercial restrictions,
+- Contains no encryption, surveillance, or defense-related functionality, and
+- Is not intended for any dual-use or sensitive application.
+
+> **Important**: These manifestations apply **only to the original, open-source version of TQEC**.
+
+All users and contributors are solely responsible for verifying compliance with relevant export, import, sanctions, and dual-use laws in their jurisdiction.
+
+## 5. Contributor Certification
+
+By contributing to this repository, you affirm your support for the following principles:
+- You have the legal right to make the contribution.
+- Your submission does not knowingly violate export control laws or proprietary rights.
+- You are not acting on behalf of any military or surveillance-related organization.
+- You acknowledge and respect the ethical values expressed in this notice.
+-
+## 6. Disclaimer of Warranty and Liability
+
+This software is provided **“AS IS”**, without any express or implied warranties, including but not limited to warranties of merchantability, fitness for a particular purpose, or non-infringement.
+
+The maintainers and contributors **accept no liability** for:
+
+- Damage or loss resulting from the use or misuse of this software.
+- Errors in compilation or simulation.
+- Indirect, incidental, or consequential damages.
+- Legal consequences arising from unauthorized or unlawful use.
+
+Use of this software is at **your sole risk and responsibility**.
+
+## 7. Final Remarks
+
+TQEC is part of a shared global effort to advance fault-tolerant quantum computing for peaceful and ethical applications. We welcome collaboration from researchers, educators, and institutions who share this vision.
+
+If you are uncertain about the legality or appropriateness of using this software in your context, please consult appropriate legal counsel and, if needed, seek clarification from the maintainers.
+---
+**Thank you for respecting the intent, integrity, and openness of this project.**
