### 2025-09-16 #8930 @BSchelpe
Hi @danielskatz, I'm sorry for the delay. The lawyer in our group is travelling, but she's kindly sent me the words to change things. At a high-level, our plan to fix this is to convert the legal_notice into an ethical intent. In summary the ethical notice now says: "This document expresses the ethical intent and values of the TQEC community. It is not a license and does not alter or restrict the rights granted under the Apache License 2.0 , which is the sole legal license for this software."

I've made a PR for the change to wording (and file name) here: https://github.com/tqec/tqec/pull/701

I will let you know as soon as it is merged.
