### 2025-09-17 f2b9623
revert the changes to si1000; clarify the current incompatibility in the docstring and error message instead
