+++ docs/ethical_notice.rst
@@ -1,0 +1,5 @@
+Ethical Notice
+============
+
+.. include:: ../ETHICAL_NOTICE.md
+   :parser: myst_parser.sphinx_
