### 2025-09-17 7d6356b
Change the legal notice to an ethical notice, on behalf of angelaelisa (#701)
This closes issue #684 
@angelaelisa is travelling and so I am submitting the PR on her behalf,
but this is entirely her work.
