### 2025-09-19 4fdef0b
fix errors in src/tqec/utils/scale.py
