### 2025-09-19 a30a96d
fix argument to `reduce` is incorrect, remove reduce
