+++ src/tqec/interop/pyzx/correlation.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 import itertools
 from collections.abc import Iterator
 from fractions import Fraction
-from functools import reduce
 
 import stim
 from pyzx.graph.graph_s import GraphS
@@ -46,7 +45,9 @@ def correlation_surface_to_pauli_web(
 
     pauli_web = PauliWeb(g)
     for e, bases in half_edge_bases.items():
-        pauli = reduce(multiply_paulis, bases, "I")
+        pauli = "I"
+        for basis in bases:
+            pauli = multiply_paulis(pauli, basis)
         pauli_web.add_half_edge(e, pauli)
     return pauli_web
 
