### 2025-09-21 13e7ef5
invalid return type in scale.py
