+++ src/tqec/compile/blocks/layers/merge.py
@@ -86,9 +86,9 @@ def merge_composed_layers(
             f"Found the following different lengths: {different_timesteps}."
         )
     if contains_only_repeated_layers(layers):
-        return merge_repeated_layers(layers, scalable_qubit_shape)
+        return merge_repeated_layers(layers, scalable_qubit_shape)  # type: ignore
     if contains_only_sequenced_layers(layers):
-        return merge_sequenced_layers(layers, scalable_qubit_shape)
+        return merge_sequenced_layers(layers, scalable_qubit_shape)  # type: ignore
     # We are left here with a mix of RepeatedLayer and SequencedLayers.
     # Check that, in case a new subclass of BaseComposedLayer has been introduced.
     if not contains_only_repeated_or_sequenced_layers(layers):
@@ -101,7 +101,7 @@ def merge_composed_layers(
             f"Found instances of {unknown_types_names} that are not yet "
             "implemented in _merge_composed_layers."
         )
-    return merge_repeated_and_sequenced_layers(layers, scalable_qubit_shape)
+    return merge_repeated_and_sequenced_layers(layers, scalable_qubit_shape)  # type: ignore
 
 
 def merge_repeated_layers(
