### 2025-09-22 33b8daf
feat: reduce branches for passthrough node
