+++ pull/692/_sources/gallery/steane_encoding.ipynb.txt
@@ -1,322 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "# Steane Encoding\n",
-    "\n",
-    "This notebook demonstrates the construction and simulation of a logical Steane encoder circuit. Steane code was introduced in [<cite data-footcite-t=\"Steane_1996\">Steane (1996)</cite>]. In surface code, this logical circuit can be used to prepare the magic state for logical S gate [<cite data-footcite-t=\"Fowler_2012\">[Fowler et.al (2012)]</cite>]. "
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "## Construction\n",
-    "\n",
-    "`tqec` provides builtin functions `tqec.gallery.steane_encoding` to construct it."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "from tqec.gallery.steane_encoding import steane_encoding\n",
-    "\n",
-    "graph = steane_encoding()\n",
-    "graph.view_as_html()"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "We can use the `find_correlation_surfaces()` method to identify all correlation surfaces, and then visualize them using the `view_as_html()` method. In this case, there are a total of seven correlation surfaces. For simplicity, only two are shown here."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "correlation_surfaces = graph.find_correlation_surfaces()\n",
-    "len(correlation_surfaces)"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "graph.view_as_html(\n",
-    "    pop_faces_at_directions=(\"-Y\",),\n",
-    "    show_correlation_surface=correlation_surfaces[0],\n",
-    ")"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "graph.view_as_html(\n",
-    "    pop_faces_at_directions=(\"-Y\",),\n",
-    "    show_correlation_surface=correlation_surfaces[1],\n",
-    ")"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "## Circuit\n",
-    "\n",
[…]
