### 2025-09-23 49da073
draft implementation of an efficient algorithm for finding correlation surfaces
