+++ src/tqec/interop/pyzx/correlation_experimental.py
@@ -1,0 +1,427 @@
+"""Defines the :class:`.CorrelationSurface` class and functions to find them in a ZX graph."""
+
+from __future__ import annotations
+
+from fractions import Fraction
+from functools import partial, reduce
+from itertools import chain, combinations, product
+
+import networkx as nx
+import stim
+from pyzx.graph.graph_s import GraphS
+from pyzx.pauliweb import PauliWeb, multiply_paulis
+from pyzx.utils import FractionLike, VertexType
+
+from tqec.computation.correlation import CorrelationSurface, ZXEdge, ZXNode
+from tqec.interop.pyzx.utils import (
+    is_boundary,
+    is_hardmard,
+    is_s,
+    is_x_no_phase,
+    is_z_no_phase,
+    is_zx_no_phase,
+)
+from tqec.utils.enums import Basis, Pauli
+from tqec.utils.exceptions import TQECError
+
+
+def correlation_surface_to_pauli_web(
+    correlation_surface: CorrelationSurface, g: GraphS
+) -> PauliWeb[int, tuple[int, int]]:
+    """Convert the correlation surface to a Pauli web.
+
+    Args:
+        correlation_surface: The correlation surface to convert.
+        g: The ZX graph the correlation surface is based on.
+
+    Returns:
+        A `PauliWeb` representation of the correlation surface.
+
+    """
+    half_edge_bases: dict[tuple[int, int], set[str]] = {}
+    for edge in correlation_surface.span:
+        u, v = edge.u, edge.v
+        half_edge_bases.setdefault((u.id, v.id), set()).add(u.basis.value)
+        half_edge_bases.setdefault((v.id, u.id), set()).add(v.basis.value)
+
+    pauli_web = PauliWeb(g)
+    for e, bases in half_edge_bases.items():
+        pauli = reduce(multiply_paulis, bases, "I")
+        pauli_web.add_half_edge(e, pauli)
+    return pauli_web
+
+
+def pauli_web_to_correlation_surface(
+    pauli_web: PauliWeb[int, tuple[int, int]],
+) -> CorrelationSurface:
+    """Create a correlation surface from a Pauli web."""
+    span: set[ZXEdge] = set()
+    half_edges: dict[tuple[int, int], str] = pauli_web.half_edges()
+    while half_edges:
+        (u, v), pauli_u = half_edges.popitem()
+        pauli_v = half_edges.pop((v, u))
+        if pauli_u == "Y":  # pragma: no cover
+            span.add(ZXEdge(ZXNode(u, Basis.X), ZXNode(v, Basis.Z)))
+            span.add(ZXEdge(ZXNode(u, Basis.Z), ZXNode(v, Basis.Z)))
+            continue
+        span.add(ZXEdge(ZXNode(u, Basis(pauli_u)), ZXNode(v, Basis(pauli_v))))
+    return CorrelationSurface(frozenset(span))
+
+
+def find_correlation_surfaces(
+    g: GraphS, reduce_to_minimal_generators: bool = True
+) -> list[CorrelationSurface]:
+    """Find the correlation surfaces in a ZX graph.
+
+    Starting from each leaf node in the graph, the function explores how can the X/Z logical
+    observable move through the graph to form a correlation surface:
+
+    - For a X/Z type leaf node, it can only support the logical observable with the opposite type.
+      Only a single type of logical observable is explored from the leaf node.
+    - For a Y type leaf node, it can only support the Y logical observable, i.e. the presence of
+      both X and Z logical observable. Both X and Z type logical observable are explored from the
+      leaf node. And the two correlation surfaces are combined to form the Y type correlation
+      surface.
+    - For the BOUNDARY node, it can support any type of logical observable. Both X and Z type
[…]
