### 2025-09-24 8ce9f5a
union type error in schedule/circuit.py - check for unit test failures
