### 2025-09-24 1d4b465
unspupported operator, check for unit test failures
