+++ src/tqec/compile/blocks/layers/atomic/layout_test.py
@@ -146,11 +146,13 @@ def test_to_template_and_plaquettes_multiple(empty_plaquette_layer: PlaquetteLay
     for t in template._layout.values():
         assert type(t) is type(empty_plaquette_layer.template)
     assert plaquettes.collection == (
-        plaquettes.collection  # type: ignore
-        | FrozenDefaultDict(
+        FrozenDefaultDict(
             {
-                i + empty_plaquette_layer.template.expected_plaquettes_number: plaq
-                for i, plaq in plaquettes.collection.items()
+                **plaquettes.collection,
+                **{
+                    i + empty_plaquette_layer.template.expected_plaquettes_number: plaq
+                    for i, plaq in plaquettes.collection.items()
+                },
             },
             default_value=plaquettes.collection.default_value,
         )
