### 2025-09-24 96415d2
func no attribute name, check for unit test failures with casting
