### 2025-09-24 1ff19e9
don't use reduce
