### 2025-09-24 454ac59
add BaseLayer type annotation, check for unit test failures
