### 2025-09-24 8ad1d5c
raise better errors
