### 2025-09-24 8df1e41
add comments for wildcard entries
