+++ src/tqec/templates/enums.py
@@ -22,5 +22,6 @@ class TemplateBorder(Enum):
                 return TemplateBorder.RIGHT
             case TemplateBorder.RIGHT:
                 return TemplateBorder.LEFT
+            # wildcard entry added as it was flagged by ty
             case _:
                 raise ValueError(f"Cannot return the opposite border of {self}.")
