+++ src/tqec/interop/pyzx/correlation_experimental.py
@@ -7,6 +7,7 @@ from functools import partial, reduce
 from itertools import chain, combinations, product
 
 import networkx as nx
+import numpy as np
 import stim
 from pyzx.graph.graph_s import GraphS
 from pyzx.pauliweb import PauliWeb, multiply_paulis
@@ -19,7 +20,6 @@ from tqec.interop.pyzx.utils import (
     is_s,
     is_x_no_phase,
     is_z_no_phase,
-    is_zx_no_phase,
 )
 from tqec.utils.enums import Basis, Pauli
 from tqec.utils.exceptions import TQECError
@@ -140,10 +140,8 @@ def _find_correlation_surfaces_from_leaf(
             bases_at_leaves[closed_leaf] = Pauli.Y
         elif is_z_no_phase(g, closed_leaf):
             bases_at_leaves[closed_leaf] = Pauli.X
-        elif is_x_no_phase(g, closed_leaf):
-            bases_at_leaves[closed_leaf] = Pauli.Z
         else:
-            raise TQECError(f"Unsupported leaf node type: {g.type(closed_leaf)}.")
+            bases_at_leaves[closed_leaf] = Pauli.Z
     neighbor = next(iter(g.neighbors(leaf)))
     pauli_graphs = []
     for basis in (
@@ -159,48 +157,115 @@ def _find_correlation_surfaces_from_leaf(
     if g.vertex_degree(neighbor) == 1:  # make sure no leaf node will be in the frontier
         frontier = []
         if neighbor in bases_at_leaves:
-            pauli_graphs = filter(
-                lambda pg: pg.nodes[neighbor][leaf] in (bases_at_leaves[neighbor], Pauli.I),
-                pauli_graphs,
+            pauli_graphs = list(
+                filter(
+                    lambda pg: pg.nodes[neighbor][leaf] in (bases_at_leaves[neighbor], Pauli.I),
+                    pauli_graphs,
+                )
             )
     else:
         frontier = [neighbor]
         explored_leaves = [leaf]
 
     while frontier:
+        stabilizer_nodes = explored_leaves + frontier
         cur = frontier.pop()
         unconnected_neighbors = sorted(
             filter(lambda n: n not in pauli_graphs[0][cur], g.neighbors(cur))
         )
-        neighbor_leaves_bases = {
-            n: _flip_pauli_based_on_edge_type(g, (cur, n), bases_at_leaves[n])
-            for n in unconnected_neighbors
-            if n in bases_at_leaves
-        }
+        # neighbor_leaves_bases = {
+        #     n: _flip_pauli_based_on_edge_type(g, (cur, n), bases_at_leaves[n])
+        #     for n in unconnected_neighbors
+        #     if n in bases_at_leaves
+        # }
         unexplored_neighbors = [n for n in unconnected_neighbors if n not in pauli_graphs[0]]
+        broadcast_basis = "X" if g.type(cur) == VertexType.Z else "Z"
+        passthrough_basis = Basis[broadcast_basis].flipped()
 
-        new_pauli_graphs = []
+        valid_graphs, invalid_graphs, syndromes = [], [], []
         stabilizer_list = []
         for pauli_graph in pauli_graphs:
-            incident_paulis = list(pauli_graph.nodes[cur].values())
-            if not is_zx_no_phase(g, cur):
-                raise TQECError(
-                    f"Unsupported non-leaf node type: {g.type(cur)} with phase {g.phase(cur)}."
-                )
-            broadcast_basis = "X" if g.type(cur) == VertexType.Z else "Z"
+            incident_paulis = pauli_graph.nodes[cur].values()
+            passthrough_parity = sum(p.has_basis(passthrough_basis) for p in incident_paulis) % 2
[…]
