### 2025-09-24 5216562
Update src/tqec/compile/blocks/enums.py
