### 2025-09-24 309af84
use dict of dicts instead of networkx to represent Pauli graphs
