### 2025-09-24 2cba989
over-pruning issue fix attempt
