+++ src/tqec/interop/pyzx/correlation_experimental.py
@@ -188,9 +188,10 @@ def _find_correlation_surfaces_from_leaf(
                 broadcast_pauli = Pauli.I
             else:  # invalid broadcast
                 valid = False
-            if not unconnected_neighbors and passthrough_parity:  # invalid passthrough
-                valid = False
-                broadcast_supports += [True]
+            if not unconnected_neighbors:  # invalid passthrough
+                broadcast_supports += [passthrough_parity]
+                if passthrough_parity:
+                    valid = False
 
             stabilizer = stim.PauliString(
                 "".join(
@@ -211,51 +212,55 @@ def _find_correlation_surfaces_from_leaf(
             else:
                 invalid_graphs.append(pauli_graph)
                 syndromes.append(
-                    np.array(broadcast_supports).dot(1 << np.arange(len(broadcast_supports)))
+                    int(np.array(broadcast_supports).dot(1 << np.arange(len(broadcast_supports))))
                 )
+
+        all_one = (1 << len(broadcast_supports)) - 1
+        if not unconnected_neighbors:
+            all_one ^= 1 << (len(broadcast_supports) - 1)
 
         for i, (pauli_graph, syndrome) in enumerate(zip(invalid_graphs, syndromes)):
             if len(stabilizer_list) == 2 * len(stabilizer_nodes):
                 break
-            for indices in chain.from_iterable(
-                combinations(range(len(invalid_graphs)), num_flips)
-                for num_flips in range(1, len(invalid_graphs) + 1)
-            ):
-                if i in indices:
+            for target in (syndrome, syndrome ^ all_one):
+                indices = _find_subset_xor(target, syndromes[:i] + syndromes[i + 1 :])
+                if indices is None:
                     continue
-                new_pauli_graph = reduce(
-                    _multiply_pauli_graphs, [invalid_graphs[j] for j in indices]
-                )
+                indices = [j if j < i else j + 1 for j in indices]
+                break
+            if not indices:
+                continue
+            new_pauli_graph = reduce(
+                _multiply_pauli_graphs, [pauli_graph] + [invalid_graphs[j] for j in indices]
+            )
 
-                incident_paulis = new_pauli_graph[cur].values()
-                passthrough_parity = (
-                    sum(p.has_basis(passthrough_basis) for p in incident_paulis) % 2
-                )
-                broadcast_supports = [p.has_basis(broadcast_basis) for p in incident_paulis]
-                valid = True
-                if all(broadcast_supports):
-                    broadcast_pauli = Pauli[broadcast_basis]
-                elif not any(broadcast_supports):
-                    broadcast_pauli = Pauli.I
-                else:  # invalid broadcast
-                    continue
-                if not unconnected_neighbors and passthrough_parity:  # invalid passthrough
-                    continue
+            incident_paulis = new_pauli_graph[cur].values()
+            passthrough_parity = sum(p.has_basis(passthrough_basis) for p in incident_paulis) % 2
+            broadcast_supports = [p.has_basis(broadcast_basis) for p in incident_paulis]
+            valid = True
+            if all(broadcast_supports):
+                broadcast_pauli = Pauli[broadcast_basis]
+            elif not any(broadcast_supports):
+                broadcast_pauli = Pauli.I
+            else:  # invalid broadcast
+                raise TQECError("This should not happen.")
+            if not unconnected_neighbors and passthrough_parity:  # invalid passthrough
+                raise TQECError("This should not happen.")
 
-                stabilizer = stim.PauliString(
-                    "".join(
-                        pauli.value
-                        for pauli in chain.from_iterable(
[…]
