### 2025-09-25 6b2e143
Update src/tqec/compile/tree/node.py
Co-authored-by: Yiming Zhang <61700160+inmzhang@users.noreply.github.com>
