### 2025-09-25 5ba3ff8
Add Astral's type checker & remove mypy (#692)
Fix #585

---------

Co-authored-by: Yiming Zhang <61700160+inmzhang@users.noreply.github.com>
