+++ pyproject.toml
@@ -64,7 +64,6 @@ Issues = "https://github.com/tqec/tqec/issues"
 [dependency-groups]
 test = [
   "pytest<=8.4.1",
-  "mypy<=1.17.0",
   "pytest-cov<=7.0.0",
   "types-networkx<=3.5.0.20250610",
   "scipy-stubs<=1.16.0.0",
@@ -88,7 +87,6 @@ bench = ["pyinstrument"]
 
 dev = [
     "jupyterlab<=4.4.3",
-    "mypy<=1.17.0",
     "myst-parser<=4.0.1",
     "nbsphinx<=0.9.7",
     "pandoc<=2.4",
@@ -101,6 +99,7 @@ dev = [
     # cannot afford someone installing a different version of ruff and being forbidden to commit
     # anything because the different version sees some errors in the code.
     "ruff==0.12.5",
+    "ty>=0.0.1a20",
     "scipy-stubs<=1.16.0.0",
     "sphinx<=8.1.3",
     "sphinxcontrib-mermaid<=1.0.0",
