+++ pull/692/_sources/_autosummary/tqec.compile.rst.txt
@@ -1,29 +1,0 @@
-﻿tqec.compile
-============
-
-.. currentmodule:: tqec
-
-.. automodule:: tqec.compile
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-      :toctree:
-   
-      compile_block_graph
-   
-   
-
-   
-   
-   
-
-   
-   
-   
