+++ src/tqec/interop/pyzx/correlation_experimental.py
@@ -2,12 +2,12 @@
 
 from __future__ import annotations
 
+from collections.abc import Iterable
 from copy import deepcopy
 from fractions import Fraction
 from functools import partial, reduce
 from itertools import chain, combinations, product
 
-import numpy as np
 import stim
 from pyzx.graph.graph_s import GraphS
 from pyzx.pauliweb import PauliWeb, multiply_paulis
@@ -137,6 +137,7 @@ def _find_correlation_surfaces_from_leaf(
         else:
             bases_at_leaves[closed_leaf] = Pauli.Z
     neighbor = next(iter(g.neighbors(leaf)))
+
     pauli_graphs = []
     for basis in (
         [bases_at_leaves[leaf]] if leaf in bases_at_leaves else [Pauli.X, Pauli.Z, Pauli.Y]
@@ -147,6 +148,7 @@ def _find_correlation_surfaces_from_leaf(
         graph[neighbor] = {}
         graph[neighbor][leaf] = _flip_pauli_based_on_edge_type(g, (leaf, neighbor), basis)
         pauli_graphs.append(graph)
+
     if g.vertex_degree(neighbor) == 1:  # make sure no leaf node will be in the frontier
         frontier = []
         if neighbor in bases_at_leaves:
@@ -162,124 +164,99 @@ def _find_correlation_surfaces_from_leaf(
 
     while frontier:
         stabilizer_nodes = explored_leaves + frontier
-        cur = frontier.pop()
-        unconnected_neighbors = sorted(
-            filter(lambda n: n not in pauli_graphs[0][cur], g.neighbors(cur))
+        stabilizer_length = sum(len(pauli_graphs[0][n]) for n in stabilizer_nodes)
+        current_node = frontier.pop()
+        unconnected_neighbors = list(
+            filter(
+                lambda n: n not in pauli_graphs[0][current_node],
+                g.neighbors(current_node),
+            )
         )
-        # neighbor_leaves_bases = {
-        #     n: _flip_pauli_based_on_edge_type(g, (cur, n), bases_at_leaves[n])
-        #     for n in unconnected_neighbors
-        #     if n in bases_at_leaves
-        # }
         unexplored_neighbors = [n for n in unconnected_neighbors if n not in pauli_graphs[0]]
-        broadcast_basis = "X" if g.type(cur) == VertexType.Z else "Z"
-        passthrough_basis = Basis[broadcast_basis].flipped()
+        flip_basis = g.type(current_node) == VertexType.X
+        broadcast_basis = Basis.Z if flip_basis else Basis.X
+        passthrough_basis = broadcast_basis.flipped()
 
-        valid_graphs, invalid_graphs, syndromes = [], [], []
-        stabilizer_list = []
+        # check if each Pauli graph candidate satisfies broadcast and passthrough constraints
+        # on the current node and is not a product of previously checked valid Pauli graphs
+        valid_graphs, invalid_graphs, syndromes, stabilizer_list = [], [], [], []
         for pauli_graph in pauli_graphs:
-            incident_paulis = pauli_graph[cur].values()
-            passthrough_parity = sum(p.has_basis(passthrough_basis) for p in incident_paulis) % 2
-            broadcast_supports = [p.has_basis(broadcast_basis) for p in incident_paulis]
+            supports = _pauli_support_on_nodes(pauli_graph, [current_node])
+            passthrough_parity = sum(supports[1 - flip_basis]) % 2
             valid = True
-            if all(broadcast_supports):
-                broadcast_pauli = Pauli[broadcast_basis]
-            elif not any(broadcast_supports):
+            syndrome = supports[flip_basis]
+            if all(syndrome):
+                broadcast_pauli = Pauli[broadcast_basis.value]
[…]
