### 2025-09-28 72e434d
Construct circuit for fixed bulk xtop ybasis init/meas
