### 2025-09-28 066371d
remove pre-commit config for now
