+++ .pre-commit-config.yaml
@@ -1,34 +1,0 @@
-repos:
-  - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v5.0.0
-    hooks:
-      - id: check-yaml
-      - id: end-of-file-fixer
-      - id: trailing-whitespace
-  - repo: https://github.com/asottile/pyupgrade
-    rev: v3.19.0
-    hooks:
-      - id: pyupgrade
-  - repo: https://github.com/crate-ci/typos
-    rev: v1.26.1
-    hooks:
-      - id: typos
-  - repo: https://github.com/gitleaks/gitleaks
-    rev: v8.21.1
-    hooks:
-      - id: gitleaks
-  - repo: https://github.com/astral-sh/ruff-pre-commit
-    # Ruff version.
-    rev: v0.12.3
-    hooks:
-      # Run the linter.
-      - id: ruff-check
-        types_or: [ python, pyi ]
-        args: [ --fix ]
-      # Run the formatter.
-      - id: ruff-format
-        types_or: [ python, pyi ]
-  # - repo: https://github.com/PyCQA/docformatter
-  #   rev: v1.7.5
-  #   hooks:
-  #     - id: docformatter
