+++ ztop.txt
@@ -1,0 +1,1 @@
+https://algassert.com/crumble#circuit=Q(1.5,2.5)0;Q(1.5,3.5)1;Q(1.5,4.5)2;Q(1.5,5.5)3;Q(2,2)4;Q(2,3)5;Q(2,4)6;Q(2,5)7;Q(2,6)8;Q(2.5,1.5)9;Q(2.5,2.5)10;Q(2.5,3.5)11;Q(2.5,4.5)12;Q(2.5,5.5)13;Q(2.5,6.5)14;Q(3,2)15;Q(3,3)16;Q(3,4)17;Q(3,5)18;Q(3,6)19;Q(3.5,1.5)20;Q(3.5,2.5)21;Q(3.5,3.5)22;Q(3.5,4.5)23;Q(3.5,5.5)24;Q(4,2)25;Q(4,3)26;Q(4,4)27;Q(4,5)28;Q(4,6)29;Q(4.5,1.5)30;Q(4.5,2.5)31;Q(4.5,3.5)32;Q(4.5,4.5)33;Q(4.5,5.5)34;Q(4.5,6.5)35;Q(5,2)36;Q(5,3)37;Q(5,4)38;Q(5,5)39;Q(5,6)40;Q(5.5,1.5)41;Q(5.5,2.5)42;Q(5.5,3.5)43;Q(5.5,4.5)44;Q(5.5,5.5)45;Q(6,2)46;Q(6,3)47;Q(6,4)48;Q(6,5)49;Q(6,6)50;Q(6.5,3.5)51;Q(6.5,5.5)52;TICK;POLYGON(0,0,1,0.25)4_15_16_5;POLYGON(0,0,1,0.25)25_36_37_26;POLYGON(0,0,1,0.25)16_26_27_17;POLYGON(0,0,1,0.25)37_47_48_38;POLYGON(0,0,1,0.25)6_17_18_7;POLYGON(0,0,1,0.25)27_38_39_28;POLYGON(0,0,1,0.25)18_28_29_19;POLYGON(0,0,1,0.25)39_49_50_40;POLYGON(0,0,1,0.25)25_15;POLYGON(0,0,1,0.25)46_36;POLYGON(0,0,1,0.25)40_29;POLYGON(0,0,1,0.25)19_8;POLYGON(1,0,0,0.25)15_25_26_16;POLYGON(1,0,0,0.25)36_46_47_37;POLYGON(1,0,0,0.25)5_16_17_6;POLYGON(1,0,0,0.25)26_37_38_27;POLYGON(1,0,0,0.25)17_27_28_18;POLYGON(1,0,0,0.25)38_48_49_39;POLYGON(1,0,0,0.25)7_18_19_8;POLYGON(1,0,0,0.25)28_39_40_29;POLYGON(1,0,0,0.25)47_48;POLYGON(1,0,0,0.25)49_50;POLYGON(1,0,0,0.25)6_7;POLYGON(1,0,0,0.25)4_5;TICK;R_9_20_30_41_10_31_43_22_12_33_45_24_14_35;RX_0_1_2_3_21_42_11_32_51_44_23_13_34_52;TICK;CX_29_24_39_33_48_43_50_45_34_40_44_49_13_19_23_28_32_38_42_47_21_26_11_17_0_5_2_7_16_10_18_12_25_20_46_41_37_31_27_22;TICK;CX_40_45_19_24_28_33_38_43_7_12_17_22_26_31_36_41_15_20_5_10_0_4_2_6_21_25_42_46_32_37_11_16_23_27_44_48_34_39_13_18;TICK;CX_6_11_1_5_3_7_12_17_10_15_16_21_22_26_4_9_25_30_31_36_38_33_28_24_47_43_49_45_40_35_19_14_34_29_44_39_52_50_51_48
[…]
