### 2025-09-29 b54c93d
finish all the circuit variants by hand
