+++ fixed_bulk_xtop.txt
@@ -1,1 +1,0 @@
-https://algassert.com/crumble#circuit=Q(1.5,2.5)0;Q(1.5,3.5)1;Q(1.5,4.5)2;Q(1.5,5.5)3;Q(2,2)4;Q(2,3)5;Q(2,4)6;Q(2,5)7;Q(2,6)8;Q(2.5,1.5)9;Q(2.5,2.5)10;Q(2.5,3.5)11;Q(2.5,4.5)12;Q(2.5,5.5)13;Q(3,2)14;Q(3,3)15;Q(3,4)16;Q(3,5)17;Q(3,6)18;Q(3.5,1.5)19;Q(3.5,2.5)20;Q(3.5,3.5)21;Q(3.5,4.5)22;Q(3.5,5.5)23;Q(3.5,6.5)24;Q(4,2)25;Q(4,3)26;Q(4,4)27;Q(4,5)28;Q(4,6)29;Q(4.5,1.5)30;Q(4.5,2.5)31;Q(4.5,3.5)32;Q(4.5,4.5)33;Q(4.5,5.5)34;Q(5,2)35;Q(5,3)36;Q(5,4)37;Q(5,5)38;Q(5,6)39;Q(5.5,1.5)40;Q(5.5,2.5)41;Q(5.5,3.5)42;Q(5.5,4.5)43;Q(5.5,5.5)44;Q(5.5,6.5)45;Q(6,2)46;Q(6,3)47;Q(6,4)48;Q(6,5)49;Q(6,6)50;Q(6.5,2.5)51;Q(6.5,4.5)52;TICK;POLYGON(0,0,1,0.25)15_26_27_16;POLYGON(0,0,1,0.25)36_47_48_37;POLYGON(0,0,1,0.25)17_28_29_18;POLYGON(0,0,1,0.25)27_37_38_28;POLYGON(0,0,1,0.25)38_49_50_39;POLYGON(0,0,1,0.25)46_47;POLYGON(0,0,1,0.25)48_49;POLYGON(0,0,1,0.25)4_14_15_5;POLYGON(0,0,1,0.25)25_35_36_26;POLYGON(0,0,1,0.25)6_16_17_7;POLYGON(0,0,1,0.25)46_35;POLYGON(0,0,1,0.25)25_14;POLYGON(1,0,0,0.25)26_36_37_27;POLYGON(1,0,0,0.25)16_27_28_17;POLYGON(1,0,0,0.25)37_48_49_38;POLYGON(1,0,0,0.25)28_38_39_29;POLYGON(1,0,0,0.25)29_18;POLYGON(1,0,0,0.25)50_39;POLYGON(1,0,0,0.25)5_15_16_6;POLYGON(1,0,0,0.25)14_25_26_15;POLYGON(1,0,0,0.25)35_46_47_36;POLYGON(1,0,0,0.25)7_17_18;POLYGON(1,0,0,0.25)6_7;POLYGON(1,0,0,0.25)4_5;TICK;R_14_25_35_46_47_36_26_37_48_49_50;RX_4_5_15_6_16_27_38_39_28_29_18_17_7;TICK;R_19_40_51_31_10_21_42_52_33_12_23_44;RX_0_20_41_32_11_2_22_43_34_13_24_45;TICK;CX_25_19_46_40_15_10_36_31_48_42_27_21_17_12_38_33_50_44_29_23_0_5_20_26_41_47_11_16_32_37_43_49_22_28_2_7_13_18_34_39;TICK;CX_20_15_41_36_11_6_32_27_43_38_22_17_34_29_14_10_35_31_26_21_47_42_16_12_37_33_49_44_28_23;TICK;CX_5_10_26_31_37_42_16_21_7_12_28_33_49_52_39_44_18
[…]

[…]
