### 2025-09-29 600f989
add ipykernel
