+++ pyproject.toml
@@ -104,6 +104,7 @@ dev = [
     "sphinx<=8.1.3",
     "sphinxcontrib-mermaid<=1.0.0",
     "types-networkx<=3.5.0.20250610",
+    "ipykernel>=6.30.1",
 ]
 
 all = [
