### 2025-09-29 d1d335f
controllable temportal height by adjusting compile_block_graph and CubeBuilder, PipeBuilder + child classes
