### 2025-09-29 3376b67
ymemory experiment success for all circuit variants
