+++ src/tqec/compile/specs/library/generators/y_basis.py
@@ -1,0 +1,583 @@
+"""Build injection blocks for Y-basis initialization and measurement."""
+
+import functools
+from collections.abc import Callable, Iterable, Set
+from typing import Any, Literal, cast
+
+import gen
+
+from tqec.utils.enums import Basis
+
+DIRS: list[complex] = [(0.5 + 0.5j) * 1j**d for d in range(4)]
+DR, DL, UL, UR = DIRS
+# Interaction orderings that are consistent with the plaquette
+# orderings used across TQEC
+ORDER_H = [UL, UR, DL, DR]
+ORDER_V = [UL, DL, UR, DR]
+
+
+def _surface_code_patch(
+    possible_data_qubits: Iterable[complex],
+    basis: Callable[[complex], Basis],
+    is_boundary_x: Callable[[complex], bool],
+    is_boundary_z: Callable[[complex], bool],
+    order_func: Callable[[complex], Iterable[complex | None]],
+) -> gen.Patch:
+    """Make a surface code patch."""
+    possible_data_qubits = set(possible_data_qubits)
+    possible_measure_qubits = {q + d for q in possible_data_qubits for d in DIRS}
+    measure_qubits = {
+        m
+        for m in possible_measure_qubits
+        if sum(m + d in possible_data_qubits for d in DIRS) > 1
+        if is_boundary_x(m) <= (basis(m) == Basis.X)
+        if is_boundary_z(m) <= (basis(m) == Basis.Z)
+    }
+    data_qubits = {
+        q for q in possible_data_qubits if sum(q + d in measure_qubits for d in DIRS) > 1
+    }
+
+    tiles = [
+        gen.Tile(
+            bases=str(basis(m)),
+            data_qubits=[
+                m + d if d is not None and m + d in data_qubits else None for d in order_func(m)
+            ],
+            measure_qubit=m,
+        )
+        for m in measure_qubits
+    ]
+    return gen.Patch(tiles)
+
+
+def _checkerboard_basis(q: complex, top_left_tile_basis: Basis) -> Basis:
+    """Classifies a coordinate as X type or Z type according to a checkerboard."""
+    if int(q.real + q.imag) & 1 == 0:
+        return top_left_tile_basis.flipped()
+    return top_left_tile_basis
+
+
+def _rectangular_surface_code_patch(
+    width: int,
+    height: int,
+    top_left_tile_basis: Basis,
+    top_basis: Basis,
+    bot_basis: Basis,
+    left_basis: Basis,
+    right_basis: Basis,
+    order_func: Callable[[complex], Iterable[complex | None]],
+) -> gen.Patch:
+    """Make a rectangular surface code patch."""
+
+    def is_boundary(m: complex, *, b: Basis) -> bool:
+        if top_basis == b and m.imag == -0.5:
+            return True
+        if left_basis == b and m.real == -0.5:
+            return True
+        if bot_basis == b and m.imag == height - 0.5:
+            return True
+        if right_basis == b and m.real == width - 0.5:
+            return True
+        return False
+
+    return _surface_code_patch(
+        possible_data_qubits=[x + 1j * y for x in range(width) for y in range(height)],
+        basis=lambda q: _checkerboard_basis(q, top_left_tile_basis),
+        is_boundary_x=functools.partial(is_boundary, b=Basis.X),
+        is_boundary_z=functools.partial(is_boundary, b=Basis.Z),
+        order_func=order_func,
+    )
+
+
+def _get_top_left_tile_basis(
+    top_boundary_basis: Basis,
+    convention: Literal["fixed_boundary", "fixed_bulk"],
+) -> Basis:
+    """Get the basis of the top-left tile given the top boundary basis and convention."""
+    if convention == "fixed_boundary" and top_boundary_basis == Basis.X:
+        return Basis.X
[…]
