+++ src/tqec/compile/specs/library/generators/y_basis_init_meas.py
@@ -1,565 +1,0 @@
-"""Factory for Y basis initialization and measurement circuit."""
-
-import functools
-from collections.abc import Callable, Iterable, Set
-from typing import Any, Literal
-
-import gen
-
-from tqec.compile.blocks.injected_block import Alignment, CircuitWithInterface, InjectedBlock
-from tqec.compile.specs.base import YHalfCubeSpec
-from tqec.utils.scale import LinearFunction, PhysicalQubitScalable2D
-
-DIRS: list[complex] = [(0.5 + 0.5j) * 1j**d for d in range(4)]
-DR, DL, UL, UR = DIRS
-# ORDER_S = [UL, UR, DL, DR]
-# ORDER_N = [UL, DL, UR, DR]
-ORDER_S = [UR, UL, DR, DL]
-ORDER_N = [UR, DR, UL, DL]
-
-
-def surface_code_patch(
-    possible_data_qubits: Iterable[complex],
-    basis: Callable[[complex], str],
-    is_boundary_x: Callable[[complex], bool],
-    is_boundary_z: Callable[[complex], bool],
-    order_func: Callable[[complex], Iterable[complex | None]],
-) -> gen.Patch:
-    """Make a surface code patch."""
-    possible_data_qubits = set(possible_data_qubits)
-    possible_measure_qubits = {q + d for q in possible_data_qubits for d in DIRS}
-    measure_qubits = {
-        m
-        for m in possible_measure_qubits
-        if sum(m + d in possible_data_qubits for d in DIRS) > 1
-        if is_boundary_x(m) <= (basis(m) == "X")
-        if is_boundary_z(m) <= (basis(m) == "Z")
-    }
-    data_qubits = {
-        q for q in possible_data_qubits if sum(q + d in measure_qubits for d in DIRS) > 1
-    }
-
-    tiles = [
-        gen.Tile(
-            bases=basis(m),
-            data_qubits=[
-                m + d if d is not None and m + d in data_qubits else None for d in order_func(m)
-            ],
-            measure_qubit=m,
-        )
-        for m in measure_qubits
-    ]
-    return gen.Patch(tiles)
-
-
-def checkerboard_basis(q: complex) -> str:
-    """Classifies a coordinate as X type or Z type according to a checkerboard."""
-    is_x = int(q.real + q.imag) & 1 == 0
-    return "X" if is_x else "Z"
-
-
-def rectangular_surface_code_patch(
-    width: int,
-    height: int,
-    top_basis: str,
-    bot_basis: str,
-    left_basis: str,
-    right_basis: str,
-    order_func: Callable[[complex], Iterable[complex | None]],
-) -> gen.Patch:
-    """Make a rectangular surface code patch."""
-
-    def is_boundary(m: complex, *, b: str) -> bool:
-        if top_basis == b and m.imag == -0.5:
-            return True
-        if left_basis == b and m.real == -0.5:
-            return True
-        if bot_basis == b and m.imag == height - 0.5:
-            return True
-        if right_basis == b and m.real == width - 0.5:
-            return True
-        return False
-
-    return surface_code_patch(
-        possible_data_qubits=[x + 1j * y for x in range(width) for y in range(height)],
-        basis=checkerboard_basis,
-        is_boundary_x=functools.partial(is_boundary, b="X"),
-        is_boundary_z=functools.partial(is_boundary, b="Z"),
-        order_func=order_func,
-    )
-
-
-def make_xtop_qubit_patch(distance: int) -> gen.Patch:
-    """Make a surface code patch with X boundaries on top and bottom edges."""
-
-    def order_func(m: complex) -> list[complex]:
-        if checkerboard_basis(m) == "X":
-            return ORDER_S
-        else:
-            return ORDER_N
-
-    return rectangular_surface_code_patch(
-        width=distance,
[…]
