### 2025-09-30 92859b2
Add explicit mention of venv creation in installation steps
