### 2025-09-30 7389249
Remove num_workers=20 from sinter call, probably a typo
