### 2025-09-30 08b756d
Move `PauliBasis` and `ExtendedBasis` enums into `tqec.utils.enums`
