### 2025-09-30 1b3fe8b
Add S gate teleportation to gallery
