+++ src/tqec/gallery/__init__.py
@@ -17,6 +17,7 @@ from .cnot import cnot as cnot
 from .cz import cz as cz
 from .memory import memory as memory
 from .move_rotation import move_rotation as move_rotation
+from .s_gate_teleportation import s_gate_teleportation as s_gate_teleportation
 from .stability import stability as stability
 from .steane_encoding import steane_encoding as steane_encoding
 from .three_cnots import three_cnots as three_cnots
