### 2025-09-30 2376a4b
Use uppercase for `PauliBasis` enum values
