### 2025-10-01 dabf8cd
Update python-sat[cryptosat] requirement from <=1.8.dev22 to <=1.8.dev23 (#710)
Updates the requirements on
[python-sat[cryptosat]](https://github.com/pysathq/pysat) to permit the
latest version.
<details>
<summary>Commits</summary>
<ul>
<li>See full diff in <a
href="https://github.com/pysathq/pysat/commits">compare view</a></li>
</ul>
</details>
<br />


Dependabot will resolve any conflicts with this PR as long as you don't
alter it yourself. You can also trigger a rebase manually by commenting
`@dependabot rebase`.

[//]: # (dependabot-automerge-start)
[//]: # (dependabot-automerge-end)

--
[…]
