+++ pyproject.toml
@@ -42,7 +42,7 @@ dependencies = [
     "platformdirs<=4.4.0",
     "pycollada<=0.9.1",
     "pycryptosat<=5.11.21",
-    "python-sat[cryptosat]<=1.8.dev22",
+    "python-sat[cryptosat]<=1.8.dev23",
     "pyzx<=0.9.0",
     "scipy<=1.15.3",
     "semver<=3.0.4",
