### 2025-10-01 ab7f183
Implement y-basis circuit with `gen`
