+++ pull/714/_sources/gallery/memory.ipynb.txt
@@ -1,0 +1,304 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Memory\n",
+    "\n",
+    "This notebook demonstrates the simplest(trivial) form of computation: logical memory. More specifically, a logical qubit with distance $d$ remains idle for $d$ error correction cycles [<cite data-footcite-t=\"Acharya_2024\"></cite>]."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Construction\n",
+    "\n",
+    "A memory experiment can be represented with a single cube. The color of the cube determines the spatial/temporal boundary types.\n",
+    "\n",
+    "`tqec` provides builtin functions `tqec.gallery.memory` to construct it."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "from tqec import Basis\n",
+    "from tqec.gallery import memory\n",
+    "\n",
+    "graph = memory(Basis.Z)\n",
+    "graph.view_as_html()"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "Memory experiment preserves the logical observable through time:"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "correlation_surfaces = graph.find_correlation_surfaces()\n",
+    "graph.view_as_html(\n",
+    "    pop_faces_at_directions=(\"-Y\",),\n",
+    "    show_correlation_surface=correlation_surfaces[0],\n",
+    ")"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Circuit\n",
+    "\n",
[…]
