+++ pull/714/_sources/gallery/cnot.ipynb.txt
@@ -1,369 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "# CNOT\n",
-    "\n",
-    "This notebook shows the construction and simulation results of the logical CNOT gate between two logical qubits with lattice surgery."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "## Construction\n",
-    "\n",
-    "A logical CNOT between two logical qubits can be implemented with the help of an ancilla qubit. It can be accomplished by the following steps:\n",
-    "\n",
-    "1. $M_{ZZ}$ parity measurement between $Q_{control}$ and $Q_{ancilla}$.\n",
-    "2. $M_{XX}$ parity measurement between $Q_{target}$ and $Q_{ancilla}$.\n",
-    "3. $M_{Z}$ measurement of $Q_{ancilla}$.\n",
-    "\n",
-    "`tqec` provides builtin functions `tqec.gallery.cnot` to construct the logical CNOT gate."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "from tqec.gallery import cnot\n",
-    "\n",
-    "graph = cnot()\n",
-    "graph.view_as_html()"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "The logical CNOT has four independent stabilizer flow generators: `XI -> XX`, `IX -> IX`, `ZI -> ZI`, `IZ -> ZZ`. Here we show the correlation surfaces for the generators."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "correlation_surfaces = graph.find_correlation_surfaces()"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "#### `XI -> XX`"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "graph.view_as_html(\n",
-    "    pop_faces_at_directions=(\"-Y\",),\n",
-    "    show_correlation_surface=correlation_surfaces[0],\n",
-    ")"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
[…]
