### 2025-10-01 #8930 @lockwo
Hi 👋 @burgholzer, @meandmytram, @peter-janderks, would one (or more) of you be willing to review this submission for JOSS?
