### 2025-10-01 #8930 @burgholzer
> Hi 👋 @burgholzer, @meandmytram, @peter-janderks, would one (or more) of you be willing to review this submission for JOSS?

Sure 👍🏻 not explicitly my domain of expertise within quantum, but close enough. I have also been loosely following the developments of the library.
