+++ src/tqec/compile/blocks/block.py
@@ -21,8 +21,8 @@ from tqec.utils.exceptions import TQECError
 from tqec.utils.scale import LinearFunction, PhysicalQubitScalable2D
 
 
-class Block(SequencedLayers):
-    """Encodes the implementation of a block.
+class LayeredBlock(SequencedLayers):
+    """Encodes the implementation of a block with a sequence of layers.
 
     This data structure is voluntarily very generic. It represents blocks as a
     sequence of layers that can be instances of either
@@ -36,8 +36,8 @@ class Block(SequencedLayers):
     """
 
     @override
-    def with_spatial_borders_trimmed(self, borders: Iterable[SpatialBlockBorder]) -> Block:
-        return Block(
+    def with_spatial_borders_trimmed(self, borders: Iterable[SpatialBlockBorder]) -> LayeredBlock:
+        return LayeredBlock(
             self._layers_with_spatial_borders_trimmed(borders),
             self.trimmed_spatial_borders | frozenset(borders),
         )
@@ -46,11 +46,11 @@ class Block(SequencedLayers):
     def with_temporal_borders_replaced(
         self,
         border_replacements: Mapping[TemporalBlockBorder, BaseLayer | None],
-    ) -> Block | None:
+    ) -> LayeredBlock | None:
         if not border_replacements:
             return self
         layers = self._layers_with_temporal_borders_replaced(border_replacements)
-        return Block(layers) if layers else None
+        return LayeredBlock(layers) if layers else None
 
     def get_atomic_temporal_border(self, border: TemporalBlockBorder) -> BaseLayer:
         """Get the layer at the provided temporal ``border``.
@@ -118,14 +118,14 @@ class Block(SequencedLayers):
         return self.is_pipe and self.dimensions[2].is_constant()
 
     def __eq__(self, value: object) -> bool:
-        return isinstance(value, Block) and super().__eq__(value)
+        return isinstance(value, LayeredBlock) and super().__eq__(value)
 
     def __hash__(self) -> int:
         raise NotImplementedError(f"Cannot hash efficiently a {type(self).__name__}.")
 
 
 def merge_parallel_block_layers(
-    blocks_in_parallel: Mapping[LayoutPosition2D, Block],
+    blocks_in_parallel: Mapping[LayoutPosition2D, LayeredBlock],
     scalable_qubit_shape: PhysicalQubitScalable2D,
 ) -> list[LayoutLayer | BaseComposedLayer]:
     """Merge several stacks of layers executed in parallel into one stack of larger layers.
