+++ src/tqec/compile/specs/library/fixed_bulk.py
@@ -3,7 +3,7 @@ from typing import Final
 
 from typing_extensions import override
 
-from tqec.compile.blocks.block import Block
+from tqec.compile.blocks.block import LayeredBlock
 from tqec.compile.blocks.layers.atomic.base import BaseLayer
 from tqec.compile.blocks.layers.atomic.plaquettes import PlaquetteLayer
 from tqec.compile.blocks.layers.composed.base import BaseComposedLayer
@@ -78,7 +78,7 @@ class FixedBulkCubeBuilder(CubeBuilder):
         )
 
     @override
-    def __call__(self, spec: CubeSpec) -> Block:
+    def __call__(self, spec: CubeSpec) -> LayeredBlock:
         kind = spec.kind
         if isinstance(kind, Port):
             raise TQECError("Cannot build a block for a Port.")
@@ -91,7 +91,7 @@ class FixedBulkCubeBuilder(CubeBuilder):
             RepeatedLayer(PlaquetteLayer(template, repeat), repetitions=_DEFAULT_BLOCK_REPETITIONS),
             PlaquetteLayer(template, measure),
         ]
-        return Block(layers)
+        return LayeredBlock(layers)
 
 
 class FixedBulkPipeBuilder(PipeBuilder):
@@ -110,7 +110,7 @@ class FixedBulkPipeBuilder(PipeBuilder):
         self._generator = FixedBulkConventionGenerator(translator, compiler)
 
     @override
-    def __call__(self, spec: PipeSpec) -> Block:
+    def __call__(self, spec: PipeSpec) -> LayeredBlock:
         if spec.pipe_kind.is_temporal:
             return self._get_temporal_pipe_block(spec)
         return self._get_spatial_pipe_block(spec)
@@ -119,7 +119,7 @@ class FixedBulkPipeBuilder(PipeBuilder):
     #    TEMPORAL PIPE    #
     #######################
 
-    def _get_temporal_pipe_block(self, spec: PipeSpec) -> Block:
+    def _get_temporal_pipe_block(self, spec: PipeSpec) -> LayeredBlock:
         """Return the block to implement a temporal pipe based on the provided ``spec``.
 
         Args:
@@ -140,7 +140,7 @@ class FixedBulkPipeBuilder(PipeBuilder):
         # Else, it is a regular temporal junction
         return self._get_temporal_non_hadamard_pipe_block(spec)
 
-    def _get_temporal_non_hadamard_pipe_block(self, spec: PipeSpec) -> Block:
+    def _get_temporal_non_hadamard_pipe_block(self, spec: PipeSpec) -> LayeredBlock:
         """Return the block to implement a regular temporal junction without Hadamard transition.
 
         Args:
@@ -166,14 +166,14 @@ class FixedBulkPipeBuilder(PipeBuilder):
             z_observable_orientation, None, None
         )
         template = self._generator.get_memory_qubit_raw_template()
-        return Block(
+        return LayeredBlock(
             [
                 PlaquetteLayer(template, memory_plaquettes)
                 for _ in range(3 if spec.at_temporal_hadamard_layer else 2)
             ]
         )
 
-    def _get_temporal_hadamard_pipe_block(self, spec: PipeSpec) -> Block:
+    def _get_temporal_hadamard_pipe_block(self, spec: PipeSpec) -> LayeredBlock:
         """Return the block to implement a temporal Hadamard pipe.
 
         Note:
@@ -209,7 +209,7 @@ class FixedBulkPipeBuilder(PipeBuilder):
             z_observable_orientation.flip(), None, None
         )
         template = self._generator.get_temporal_hadamard_raw_template()
-        return Block(
+        return LayeredBlock(
             [
[…]
