+++ src/tqec/interop/collada/read_write.py
@@ -262,7 +262,8 @@ def write_block_graph_to_dae_file(
     base.mesh.write(file_like)
 
 
-# JSON IMPORTER
+# JSON IMPORTER USING WEBGL-LIKE COORDINATES
+# Potentially useful for other WEBGL tools besides SketchUp
 # Why no JSON exporter? No need. It can be done with blockgraph.to_dict
 def read_block_graph_from_json(
     filepath: str | pathlib.Path,
@@ -411,6 +412,119 @@ def read_block_graph_from_json(
         graph.add_pipe(head_pos, tail_pos, pipe_kind)
 
     return graph
+
+
+# IMPORTER FOR DICTIONARY-BASED SPACE-TIME DIAGRAMS USING WEBGL-COMPATIBLE METHODS
+def read_from_lattice_dicts(
+    lattice_nodes: dict[int, tuple[tuple[int, int, int], str]],
+    lattice_edges: dict[tuple[int, int], str],
+    graph_name: str = "",
+) -> BlockGraph:
+    """Construct a :class:`.BlockGraph` from a space-time diagram given as a 3D-graph dictionary.
+
+    Args:
+        lattice_nodes: {id: ((x, y, z), kind)} for each node (cube) in intended blockgraph
+        lattice_edges: {(source_id, target_id): kind} for each edge (pipes) in intended blockgraph
+        graph_name: The name of the block graph. Default is an empty string.
+
+
+    Returns:
+        The constructed :py:class:`~tqec.computation.block_graph.BlockGraph` object.
+
+    Raises:
+        ValueError: If incoming objects cannot be parsed and converted to a block graph.
+
+    """
+    # Reject malforned inputs
+    if not lattice_nodes.values():
+        raise ValueError(
+            "Lattice surgery objects not appropriately constructed: no nodes/cubes detected"
+        )
+
+    if not lattice_edges.values():
+        raise ValueError(
+            "Lattice surgery objects not appropriately constructed: node edges/pipes detected"
+        )
+
+    # Helper variables
+    pipe_length: float | None = None
+    parsed_ports: list[FloatPosition3D] = []
+    parsed_cubes: list[tuple[FloatPosition3D, CubeKind]] = []
+    parsed_pipes: list[tuple[FloatPosition3D, PipeKind, int]] = []
+
+    # Unpack nodes/cubes
+    try:
+        for v in lattice_nodes.values():
+            coords = v[0]
+            translation = FloatPosition3D(*coords)
+            if v[1] != "ooo":
+                kind = block_kind_from_str(v[1].upper())
+                if isinstance(kind, CubeKind):
+                    parsed_cubes.append((translation, kind))
+
+            else:
+                parsed_ports.append(translation)
+    except (ValueError, TypeError, IndexError, KeyError) as e:
+        raise ValueError("Error reading nodes/cubes from lattice_nodes dictionary:", e)
+
+    try:
+        for (src, tgt), v in lattice_edges.items():
+            kind = block_kind_from_str(v.upper())
+            if isinstance(kind, PipeKind):
+                src_pos = lattice_nodes[src][0]
+                tgt_pos = lattice_nodes[tgt][0]
+
+                shift_coords_from_src = tuple([(u - v) / 3 for u, v in zip(tgt_pos, src_pos)])
+                directional_multiplier = int(sum(shift_coords_from_src))
+
+                coords = [u + v for u, v in zip(src_pos, shift_coords_from_src)]
+                translation = FloatPosition3D(*coords)
+
+                parsed_pipes.append((translation, kind, directional_multiplier))
+    except (ValueError, TypeError, IndexError, KeyError) as e:
+        raise ValueError("Error reading edges/pipes from lattice_edges dictionary:", e)
+
+    # Construct graph
+    # Create graph
[…]
