### 2025-10-01 3e65cb9
Fix the issues resulted from uppercasing the `PauliBasis` and `ExtendedBasis` values
