### 2025-10-01 7e17f0a
Introduces `InjectedBlock` type and `Block` base type
