+++ src/tqec/compile/blocks/block.py
@@ -1,11 +1,15 @@
 from __future__ import annotations
 
+from abc import ABC, abstractmethod
 from collections.abc import Iterable, Mapping
-from typing import Final
+from dataclasses import dataclass, field
+from typing import Final, Protocol
 
+import gen
+import stim
 from typing_extensions import override
 
-from tqec.compile.blocks.enums import SpatialBlockBorder, TemporalBlockBorder
+from tqec.compile.blocks.enums import Alignment, SpatialBlockBorder, TemporalBlockBorder
 from tqec.compile.blocks.layers.atomic.base import BaseLayer
 from tqec.compile.blocks.layers.atomic.layout import LayoutLayer
 from tqec.compile.blocks.layers.composed.base import BaseComposedLayer
@@ -21,7 +25,39 @@ from tqec.utils.exceptions import TQECError
 from tqec.utils.scale import LinearFunction, PhysicalQubitScalable2D
 
 
-class LayeredBlock(SequencedLayers):
+class Block(ABC):
+    """Base class for all block instantiations in the compilation framework.
+
+    A block represents a unit of quantum computation with a defined spatial
+    and temporal footprint that scales with the scaling factor ``k``.
+    """
+
+    @property
+    @abstractmethod
+    def scalable_timesteps(self) -> LinearFunction:
+        """Get the scalable timesteps (temporal extent) of the block."""
+        pass
+
+    @property
+    @abstractmethod
+    def scalable_shape(self) -> PhysicalQubitScalable2D:
+        """Get the scalable shape (spatial extent) of the block."""
+        pass
+
+    @property
+    @abstractmethod
+    def is_cube(self) -> bool:
+        """Return ``True`` if ``self`` represents a cube, else ``False``."""
+        pass
+
+    @property
+    @abstractmethod
+    def is_pipe(self) -> bool:
+        """Return ``True`` if ``self`` represents a pipe, else ``False``."""
+        pass
+
+
+class LayeredBlock(SequencedLayers, Block):
     """Encodes the implementation of a block with a sequence of layers.
 
     This data structure is voluntarily very generic. It represents blocks as a
@@ -179,3 +215,111 @@ def merge_parallel_block_layers(
                 f"\n{list(layers.values())}"
             )
     return merged_layers
+
+
+@dataclass(frozen=True)
+class CircuitWithInterface:
+    """A quantum circuit with its expected flow interface.
+
+    Attributes:
+        circuit: the quantum circuit.
+        det_interface: the expected detector flows interface of the circuit.
+        obs_interface: the expected observable flow interface of the circuit.
+
+    """
+
+    circuit: stim.Circuit
+    det_interface: gen.ChunkInterface = field(default_factory=lambda: gen.ChunkInterface(()))
+    obs_interface: gen.ChunkInterface = field(default_factory=lambda: gen.ChunkInterface(()))
+
+    def __post_init__(self) -> None:
+        if not self.circuit:
+            raise TQECError("The provided circuit is empty.")
+        if len(self.obs_interface.ports) > 1:
+            raise TQECError("At most one observable flow is supported at each injection block.")
+
+
+class InjectionFactory(Protocol):
+    def __call__(self, k: int) -> CircuitWithInterface:
+        """Generate the quantum circuit with expected flow interface from the scaling factor.
+
+        Args:
+            k: The scaling factor.
+
+        Returns:
+            A quantum circuit with expected flow interface.
+
+        """
+        ...
+
+
+class InjectedBlock(Block):
+    """Represent temporally injected block like ``YHalfCube``.
+
+    Some blocks like ``YHalfCube`` are connected to the computation in time
+    direction and do not have timestep schedules that are consistent with
[…]
