### 2025-10-01 0bef212
Add spec information for `YHalfCube`
