+++ src/tqec/compile/specs/base.py
@@ -3,23 +3,63 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Literal, Protocol, cast
 
-from tqec.compile.blocks.block import LayeredBlock
+from tqec.compile.blocks.block import Block, LayeredBlock
 from tqec.compile.specs.enums import SpatialArms
 from tqec.computation.block_graph import BlockGraph
-from tqec.computation.cube import Cube, CubeKind, ZXCube
+from tqec.computation.cube import Cube, CubeKind, YHalfCube, ZXCube
 from tqec.computation.pipe import PipeKind
 from tqec.templates.base import RectangularTemplate
+from tqec.utils.enums import Basis
 from tqec.utils.exceptions import TQECError
 from tqec.utils.position import Direction3D
+
+
+@dataclass(frozen=True)
+class YHalfCubeSpec:
+    """Specification of a Y half cube in a block graph.
+
+    Attributes:
+        initialization: Flag indicating whether the half cube is for Y-basis
+            initialization or measurement.
+        horizontal_boundary_basis: The basis of the horizontal boundary of the
+            pipe connecting to the cube.
+
+    """
+
+    initialization: bool
+    horizontal_boundary_basis: Basis
+
+    @staticmethod
+    def from_cube(
+        cube: Cube,
+        graph: BlockGraph,
+    ) -> YHalfCubeSpec:
+        """Return the spec from a Y half cube in a block graph."""
+        pos = cube.position
+        pos_up, pos_down = pos.shift_by(dz=1), pos.shift_by(dz=-1)
+        connect_up = graph.has_pipe_between(pos, pos_up)
+        connect_down = graph.has_pipe_between(pos, pos_down)
+        assert connect_down + connect_up == 1
+        if connect_up:
+            pipe = graph.get_pipe(pos, pos_up)
+            at_head = True
+        else:
+            pipe = graph.get_pipe(pos, pos_down)
+            at_head = False
+
+        horizontal_boundary_basis = pipe.kind.get_basis_along(Direction3D.Y, at_head)
+        assert horizontal_boundary_basis is not None
+        return YHalfCubeSpec(
+            initialization=connect_up,
+            horizontal_boundary_basis=horizontal_boundary_basis,
+        )
 
 
 @dataclass(frozen=True)
 class CubeSpec:
     """Specification of a cube in a block graph.
 
-    The template of the `CompiledBlock` will be determined based on the specification.
-    This class can be used as a key to look up the corresponding `CompiledBlock` before
-    applying the substitution rules.
+    The block instantiation will be determined based on the specification.
 
     Attributes:
         cube_kind: The kind of the cube.
@@ -30,12 +70,15 @@ class CubeSpec:
             pipe at the top or bottom of a spatial cube is executed on the same
             timeslice as this cube. This information is needed for the fixed
             boundary convention.
+        y_half_cube_spec: If the cube is a Y half cube, this attribute contains
+            the specification of the half cube. Otherwise, it is ``None``.
 
     """
 
     kind: CubeKind
     spatial_arms: SpatialArms = SpatialArms.NONE
     has_spatial_up_or_down_pipe_in_timeslice: bool = False
+    y_half_cube_spec: YHalfCubeSpec | None = None
 
     def __post_init__(self) -> None:
         if self.spatial_arms != SpatialArms.NONE:
@@ -44,11 +87,21 @@ class CubeSpec:
                     "The `spatial_arms` attribute should be `SpatialArms.NONE` "
                     "for non-spatial cubes."
                 )
[…]
