### 2025-10-01 67b4398
Implement Y basis block generator
