+++ src/tqec/compile/specs/library/fixed_boundary.py
@@ -1,7 +1,7 @@
 from collections.abc import Callable
 from typing import Final, Protocol
 
-from tqec.compile.blocks.block import LayeredBlock
+from tqec.compile.blocks.block import Block, LayeredBlock
 from tqec.compile.blocks.layers.atomic.base import BaseLayer
 from tqec.compile.blocks.layers.atomic.plaquettes import PlaquetteLayer
 from tqec.compile.blocks.layers.composed.base import BaseComposedLayer
@@ -10,6 +10,7 @@ from tqec.compile.blocks.layers.composed.sequenced import SequencedLayers
 from tqec.compile.specs.base import CubeBuilder, CubeSpec, PipeBuilder, PipeSpec
 from tqec.compile.specs.enums import SpatialArms
 from tqec.compile.specs.library.generators.fixed_boundary import FixedBoundaryConventionGenerator
+from tqec.compile.specs.library.generators.y_basis import get_y_half_cube_block
 from tqec.computation.cube import Port, YHalfCube, ZXCube
 from tqec.plaquette.compilation.base import IdentityPlaquetteCompiler, PlaquetteCompiler
 from tqec.plaquette.plaquette import Plaquettes
@@ -157,13 +158,15 @@ class FixedBoundaryCubeBuilder(CubeBuilder):
             _spatial_plaquettes_generator,
         )
 
-    def __call__(self, spec: CubeSpec) -> LayeredBlock:
+    def __call__(self, spec: CubeSpec) -> Block:
         """Instantiate a :class:`.Block` instance implementing the provided ``spec``."""
         kind = spec.kind
         if isinstance(kind, Port):
             raise TQECError("Cannot build a block for a Port.")
         elif isinstance(kind, YHalfCube):
-            raise NotImplementedError("Y cube is not implemented.")
+            y_spec = spec.y_half_cube_spec
+            assert y_spec is not None
+            return get_y_half_cube_block(y_spec, "fixed_boundary")
         template, pgen = self._get_template_and_plaquettes_generator(spec)
         return _get_block(
             z_basis=kind.z,
