### 2025-10-01 a09b842
Enable adding `InjectedBlock` instances to `TopologicalComputationGraph`
