+++ src/tqec/compile/graph.py
@@ -48,7 +48,12 @@ from typing import Final
 
 import stim
 
-from tqec.compile.blocks.block import LayeredBlock, merge_parallel_block_layers
+from tqec.compile.blocks.block import (
+    Block,
+    InjectedBlock,
+    LayeredBlock,
+    merge_parallel_block_layers,
+)
 from tqec.compile.blocks.enums import (
     SpatialBlockBorder,
     TemporalBlockBorder,
@@ -115,7 +120,8 @@ class TopologicalComputationGraph:
         observables: list[AbstractObservable] | None = None,
     ) -> None:
         """Represent a topological computation with :class:`.Block` instances."""
-        self._blocks: dict[LayoutPosition3D, LayeredBlock] = {}
+        self._layered_blocks: dict[LayoutPosition3D, LayeredBlock] = {}
+        self._injected_blocks: dict[BlockPosition3D, InjectedBlock] = {}
         # For fixed-bulk convention, temporal Hadamard pipe has its on space-time
         # extent. We need to keep track of the temporal pipes that are at the
         # same layer of at least one temporal Hadamard pipe.
@@ -126,24 +132,40 @@ class TopologicalComputationGraph:
         self._observables: list[AbstractObservable] | None = observables
         self._observable_builder = observable_builder
 
-    def add_cube(self, position: BlockPosition3D, block: LayeredBlock) -> None:
+    @property
+    def layout_positions(self) -> set[LayoutPosition3D]:
+        """Get all the positions at which a block has been added."""
+        return set(self._layered_blocks.keys()).union(
+            {LayoutPosition3D.from_block_position(p) for p in self._injected_blocks.keys()}
+        )
+
+    def add_cube(self, position: BlockPosition3D, block: Block) -> None:
         """Add a new cube at ``position`` implemented by the provided ``block``."""
+        self._check_block_spatial_shape(block)
+        if isinstance(block, InjectedBlock):
+            if position in self._injected_blocks:
+                raise TQECError(
+                    "Cannot override an injected block with ``add_cube``. There is already "
+                    f"an injected block at {position}."
+                )
+            self._injected_blocks[position] = block
+            return
+        assert isinstance(block, LayeredBlock)
+        layout_position = LayoutPosition3D.from_block_position(position)
         if not block.is_cube:
             raise TQECError(
                 f"Cannot add the block as a cube. The provided block({block}) "
                 "has at least one non-scalable dimension."
             )
-        self._check_block_spatial_shape(block)
-        layout_position = LayoutPosition3D.from_block_position(position)
-        if layout_position in self._blocks:
+        if layout_position in self._layered_blocks:
             raise TQECError(
                 "Cannot override a block with ``add_cube``. There is already "
                 f"an entry at {layout_position}."
             )
-        self._blocks[layout_position] = block
+        self._layered_blocks[layout_position] = block
 
-    def get_cube(self, position: BlockPosition3D) -> LayeredBlock:
-        """Recover the :class:`.Block` instance at the provided ``position``.
+    def get_layered_block(self, position: BlockPosition3D) -> LayeredBlock:
+        """Recover the :class:`.LayeredBlock` instance at the provided ``position``.
 
         Args:
             position: position of the block to recover.
@@ -152,11 +174,11 @@ class TopologicalComputationGraph:
             KeyError: if the provided ``position`` has no block associated.
 
         Returns:
-            the :class:`.Block` instance at the provided ``position``.
+            the :class:`.LayeredBlock` instance at the provided ``position``.
 
         """
[…]
