### 2025-10-01 715ddf4
Do not shift the block graph's z position and attach layer's z information to the layer tree with metadata stored in `SequencedLayer`
