+++ src/tqec/compile/observables/abstract_observable.py
@@ -113,6 +113,8 @@ class AbstractObservable:
             by the pipe might be included in the logical observable. It is only
             relevant for the fixed-bulk convention, where the temporal Hadamard
             includes a realignment layer of stabilizers.
+        y_half_cubes: A set of Y half cubes of which some measurements should be
+            included in the observable because Y-basis initialization or measurement.
 
     """
 
@@ -121,6 +123,7 @@ class AbstractObservable:
     bottom_stabilizer_cubes: frozenset[CubeWithArms] = frozenset()
     bottom_stabilizer_pipes: frozenset[PipeWithArms] = frozenset()
     temporal_hadamard_pipes: frozenset[PipeWithObservableBasis] = frozenset()
+    y_half_cubes: frozenset[Cube] = frozenset()
 
     def slice_at_z(self, z: int) -> AbstractObservable:
         """Get the observable slice at the given z position."""
@@ -130,6 +133,7 @@ class AbstractObservable:
             frozenset(c for c in self.bottom_stabilizer_cubes if c.cube.position.z == z),
             frozenset(p for p in self.bottom_stabilizer_pipes if p.pipe.u.position.z == z),
             frozenset(p for p in self.temporal_hadamard_pipes if p.pipe.u.position.z == z),
+            frozenset(c for c in self.y_half_cubes if c.position.z == z),
         )
 
 
@@ -231,6 +235,7 @@ def compile_correlation_surface_to_abstract_observable(
     top_readout_pipes: set[PipeWithArms] = set()
     bottom_stabilizer_pipes: set[PipeWithArms] = set()
     temporal_hadamard_pipes: set[PipeWithObservableBasis] = set()
+    y_half_cubes: set[Cube] = set()
 
     # 1. Handle spatial cubes top readouts
     for node in correlation_surface.span_vertices():
@@ -272,8 +277,9 @@ def compile_correlation_surface_to_abstract_observable(
     # 2. Handle all the pipes
     def has_obs_include(cube: Cube, correlation: Basis) -> bool:
         """Check if the top data qubit readout should be included in the observable."""
+        # Y cubes should have been handled
         if cube.is_y_cube:
-            return True
+            return False
         assert isinstance(cube.kind, ZXCube)
         # No pipe at the top
         if block_graph.has_pipe_between(cube.position, cube.position.shift_by(0, 0, 1)):
@@ -290,6 +296,9 @@ def compile_correlation_surface_to_abstract_observable(
             # during realignment of plaquettes under fixed-bulk convention
             if include_temporal_hadamard_pipes and pipe.kind.has_hadamard:
                 temporal_hadamard_pipes.add(PipeWithObservableBasis(pipe, edge.u.basis))
+            for cube in pipe:
+                if cube.is_y_cube:
+                    y_half_cubes.add(cube)
             if has_obs_include(pipe.v, edge.v.basis):
                 top_readout_cubes.add(CubeWithArms(pipe.v))
             continue
@@ -323,6 +332,7 @@ def compile_correlation_surface_to_abstract_observable(
         top_readout_pipes=frozenset(top_readout_pipes),
         bottom_stabilizer_pipes=frozenset(bottom_stabilizer_pipes),
         temporal_hadamard_pipes=frozenset(temporal_hadamard_pipes),
+        y_half_cubes=frozenset(y_half_cubes),
     )
 
 
@@ -358,7 +368,7 @@ def _check_correlation_surface_validity(correlation_surface: CorrelationSurface,
                 )
             continue
         v_basis = Basis.Z if is_z_no_phase(g, v) else Basis.X
-        if counts[v_basis.flipped()] not in [0, len(g.incident_edges(v))]:  # type: ignore
[…]
