### 2025-10-01 d54f927
Fix left PauliBasis uppercase issue
