+++ src/tqec/compile/tree/annotations.py
@@ -64,7 +64,7 @@ class Polygon:
         """Convert the polygon to the representation in a crumble url."""
         # default grey color for polygons with no basis information
         rgba = [0, 0, 0, 0.25]
-        rgba["xyz".index(self.basis.value)] = 1
+        rgba["XYZ".index(self.basis.value)] = 1
         rgba_str = ",".join(str(i) for i in rgba)
         qubits_idx = [qubit_map[q] for q in self._sorted_qubits()]
         qubits_str = "_".join(str(i) for i in qubits_idx)
