### 2025-10-01 2263bc0
Deduplicate position z values to fix tests
