### 2025-10-02 55a1d02
Fix abstract observable test
