### 2025-10-02 2496e23
NB! -> hint VENV & explicit multi-thread count on logical_cnot
