### 2025-10-02 9695fbf
Update docs/contributor_guide.rst
Co-authored-by: Purva Thakre <66048318+purva-thakre@users.noreply.github.com>
