### 2025-10-02 a63e41b
Update docs/contributor_guide.rst
Co-authored-by: Purva Thakre <66048318+purva-thakre@users.noreply.github.com>
