+++ pull/713/_sources/contributor_guide.rst.txt
@@ -22,8 +22,8 @@ Currently, ``tqec`` is compatible with Python 3.10, 3.11, 3.12 and 3.13. You can
 of ``tqec`` through ``pip`` or ``uv``.
 
 .. hint::
-    Creating an environment before running `pip install` is optional but recommended to avoid everything installing globally.
-    Click [here](https://docs.python.org/3/library/venv.html) for a common approach.
+    Creating an environment before running ``pip install`` is optional but recommended to avoid everything installing globally.
+    Click `here <https://docs.python.org/3/library/venv.html>`_ for a common approach.
 
 .. tab-set::
 
