### 2025-10-02 816376f
Add observable instructions at the factory level
