+++ src/tqec/compile/specs/library/generators/y_basis.py
@@ -326,7 +326,7 @@ def _make_y_transition_round(
     distance: int,
     top_boundary_basis_before_transition: Basis,
     convention: Literal["fixed_boundary", "fixed_bulk"],
-    include_observable_flow: bool = True,
+    include_observables: Iterable[int] = (0,),
 ) -> gen.Chunk:
     """Make the circuit chunk for transition from qubit patch to degenerate patch."""
     start = _make_qubit_patch(distance, top_boundary_basis_before_transition, convention)
@@ -447,7 +447,8 @@ def _make_y_transition_round(
     # Annotate how observable flows through the system.
     # Contrast to the observable in the paper, the annotated observable is composed
     # of the two X/Z midline operators.
-    if include_observable_flow:
+    observables = sorted(include_observables)
+    if observables:
         half_d = distance // 2
         center_dq = complex(half_d, half_d)
         vertical_basis = top_boundary_basis_before_transition
@@ -455,33 +456,32 @@ def _make_y_transition_round(
 
         mset_ur = xs if top_basis == "Z" else zs
         mset_dl = xs if top_basis == "X" else zs
-        builder.add_flow(
-            center=center_dq,
-            start=gen.PauliMap(
-                {
-                    center_dq: "Y",
-                    **{
-                        complex(q, half_d): str(horizontal_basis)
-                        for q in range(distance)
-                        if q != half_d
-                    },
-                    **{
-                        complex(half_d, q): str(vertical_basis)
-                        for q in range(distance)
-                        if q != half_d
-                    },
-                }
-            ),
-            ms=[
-                (m, "solo")
-                for m in [
-                    my_target,
-                    *[q for q in mset_ur | top_row if q.real > half_d and q.imag < half_d],
-                    *[q for q in mset_dl | left_col if q.real < half_d and q.imag > half_d],
-                ]
-            ],
-            obs_key=0,
+        start = gen.PauliMap(
+            {
+                center_dq: "Y",
+                **{
+                    complex(q, half_d): str(horizontal_basis)
+                    for q in range(distance)
+                    if q != half_d
+                },
+                **{complex(half_d, q): str(vertical_basis) for q in range(distance) if q != half_d},
+            }
         )
+        ms = [
+            (m, "solo")
+            for m in [
+                my_target,
+                *[q for q in mset_ur | top_row if q.real > half_d and q.imag < half_d],
+                *[q for q in mset_dl | left_col if q.real < half_d and q.imag > half_d],
+            ]
+        ]
+        for idx in observables:
+            builder.add_flow(
+                center=center_dq,
+                start=start,
+                ms=ms,
+                obs_key=idx,
+            )
 
     return builder.finish_chunk()
 
@@ -540,7 +540,8 @@ def _make_y_basis_measurement_chunks(
     convention: Literal["fixed_boundary", "fixed_bulk"],
     padding_rounds: int,
     transform: Callable[[complex], complex] = lambda x: x,
-    include_observable_flow: bool = True,
+    include_observables: Iterable[int] = (0,),
+    close_obs_flows: bool = False,
     include_open_flows: bool = True,
 ) -> list[gen.Chunk | gen.ChunkLoop]:
     """Make circuit chunks for Y-basis measurement."""
@@ -548,7 +549,7 @@ def _make_y_basis_measurement_chunks(
         distance, top_boundary_basis_before_measure, convention
     )
     qubit_to_boundary_round = _make_y_transition_round(
[…]
