### 2025-10-02 136e7bd
Add very ugly proof of concept implementation of tree circuit injection
