+++ src/tqec/compile/tree/injection.py
@@ -1,0 +1,413 @@
+from collections.abc import Iterable
+
+import gen
+import stim
+from gen._layers._det_obs_annotation_layer import DetObsAnnotationLayer
+from gen._layers._empty_layer import EmptyLayer
+from gen._layers._layer import Layer
+
+from tqec.compile.blocks.block import InjectedBlock
+from tqec.compile.blocks.enums import Alignment
+from tqec.utils.position import BlockPosition2D
+
+
+class InjectionBuilder:
+    def __init__(
+        self,
+        k: int,
+        q2i: dict[complex, int],
+        o2i: dict[int, int],
+    ) -> None:
+        """Create a helper class to inject blocks into the layered circuits.
+
+        Args:
+            k: The scaling factor of the scalable circuit.
+            q2i: A mapping from qubit coordinates to global qubit indices.
+            o2i: A mapping from observable keys to observable indices.
+
+        """
+        self._k = k
+        self._q2i = q2i
+        self._o2i = o2i
+        self._chunks: list[gen.Chunk] = []
+
+        # Track the previous and current time slice (different z) circuits
+        self._prev_circuit = stim.Circuit()
+        self._cur_circuit = stim.Circuit()
+
+        # Track the semi-auto open flows for the previous, current, and next circuits
+        self._prev_flows: list[gen.FlowSemiAuto] = []
+        self._cur_flows: list[gen.FlowSemiAuto] = []
+        self._next_flows: list[gen.FlowSemiAuto] = []
+
+        # As the circuits built from the layer tree include detectors/observables
+        # with respect to measurement indices in the whole tree, we need to
+        # track the measurement indices in the tree circuit
+
+        # For those measurements from the tree circuits, i.e. circuits excluding
+        # all the injected block, track their global indices in the built circuit
+        # so far to correctly fix the lookback indices of detectors/observables
+        self._tree_circuits_global_mids: list[int] = []
+        self._cur_tree_circuit_global_mids: list[int] = []
+        self._cur_tree_circuit_local_mids: set[int] = set()
+        self._num_global_measurements = 0
+
+        self._has_injection: bool = False
+        self._out = stim.Circuit()
+
+    def finish(self) -> stim.Circuit:
+        """Finalize the building process and return the resulting circuit."""
+        # commit 2 extra times for prev and cur circuits
+        for _ in range(2):
+            self.append_tree_circuit(stim.Circuit())
+        return gen.compile_chunks_into_circuit(self._chunks)  # type: ignore
+
+    def _commit_prev_circuit(self) -> None:
+        """Commit the previous circuit."""
+        if self._prev_circuit:
+            # solve the measurement indices that should be included in the flows
+            chunk = gen.ChunkSemiAuto(
+                circuit=self._prev_circuit,
+                flows=self._prev_flows,
+                q2i=self._q2i,
+                o2i=self._o2i,
+            ).solve()
+            chunk.verify()
+            self._chunks.append(chunk)
+
+        # commit the measurement indices from the tree circuit
+        self._tree_circuits_global_mids.extend(self._cur_tree_circuit_global_mids)
+        if not self._has_injection:
+            self._num_global_measurements += len(self._cur_tree_circuit_global_mids)
+        self._cur_tree_circuit_global_mids.clear()
+        self._cur_tree_circuit_local_mids.clear()
+
+        self._prev_circuit = self._cur_circuit
+        self._prev_flows = self._cur_flows
+        self._cur_flows = self._next_flows
+        self._next_flows = []
+
+    def append_tree_circuit(self, circuit: stim.Circuit) -> None:
[…]
