+++ src/tqec/compile/tree/tree.py
@@ -21,6 +21,7 @@ from tqec.compile.tree.annotators.circuit import AnnotateCircuitOnLayerNode
 from tqec.compile.tree.annotators.detectors import AnnotateDetectorsOnLayerNode
 from tqec.compile.tree.annotators.observables import annotate_observable
 from tqec.compile.tree.annotators.polygons import AnnotatePolygonOnLayerNode
+from tqec.compile.tree.injection import InjectionBuilder
 from tqec.compile.tree.node import LayerNode, NodeWalker
 from tqec.post_processing.shift import shift_to_only_positive
 from tqec.utils.exceptions import TQECError, TQECWarning
@@ -362,13 +363,51 @@ class LayerTree:
             parallel_process_count=parallel_process_count,
         )
         annotations = self._get_annotation(k)
-        assert annotations.qubit_map is not None
+        qubit_map = annotations.qubit_map
+        assert qubit_map is not None
 
         circuit = stim.Circuit()
         if include_qubit_coords:
-            circuit += annotations.qubit_map.to_circuit()
-        circuit += self._root.generate_circuit(k, annotations.qubit_map)
-        return circuit
+            circuit += qubit_map.to_circuit()
+        if not self._injected_blocks:
+            circuit += self._root.generate_circuit(k, qubit_map)
+            return circuit
+
+        return self._get_circuit_after_injection(k)
+
+    def _get_circuit_after_injection(self, k: int) -> stim.Circuit:
+        qubit_map = self._get_annotation(k).qubit_map
+        assert qubit_map is not None
+
+        circuits_by_z: dict[int, stim.Circuit] = {}
+        for node in self._root.children:
+            assert isinstance(node._layer, SequencedLayers)
+            layer_z = node._layer.additional_metadata.get("z")
+            assert isinstance(layer_z, int)
+            circuits_by_z[layer_z] = node.generate_circuit(k, qubit_map)
+
+        # included observable indices at each injection position
+        injection_obs_indices: dict[BlockPosition3D, list[int]] = {}
+        for obs_idx, obs in enumerate(self._abstract_observables):
+            for cube in obs.y_half_cubes:
+                pos = BlockPosition3D(*cube.position.as_tuple())
+                injection_obs_indices.setdefault(pos, []).append(obs_idx)
+
+        q2i = {complex(q.x, q.y): i for q, i in qubit_map.q2i.items()}
+        o2i = {i: i for i in range(len(self._abstract_observables))}
+        builder = InjectionBuilder(k, q2i, o2i)
+
+        all_zs = sorted(set(circuits_by_z.keys()) | {pos.z for pos in self._injected_blocks})
+        for z in all_zs:
+            injection_blocks = {
+                pos: block for pos, block in self._injected_blocks.items() if pos.z == z
+            }
+            circuit_at_z = circuits_by_z.get(z, stim.Circuit())
+            assert injection_blocks or circuit_at_z
+            builder.append_tree_circuit(circuit_at_z)
+            for pos, block in injection_blocks.items():
+                builder.inject(pos.as_2d(), block, injection_obs_indices.get(pos, []))
+        return builder.finish()
 
     def _get_annotation(self, k: int) -> LayerTreeAnnotations:
         return self._annotations.setdefault(k, LayerTreeAnnotations())
