### 2025-10-03 468e9b6
refactor the injection code but still complex
