+++ src/tqec/compile/tree/injection.py
@@ -1,4 +1,6 @@
+from collections import deque
 from collections.abc import Iterable
+from dataclasses import dataclass, field
 
 import gen
 import stim
@@ -8,10 +10,21 @@ from gen._layers._layer import Layer
 
 from tqec.compile.blocks.block import InjectedBlock
 from tqec.compile.blocks.enums import Alignment
+from tqec.utils.exceptions import TQECError
 from tqec.utils.position import BlockPosition2D
 
 
 class InjectionBuilder:
+    """Incrementally weave injected blocks into tree-generated circuit slices.
+
+    The builder maintains three moving windows of circuit data (previous,
+    current, next) so that semi-auto flows, detector lookbacks, and measurement
+    indices remain consistent as blocks are injected between tree layers.  It
+    exposes a minimal API mirroring the tree traversal: append a raw tree slice,
+    optionally inject blocks at specific positions, and finally produce the
+    composed ``stim.Circuit`` once all slices are processed.
+    """
+
     def __init__(
         self,
         k: int,
@@ -40,30 +53,33 @@ class InjectionBuilder:
         self._cur_flows: list[gen.FlowSemiAuto] = []
         self._next_flows: list[gen.FlowSemiAuto] = []
 
-        # As the circuits built from the layer tree include detectors/observables
-        # with respect to measurement indices in the whole tree, we need to
-        # track the measurement indices in the tree circuit
-
         # For those measurements from the tree circuits, i.e. circuits excluding
-        # all the injected block, track their global indices in the built circuit
+        # all the injected blocks, track their global indices in the built circuit
         # so far to correctly fix the lookback indices of detectors/observables
-        self._tree_circuits_global_mids: list[int] = []
-        self._cur_tree_circuit_global_mids: list[int] = []
-        self._cur_tree_circuit_local_mids: set[int] = set()
+        self._mtracker = _TreeMeasurementTracker()
         self._num_global_measurements = 0
 
-        self._has_injection: bool = False
+        self._current_slice_has_injection: bool = False
         self._out = stim.Circuit()
 
     def finish(self) -> stim.Circuit:
-        """Finalize the building process and return the resulting circuit."""
+        """Finalize the building process and return the resulting circuit.
+
+        Two no-op commits flush any remaining buffered slices (``prev`` and
+        ``cur``) through the chunk compiler before producing the final circuit.
+        """
         # commit 2 extra times for prev and cur circuits
         for _ in range(2):
-            self.append_tree_circuit(stim.Circuit())
+            self._commit_prev_circuit()
         return gen.compile_chunks_into_circuit(self._chunks)  # type: ignore
 
     def _commit_prev_circuit(self) -> None:
-        """Commit the previous circuit."""
+        """Commit the previous circuit slice and related metadata.
+
+        This resolves semi-auto flows into a concrete chunk, persists detector
+        bookkeeping from the tree portion, and rolls the window forward so the
+        builder is ready for the next slice.
+        """
         if self._prev_circuit:
             # solve the measurement indices that should be included in the flows
             chunk = gen.ChunkSemiAuto(
@@ -76,11 +92,9 @@ class InjectionBuilder:
             self._chunks.append(chunk)
 
         # commit the measurement indices from the tree circuit
-        self._tree_circuits_global_mids.extend(self._cur_tree_circuit_global_mids)
-        if not self._has_injection:
-            self._num_global_measurements += len(self._cur_tree_circuit_global_mids)
-        self._cur_tree_circuit_global_mids.clear()
-        self._cur_tree_circuit_local_mids.clear()
+        num_commit_measurements = self._mtracker.commit_current_slice()
+        if not self._current_slice_has_injection:
+            self._num_global_measurements += num_commit_measurements
 
         self._prev_circuit = self._cur_circuit
[…]
