### 2025-10-03 d82359a
Split large chunks into smaller one to get the auto-solved measurements localized to single cycle
