+++ src/tqec/interop/shared.py
@@ -1,11 +1,52 @@
-"""Defines the color representation used in the TQEC interop layer."""
+"""Defines common classes and operations used in the TQEC interop layer."""
 
 from __future__ import annotations
 
 from dataclasses import dataclass
 from enum import Enum
 
+import numpy as np
 
+from tqec.utils.position import FloatPosition3D, Position3D
+from tqec.utils.scale import round_or_fail
+
+
+# TRANSFORMATIONS
+def int_position_before_scale(pos: FloatPosition3D, pipe_length: float) -> Position3D:
+    """Exchanges a float-based position with an integer-based position considering length of pipes.
+
+    Args:
+        pos: (x, y, z) position where x, y, and z are floats.
+        pipe_length: the length of the pipes in the model.
+
+    Returns:
+        An (x, y, z)  position where x, y, and z are integers.
+
+    """
+    return Position3D(
+        x=round_or_fail(pos.x / (1 + pipe_length), atol=0.35),
+        y=round_or_fail(pos.y / (1 + pipe_length), atol=0.35),
+        z=round_or_fail(pos.z / (1 + pipe_length), atol=0.35),
+    )
+
+
+def offset_y_cube_position(pos: FloatPosition3D, pipe_length: float) -> FloatPosition3D:
+    """Offsets the position of a Y-cube according to the length of pipes in the model.
+
+    Args:
+        pos: (x, y, z) position where x, y, and z are floats.
+        pipe_length: the length of the pipes in the model.
+
+    Returns:
+        An offset (x, y, z)  position where x, y, and z are floats.
+
+    """
+    if np.isclose(pos.z - 0.5, np.floor(pos.z), atol=1e-9):
+        pos = pos.shift_by(dz=-0.5)
+    return FloatPosition3D(pos.x, pos.y, pos.z / (1 + pipe_length))
+
+
+# CLASSES
 @dataclass(frozen=True)
 class RGBA:
     """RGBA representation of a color.
