+++ src/tqec/interop/collada/read_write.py
@@ -18,12 +18,11 @@ from tqec.computation.correlation import CorrelationSurface
 from tqec.computation.cube import CubeKind, Port, YHalfCube
 from tqec.computation.pipe import PipeKind
 from tqec.interop.collada._geometry import BlockGeometries, Face, get_correlation_surface_geometry
-from tqec.interop.color import TQECColor
+from tqec.interop.shared import TQECColor, int_position_before_scale, offset_y_cube_position
 from tqec.utils.enums import Basis
 from tqec.utils.exceptions import TQECError
 from tqec.utils.position import FloatPosition3D, Position3D, SignedDirection3D
 from tqec.utils.rotations import adjust_hadamards_direction, get_axes_directions, rotate_on_import
-from tqec.utils.scale import round_or_fail
 
 _ASSET_AUTHOR = "TQEC Community"
 _ASSET_AUTHORING_TOOL_TQEC = "https://github.com/tqec/tqec"
@@ -32,21 +31,6 @@ _ASSET_UNIT_METER = 0.02539999969303608
 
 _MATERIAL_SYMBOL = "MaterialSymbol"
 _CORRELATION_SUFFIX = "_CORRELATION"
-
-
-# SHARED FUNCTIONS
-def _int_position_before_scale(pos: FloatPosition3D, pipe_length: float) -> Position3D:
-    return Position3D(
-        x=round_or_fail(pos.x / (1 + pipe_length), atol=0.35),
-        y=round_or_fail(pos.y / (1 + pipe_length), atol=0.35),
-        z=round_or_fail(pos.z / (1 + pipe_length), atol=0.35),
-    )
-
-
-def _offset_y_cube_position(pos: FloatPosition3D, pipe_length: float) -> FloatPosition3D:
-    if np.isclose(pos.z - 0.5, np.floor(pos.z), atol=1e-9):
-        pos = pos.shift_by(dz=-0.5)
-    return FloatPosition3D(pos.x, pos.y, pos.z / (1 + pipe_length))
 
 
 # DAE EXPORTER/IMPORTER
@@ -164,11 +148,11 @@ def read_block_graph_from_dae_file(
     for pos, cube_kind, axes_directions in parsed_cubes:
         if isinstance(cube_kind, YHalfCube):
             graph.add_cube(
-                _int_position_before_scale(_offset_y_cube_position(pos, pipe_length), pipe_length),
+                int_position_before_scale(offset_y_cube_position(pos, pipe_length), pipe_length),
                 cube_kind,
             )
         else:
-            graph.add_cube(_int_position_before_scale(pos, pipe_length), cube_kind)
+            graph.add_cube(int_position_before_scale(pos, pipe_length), cube_kind)
     port_index = 0
 
     # Add pipes
@@ -176,7 +160,7 @@ def read_block_graph_from_dae_file(
         # Draw pipes in +1/-1 direction using position, kind of pipe, and directional pointers
         # from previous operations
         directional_multiplier = axes_directions[str(pipe_kind.direction)]
-        head_pos = _int_position_before_scale(
+        head_pos = int_position_before_scale(
             pos.shift_in_direction(pipe_kind.direction, -1 * directional_multiplier),
             pipe_length,
         )
@@ -269,7 +253,7 @@ def read_block_graph_from_json(
     filepath: str | pathlib.Path,
     graph_name: str = "",
 ) -> BlockGraph:
-    """Read a Collada JSON file and construct a :py:class:`.BlockGraph` from it.
+    """Construct a :py:class:`.BlockGraph` from a JSON file with COLLADA-like information.
 
     Args:
         filepath: The input dae file path.
@@ -390,17 +374,17 @@ def read_block_graph_from_json(
[…]
