### 2025-10-04 26699d9
Bump the detector database version as a breaking change
