### 2025-10-04 e710d71
Fix code block syntax of `uv`/`pip` install in readme (#714)
Fixes the code blocks below in the readme:

<img width="879" height="407" alt="截屏2025-10-01 10 03 19"
src="https://github.com/user-attachments/assets/7af8124b-ecd1-4ae3-965e-700f6194864b"
/>
