### 2025-10-04 1aa1ec6
delete directort and add cli rst file
