### 2025-10-04 569644f
error with bash code block
