### 2025-10-04 d79d47e
Fix some basis specifications and a hadamard flag
