+++ src/tqec/compile/specs/library/fixed_boundary.py
@@ -283,13 +283,6 @@ class FixedBoundaryPipeBuilder(PipeBuilder):
         is_hadamard = spec.pipe_kind.has_hadamard
         assert x is not None or y is not None
         spatial_boundary_basis: Basis = x if x is not None else y  # type: ignore
-        pipe_in_x_direction = spec.pipe_kind.direction == Direction3D.X
-        top_left_basis = (
-            spec.pipe_kind.get_basis_along(Direction3D.Y, at_head=True)
-            if pipe_in_x_direction
-            else spec.pipe_kind.get_basis_along(Direction3D.X, at_head=True)
-        )
-        assert top_left_basis is not None
         arms = FixedBoundaryPipeBuilder._get_spatial_cube_arms(
             spec
         )  # this tells me whether this is a vertical (up+down)
@@ -301,7 +294,6 @@ class FixedBoundaryPipeBuilder(PipeBuilder):
         def plaquettes_generator(is_reversed: bool, r: Basis | None, m: Basis | None) -> Plaquettes:
             return self._generator.get_spatial_cube_arm_plaquettes(
                 spatial_boundary_basis,
-                top_left_basis,
                 arms,
                 spec.cube_specs,
                 is_reversed,
