+++ src/tqec/compile/specs/library/generators/fixed_boundary.py
@@ -160,7 +160,7 @@ class FixedBoundaryConventionGenerator:
                 reversed or not. Useful to limit the loss of code distance when
                 hook errors are not correctly oriented by alternating regular
                 and reversed plaquettes.
-            hadamard: ``True`` if the plaquette should contain a Hadamard gate.
+            hadamard: ``True`` if the plaquette should contain a temporal Hadamard gate.
 
         Note:
             The 2-body stabilizer measurement plaquettes returned by this function
@@ -246,6 +246,63 @@ class FixedBoundaryConventionGenerator:
             }
             for basis in Basis
         }
+
+    def get_2_body_horizontal_hadamard_rpng_descriptions(
+        self, is_reversed: bool
+    ) -> dict[Basis, dict[PlaquetteOrientation, RPNGDescription]]:
+        """Get plaquettes that are supposed to be used on the boundaries.
+
+        This function returns the four 2-body stabilizer measurement plaquettes
+        which could be needed in a horizontal hadamard (ie only LEFT and RIGHT
+        orientations).
+
+        Args:
+            is_reversed: flag indicating if the plaquette schedule should be
+                reversed or not. Useful to limit the loss of code distance when
+                hook errors are not correctly oriented by alternating regular
+                and reversed plaquettes.
+
+        Note:
+            The 2-body stabilizer measurement plaquettes returned by this function
+            all follow the same schedule: ``1-2-3-5``.
+
+        Warning:
+            By convention, the 2-body stabilizers never reset/measure any
+            data-qubit. This is done because it is way simpler to reset the correct
+            data-qubits in 4-body stabilizers, and the resets/measurements in 2-body
+            stabilizers would be redundant.
+
+        Warning:
+            This function uses the :class:`~tqec.plaquette.enums.PlaquetteOrientation`
+            class. For a 2-body stabilizer measurement plaquette, the "orientation"
+            corresponds to the direction in which the rounded side is pointing.
+            So a plaquette with the orientation ``PlaquetteOrientation.DOWN`` has the
+            following properties:
+            - it measures the 2 data-qubits on the **top** side of the usual 4-body
+            stabilizer measurement plaquette,
+            - it can be used for a bottom boundary,
+            - its rounded side points downwards.
+
+        Returns:
+            a mapping with 4 plaquettes: one for each top-left basis
+            (either ``X`` or ``Z``) and for each plaquette orientation
+            (``LEFT`` or ``RIGHT``).
+
+        """
+        _po = PlaquetteOrientation
+        ret: dict[Basis, dict[PlaquetteOrientation, RPNGDescription]] = {}
+        # Note: the schedule of CNOT gates in weight-2 plaquettes is less
+        # important because hook errors do not exist. We arbitrarily chose the
+        # vertical schedule.
+        s = VERTICAL_HOOK_SCHEDULES[is_reversed]
+        for basis in Basis:
+            b = basis.value.lower()
+            o = basis.flipped().value.lower()
+            ret[basis] = {
+                _po.LEFT: RPNGDescription.from_string(f"---- -{b}{s[1]}- ---- -{o}{s[3]}-"),
+                _po.RIGHT: RPNGDescription.from_string(f"-{b}{s[0]}- ---- -{o}{s[2]}- ----"),
+            }
+        return ret
 
     def get_spatial_x_hadamard_rpng_descriptions(
         self,
@@ -1118,7 +1175,6 @@ class FixedBoundaryConventionGenerator:
[…]
