### 2025-10-07 d73351e
Add qiskit & topologiq to main deps & add qrisp to new deps group for all supported external nisq frameworks
