### 2025-10-07 9e308fa
add qiskit & topologiq both in docs and integration groups
