### 2025-10-07 eadd290
adjust qiskit v. to fully compatible v.
