+++ pyproject.toml
@@ -110,7 +110,7 @@ dev = [
 
 integration = [
   "topologiq@git+https://github.com/tqec/topologiq.git",
-  "qiskit>=2.1.1",
+  "qiskit>=2.1.2",
   "qrisp>=0.7.9"
 ]
 
