### 2025-10-07 99afd6f
Add steane.qpy as independent asset & related adjustments
