### 2025-10-08 45a3463
Add nisq notebook to docs/gallery index
