+++ docs/gallery/index.rst
@@ -12,3 +12,4 @@ This section contains a collection of constructed logical computations and their
    move_rotation.ipynb
    three_cnots.ipynb
    steane_encoding.ipynb
+   using_nisq_circuits.ipynb
