### 2025-10-08 28e2e01
Load circuits directly from ASSETS to notebook
