+++ src/tqec/interop/nisq/__init__.py
@@ -1,5 +1,0 @@
-"""Interoperability with Noisy Intermediate Scale Quantum (NISQ) circuit design frameworks."""
-
-from tqec.interop.nisq.load import (
-    load_nisq_circuit_as_qasm as load_nisq_circuit_as_qasm,
-)
