### 2025-10-08 1280cc9
Update uv-build requirement from <0.9.0,>=0.7.19 to >=0.7.19,<0.10.0
Updates the requirements on [uv-build](https://github.com/astral-sh/uv) to permit the latest version.
- [Release notes](https://github.com/astral-sh/uv/releases)
- [Changelog](https://github.com/astral-sh/uv/blob/main/CHANGELOG.md)
- [Commits](https://github.com/astral-sh/uv/compare/0.7.19...0.9.0)

---
updated-dependencies:
- dependency-name: uv-build
  dependency-version: 0.9.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
