+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["uv_build >= 0.7.19, <0.9.0"]
+requires = ["uv_build >= 0.7.19, < 0.10.0"]
 build-backend = "uv_build"
 
 [project]
