### 2025-10-08 110548d
Rewritting notebook & renaming related assets for clarity
