+++ pull/720/_sources/gallery/qiskit_integration.ipynb.txt
@@ -5,7 +5,7 @@
    "id": "695f75d2",
    "metadata": {},
    "source": [
-    "# Using circuits designed with external NISQ frameworks"
+    "# Importing a Qiskit circuit and running simulations with it"
    ]
   },
   {
@@ -34,9 +34,11 @@
    "id": "27f0f36b",
    "metadata": {},
    "source": [
-    "This notebook illustrates the process of importing into TQEC circuits originally created with external NISQ circuit design software/frameworks, such as, for instance, Qiskit and Qrisp.\n",
+    "This notebook illustrates the process of importing into TQEC circuits originally created with Qiskit.\n",
     "\n",
-    "> **NB!** Run `uv add --only-group integration` to install all libraries used in this notebook (ignore if you installed all dependency groups when setting up TQEC)."
+    "The notebook assumes you have already designed a Qiskit circuit and saved it using Qiskit's `.qpy` format. \n",
+    "\n",
+    "> **NB!** Run `uv add --only-group integration` to install the libraries used in this notebook (ignore if you installed all dependency groups when setting up TQEC)."
    ]
   },
   {
@@ -44,7 +46,7 @@
    "id": "2be0a675",
    "metadata": {},
    "source": [
-    "## Load the foundational circuit"
+    "## Load Qiskit circuit"
    ]
   },
   {
@@ -52,32 +54,7 @@
    "id": "2f541923",
    "metadata": {},
    "source": [
-    "The first step is, naturally, to load your foundational circuit. This step varies slightly according the your chosen framework, but it comes down to loading the circuit and converting it into OpenQASM.\n",
-    "\n",
-    "So, let's start by selecting the NISQ framework used to design the foundational circuit (currently supported: qiskit, qrisp)."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "50825bec",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "# SELECT NISQ FRAMEWORK AND NAME OF CIRCUIT\n",
-    "# Note. This block assumes the corresponding circuit is saved using the circuit name\n",
-    "source_nisq_framework = \"qiskit\"  # options currently available [\"qiskit\", \"qrisp\"]\n",
-    "circuit_name = \"steane\""
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "98aeab11",
-   "metadata": {},
-   "source": [
-    "Let's now load the circuit!\n",
-    "\n",
-    "> **NB!** We use conditionals in the block below to allow anyone who downloads the notebook to easily run tests using any supported framework. In your own notebook, you would use only the method for your chosen framework."
+    "The first step is, naturally, to load the Qiskit circuit."
    ]
   },
   {
@@ -87,39 +64,18 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "# LOAD CORRESPONDING CIRCUIT AS QASM\n",
-    "qc = None\n",
-    "qasm_str = None\n",
+    "from qiskit import qasm2, qpy\n",
     "\n",
-    "if source_nisq_framework == \"qiskit\":\n",
-    "    \n",
-    "    # Load applicable Qiskit modules\n",
-    "    from qiskit import qasm2, qpy\n",
-    "    \n",
-    "    # Load circuit\n",
-    "    path_to_c = ASSETS_FOLDER / f\"{circuit_name}_{source_nisq_framework}.qpy\"\n",
-    "    with open(path_to_c, \"rb\") as f:\n",
-    "        qc = qpy.load(f)[0]\n",
+    "# NAME OF CIRCUIT\n",
[…]
