+++ pull/720/_sources/gallery/qiskit_integration.ipynb.txt
@@ -15,7 +15,6 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "import sys\n",
     "import random\n",
     "import numpy as np\n",
     "import matplotlib.pyplot as plt\n",
@@ -211,9 +210,6 @@
     "    random.seed(12)\n",
     "    zx.draw(zx_graph, labels = True, auto_layout=True)\n",
     "\n",
-    "    # Object below is not typical in a PyZX workflow but useful for visualisations in the next section.\n",
-    "    fig_data = zx.draw_matplotlib(zx_graph, labels=True)\n",
-    "\n",
     "else:\n",
     "    print(\"ERROR. Block didn't run because there is no PyZX graph to manipulate.\")"
    ]
@@ -297,7 +293,6 @@
     "        simple_graph,  # The PyZX input graph, simplified\n",
     "        full_circuit_name,  # The name of the circuit\n",
     "        visualise=(visualisation, animation),\n",
-    "        fig_data=fig_data,\n",
     "        **kwargs\n",
     "    )\n",
     "\n",
