### 2025-10-09 a527f54
added temporal height test, adjusted crashed graph_tests because of new interface
