### 2025-10-09 8e80eb6
Fix topologiq coordinate transformation and add comprehensive tests
This commit addresses a bug in the topologiq lattice surgery integration
where pipe positions were calculated incorrectly using midpoint logic.

Changes:
- Fixed pipe positioning to directly transform source and target cube
  coordinates instead of calculating midpoints and shifting
- This approach matches the existing JSON importer in collada/read_write.py
- Removed unused parsed_ports variable
- Added clarifying comments explaining automatic port creation mechanism

Testing:
- Added 17 comprehensive unit tests (previously 14
[…]
