### 2025-10-10 daa1b47
docs: Add comprehensive documentation contribution guide
- Added 'Contributing to Documentation' section to contributor guide
- Documented step-by-step process for adding user guide pages
- Documented step-by-step process for adding gallery examples
- Included best practices and common pitfalls
- Addresses issue #718 with detailed RST and Jupyter notebook workflows
- Explains jupyter-sphinx, nbsphinx, and sphinxcontrib-bibtex usage
