+++ docs/contributor_guide.rst
@@ -157,3 +157,240 @@ Once your code has been reviewed and accepted by at least one of the developers,
 will be able to merge it to the ``main`` branch.
 You (the PR owner) are responsible to click on the "Merge" button. If you prefer someone
 else to do it, you should send a clear comment asking for it.
+
+Contributing to Documentation
+------------------------------
+
+The ``tqec`` documentation is built using Sphinx and includes both a user guide and an
+example gallery. This section explains how to add new content to either.
+
+How to add a page to the user guide
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The user guide pages are located in ``docs/user_guide/`` and written in reStructuredText
+(``.rst``) format. Here's how to add a new page:
+
+1. **Create a new .rst file** in ``docs/user_guide/``
+
+   Name your file descriptively (e.g., ``advanced_compilation.rst``). Start with a title:
+
+   .. code-block:: rst
+
+       My Page Title
+       =============
+
+2. **Add your content using reStructuredText**
+
+   See the `Sphinx reStructuredText documentation <https://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html>`_
+   for syntax reference. Common elements include:
+
+   - **Sections**: Use ``=====`` under titles, ``-----`` for subsections
+   - **Code blocks**: Use ``.. code-block:: python`` for static code
+   - **Executable code**: Use ``.. jupyter-execute::`` for code that runs during the docs build
+   - **Links**: Use ``:ref:`label-name``` to reference other pages
+
+3. **Use executable code blocks when possible**
+
+   The ``tqec`` documentation uses `jupyter-sphinx <https://jupyter-sphinx.readthedocs.io/en/latest/>`_
+   to execute code during the documentation build. This ensures examples stay up-to-date:
+
+   .. code-block:: rst
+
+       .. jupyter-execute::
+
+           from tqec import BlockGraph
+           # Your working code here
+           print("This code will be executed during docs build")
+
+   .. warning::
+       Make sure your code executes quickly (< 30 seconds) to avoid slowing down the docs build.
+
+4. **Add references if needed**
+
+   If you cite academic papers or external resources:
+
+   a. Add the reference to ``docs/refs.bib`` in alphabetical order by last name:
+
+      .. code-block:: bibtex
+
+          @article{AuthorName_2024,
+             title={Paper Title},
+             ...
+          }
+
+   b. Cite in your ``.rst`` file using ``footcite``:
+
+      .. code-block:: rst
+
+          This approach was introduced by [<cite data-footcite-t="AuthorName_2024"></cite>].
+
+   c. Add a References section at the bottom of your page:
+
+      .. code-block:: rst
+
+          References
+          ----------
+          .. footbibliography::
+
+5. **Add your page to the index**
+
+   Edit ``docs/user_guide/index.rst`` and add your page to the ``toctree``:
+
+   .. code-block:: rst
+
+       .. toctree::
+          :maxdepth: 2
+
+          installation
+          quick_start
+          my_new_page
+
+6. **Build the docs locally to verify**
+
+   .. code-block:: bash
+
+       # From the repository root
+       cd docs
+       make html
+       # View the result by opening docs/_build/html/index.html in a browser
+
+   Check that:
+
+   - Your page appears in the navigation
+   - All code blocks execute successfully
+   - Images and links work correctly
+   - References render properly
+
+How to add an example to the docs gallery
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The gallery showcases complete computation examples and is located in ``docs/gallery/``.
[…]
