### 2025-10-10 9c28c40
Fix topologiq coordinate transformation in read_from_lattice_dicts (#726)
## Summary

Fixes a coordinate transformation bug in `topologiq.py` where pipe
positions were calculated using midpoints instead of direct endpoint
transformation.

## Problem
The original implementation calculated pipe positions using midpoints
and directional multipliers. This was conceptually incorrect because:
- Pipes in TQEC are logical connections, not positioned objects
- The midpoint calculation introduced errors for certain lattice
configurations
- It didn't match the proven pattern in `collada/read_write.py`

#
[…]
