+++ src/tqec/interop/pyzx/topologiq.py
@@ -1,4 +1,27 @@
-"""Read and write block graphs to and from Collada DAE files."""
+"""Read and write block graphs to and from topologiq space-time diagrams.
+
+This module provides interoperability between topologiq's lattice surgery
+representation and TQEC's BlockGraph representation. The primary function
+`read_from_lattice_dicts` converts topologiq's output (lattice nodes and edges)
+into a TQEC BlockGraph that can be compiled and simulated.
+
+Typical usage:
+    ```python
+    from topologiq.scripts.runner import runner
+    from topologiq.utils.interop_pyzx import pyzx_g_to_simple_g
+    from tqec.interop.pyzx.topologiq import read_from_lattice_dicts
+
+    # Convert PyZX graph to topologiq format
+    simple_graph = pyzx_g_to_simple_g(zx_graph)
+
+    # Run topologiq
+    _, _, lattice_nodes, lattice_edges = runner(simple_graph, "circuit_name")
+
+    # Convert to TQEC BlockGraph
+    lattice_edges_min = dict([(k, v[0]) for k, v in lattice_edges.items()])
+    block_graph = read_from_lattice_dicts(lattice_nodes, lattice_edges_min)
+    ```
+"""
 
 from __future__ import annotations
 
@@ -16,19 +39,85 @@ def read_from_lattice_dicts(
 ) -> BlockGraph:
     """Construct a :class:`.BlockGraph` from a space-time diagram produced by topologiq.
 
-    Args:
-        lattice_nodes: {id: ((x, y, z), kind)} for each node (cube) in intended blockgraph
-        lattice_edges: {(source_id, target_id): kind} for each edge (pipes) in intended blockgraph
-        graph_name: The name of the block graph. Default is an empty string.
+    This function converts topologiq's lattice surgery representation into TQEC's
+    BlockGraph format. It handles coordinate transformations, node/edge parsing,
+    and automatic port creation for boundary nodes.
 
+    The conversion process:
+    1. Validates input dictionaries
+    2. Parses nodes (cubes) and their positions
+    3. Parses edges (pipes) and calculates their positions
+    4. Creates a BlockGraph with appropriate transformations
+    5. Automatically adds Port cubes for boundary connections
+
+    Args:
+        lattice_nodes: Dictionary mapping node IDs to (position, kind) tuples.
+            - ID: Unique integer identifier for the node
+            - position: (x, y, z) coordinate tuple in topologiq space
+            - kind: String identifier for cube type (e.g., 'ZXZ', 'ZXX', 'ooo' for ports)
+            Example: {0: ((0, 0, 0), 'ZXZ'), 1: ((3, 0, 0), 'ZXX')}
+
+        lattice_edges: Dictionary mapping (source_id, target_id) pairs to pipe kinds.
+            - source_id, target_id: Node IDs from lattice_nodes
+            - kind: String identifier for pipe type (e.g., 'X', 'Z')
+            Example: {(0, 1): 'X', (1, 2): 'Z'}
+            Note: topologiq typically returns lists as values; extract first element
+                  before passing to this function
+
+        graph_name: Optional name for the resulting BlockGraph. Used for visualization
+            and debugging. Default is an empty string.
 
     Returns:
-        The constructed :py:class:`~tqec.computation.block_graph.BlockGraph` object.
+        BlockGraph: A fully constructed BlockGraph with:
+            - All cubes positioned and typed correctly
+            - All pipes connecting cubes
+            - Port cubes automatically added at boundaries
+            - Coordinates transformed from topologiq space to TQEC space
 
     Raises:
-        ValueError: If incoming objects cannot be parsed and converted to a block graph.
+        ValueError: If input validation fails or conversion encounters errors:
[…]
