### 2025-10-10 7dd52ae
feat: Significantly improve error messages in PyZX and COLLADA interop
- Enhanced positioned.py error messages with detailed context:
  * Vertex mismatch now shows missing/extra IDs
  * Non-neighbor edges show Manhattan distance
  * Unsupported vertex types list valid combinations
  * Dangling spider errors show neighbor positions
  * 3D corner errors enumerate connecting directions

- Enhanced read_write.py error messages with actionable guidance:
  * Scene/structure errors explain expected format
  * Inconsistent pipe lengths show actual vs expected values
  * Invalid scaling errors specify 
[…]
