+++ src/tqec/interop/collada/read_write.py
@@ -74,12 +74,19 @@ def read_block_graph_from_dae_file(
 
     # Check some invariants about the DAE file
     if mesh.scene is None:
-        raise TQECError("No scene found in the DAE file.")
+        raise TQECError(
+            f"No scene found in the COLLADA file '{filepath}'. "
+            "The DAE file must contain a <visual_scene> element. "
+            "Ensure the file was exported correctly from SketchUp or your modeling tool."
+        )
     scene: collada.scene.Scene = mesh.scene
 
     if not (len(scene.nodes) == 1 and scene.nodes[0].name == "SketchUp"):
+        actual_nodes = [node.name for node in scene.nodes] if scene.nodes else []
         raise TQECError(
-            "The <visual_scene> node must have a single child node with the name 'SketchUp'."
+            f"Invalid scene structure in '{filepath}'. "
+            f"Expected single 'SketchUp' node, found {len(scene.nodes)} node(s): {actual_nodes}. "
+            "The COLLADA file must be exported from SketchUp with default settings."
         )
 
     sketchup_node: collada.scene.Node = scene.nodes[0]
@@ -136,13 +143,20 @@ def read_block_graph_from_dae_file(
                 if pipe_length is None:
                     pipe_length = scale * 2.0
                 elif not np.isclose(pipe_length, scale * 2.0, atol=1e-9):
-                    raise TQECError("All pipes must have the same length.")
+                    raise TQECError(
+                        f"Inconsistent pipe lengths in '{filepath}'. "
+                        f"Expected length {pipe_length:.4f}, but pipe '{name}' at {translation} "
+                        f"has length {scale * 2.0:.4f}. "
+                        "All pipes in the COLLADA model must have identical scaling."
+                    )
                 expected_scale = np.ones(3)
                 expected_scale[pipe_direction.value] = scale
                 if not np.allclose(transformation.scale, expected_scale, atol=1e-9):
                     raise TQECError(
-                        "Only the dimension along the pipe can be scaled, "
-                        f"which is not the case at {translation}."
+                        f"Invalid pipe scaling at {translation} in '{filepath}'. "
+                        f"Pipe '{name}' in {pipe_direction} direction has scale {transformation.scale}, "
+                        f"but expected {expected_scale}. "
+                        "Only the dimension along the pipe direction can be scaled."
                     )
                 # Append
                 parsed_pipes.append((translation, kind, axes_directions))
@@ -150,7 +164,12 @@ def read_block_graph_from_dae_file(
             else:
                 # Checks
                 if not np.allclose(transformation.scale, np.ones(3), atol=1e-9):
-                    raise TQECError(f"Cube at {translation} has a non-identity scale.")
+                    raise TQECError(
+                        f"Invalid cube scaling in '{filepath}'. "
+                        f"Cube '{name}' at {translation} has scale {transformation.scale}, "
+                        "but cubes must have identity scale [1.0, 1.0, 1.0]. "
+                        "Ensure cubes are not scaled in SketchUp."
+                    )
                 # Append
                 parsed_cubes.append((translation, kind, axes_directions))
 
@@ -285,17 +304,40 @@ def read_block_graph_from_json(
     try:
         with open(filepath) as f:
             data = json.load(f)
-    except Exception:
-        raise TQECError("JSON file not found.")
+    except FileNotFoundError as e:
+        raise TQECError(
+            f"JSON file not found at '{filepath}'. "
+            "Verify the file path is correct and the file exists."
+        ) from e
+    except json.JSONDecodeError as e:
+        raise TQECError(
+            f"Invalid JSON format in '{filepath}' at line {e.lineno}, column {e.colno}: {e.msg}. "
[…]
