### 2025-10-10 b72f581
docs: Add comprehensive development workflow documentation
- Documented multi-branch feature development strategy
- Detailed all active branches with status and next steps
- Workflow steps for starting new work and opening PRs
- Testing strategy and commands
- Communication guidelines for respectful collaboration
- Contribution metrics and learning outcomes
- Resource links and future directions

This document serves as the single source of truth for how this fork
operates and contributes to upstream TQEC.
