+++ DEVELOPMENT.md
@@ -1,0 +1,329 @@
+# Development Workflow for SMC17/tqec Fork
+
+This document describes the development strategy and branching model used in this fork to contribute to the upstream TQEC project.
+
+## Fork Strategy
+
+**Fork Owner**: Sean Collins (SMC17)  
+**Upstream**: [tqec/tqec](https://github.com/tqec/tqec)  
+**Fork**: [SMC17/tqec](https://github.com/SMC17/tqec)
+
+## Philosophy
+
+This fork follows a **multi-branch feature development** approach:
+
+1. **Each issue gets its own feature branch** - Allows parallel work on multiple contributions
+2. **Branches track upstream/main** - Ensures compatibility with latest code
+3. **Comprehensive documentation** - Each branch includes detailed analysis and rationale
+4. **Test locally before PR** - Quality over speed
+5. **Respectful collaboration** - Never assume maintainers will accept; always ask first
+
+## Branch Naming Convention
+
+```
+feature/<descriptive-name>
+fix/<bug-description>
+docs/<documentation-update>
+```
+
+Examples:
+- `feature/docs-contributing-guide` - Issue #718 documentation improvements
+- `feature/improve-error-messages` - Issue #262 complementary work
+- `fix/topologiq-coordinate-transformation` - Pipe position calculation bug
+
+## Current Active Branches
+
+### 1. `feature/docs-contributing-guide`
+**Status**: ✅ Complete, pushed  
+**Related Issue**: [#718](https://github.com/tqec/tqec/issues/718)  
+**Description**: Adds comprehensive documentation on how to contribute to the user guide and gallery  
+
+**Changes**:
+- Added "Contributing to Documentation" section to `docs/contributor_guide.rst`
+- Step-by-step guide for adding user guide pages (RST format)
+- Step-by-step guide for adding gallery examples (Jupyter notebooks)
+- Best practices and common pitfalls
+- Explains `jupyter-sphinx`, `nbsphinx`, and `sphinxcontrib-bibtex` usage
+
+**Testing**: RST syntax validated, ready for full docs build
+
+**Next Steps**: Wait for maintainer feedback before opening PR
+
+---
+
+### 2. `feature/improve-error-messages`
+**Status**: ✅ Complete, pushed  
+**Related Issue**: [#262](https://github.com/tqec/tqec/issues/262) (complementary)  
+**Description**: Improves error message clarity and actionability across user-facing modules
+
+**Changes**:
+- Enhanced `interop/pyzx/positioned.py` error messages (6 improvements)
+- Enhanced `interop/collada/read_write.py` error messages (10 improvements)
+- Created `ERROR_MESSAGE_IMPROVEMENTS.md` documenting patterns and strategy
+- All errors now include:
+  - Specific values that caused the error
+  - Actionable suggestions for fixes
+  - Clear explanations of technical terms
+  - File paths and relevant context
+
+**Testing**: Need to run existing test suite to ensure error messages appear correctly
+
+**Next Steps**: 
+1. Run tests: `pytest src/tqec/interop/`
+2. Document any test failures
+3. Wait for #262 exception type refactoring to complete
+4. Coordinate with @purva-thakre on integration
+
+---
+
+### 3. `fix/topologiq-coordinate-transformation`
+**Status**: ✅ Complete, PR #726 open  
+**Related Issue**: [#723](https://github.com/tqec/tqec/issues/723)  
+**Description**: Fixes critical bug in PyZX → COLLADA pipe position calculations
+
+**Changes**:
+- Corrected `write_block_graph_to_dae_file()` pipe positioning logic
[…]
