### 2025-10-10 b79be3a
fix: Resolve ruff formatting and line length violations
- Split long error messages to stay under 100 char limit
- Extract complex ternary into variable for clarity
- All pre-commit hooks now pass

This addresses the CI failure that would have occurred
when opening a PR for this branch.
