### 2025-10-10 739a9c9
test: Update test assertions to match improved error messages
- Fixed test regex patterns to match new error message format
- Corrected manhattan_distance calculation (Position3D doesn't have that method)
- All 6 tests in positioned_test.py now pass

This ensures tests validate the improved error messages are working correctly.
