### 2025-10-10 ff3808a
chore: automatic check format when using detector database
