### 2025-10-10 357143a
chore: remove legacy args for db from file
