+++ pull/720/_sources/gallery/qrisp_integration.ipynb.txt
@@ -1,599 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "695f75d2",
-   "metadata": {},
-   "source": [
-    "# Importing a Qrisp circuit and running simulations with it"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "5986b007",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "import sys\n",
-    "import random\n",
-    "import numpy as np\n",
-    "import matplotlib.pyplot as plt\n",
-    "\n",
-    "from IPython.display import display\n",
-    "from pathlib import Path\n",
-    "\n",
-    "ASSETS_FOLDER = Path(\"../../assets/\").resolve()\n",
-    "\n",
-    "# To display interactive 3D visualisations, uncomment the following line.\n",
-    "# %matplotlib widget"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "27f0f36b",
-   "metadata": {},
-   "source": [
-    "This notebook illustrates the process of importing into TQEC circuits originally created with Qrisp.\n",
-    "\n",
-    "The notebook assumes you have already designed a Qrisp circuit. \n",
-    "\n",
-    "> **NB!** Run `uv add --only-group integration` to install the libraries used in this notebook (ignore if you installed all dependency groups when setting up TQEC)."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "2be0a675",
-   "metadata": {},
-   "source": [
-    "## Load Qrisp circuit"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "2f541923",
-   "metadata": {},
-   "source": [
-    "The first step is, naturally, to load the Qrisp circuit."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "66009227",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "from qrisp.jasp import make_jaspr\n",
-    "\n",
-    "# CIRCUIT NAME\n",
-    "circuit_name = \"steane\"\n",
-    "\n",
-    "# LOAD CIRCUIT\n",
-    "# Since Qrisp doesn't have a native file format,\n",
-    "# we import a function which returns a Qrisp circuit.\n",
-    "sys.path.insert(0, str(ASSETS_FOLDER))\n",
-    "import steane_qrisp # type: ignore[reportMissingImports]\n",
-    "\n",
-    "qc_encoding_function = getattr(steane_qrisp, f\"{circuit_name}_qrisp\")\n",
-    "qc = qc_encoding_function()[0].qs\n",
-    "\n",
-    "# CONVERT to QASM\n",
-    "jaspr = make_jaspr(qc_encoding_function)()\n",
-    "qasm_str = jaspr.to_qasm()"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "3fac4bac",
-   "metadata": {},
-   "source": [
-    "We can print the original circuit for inspection."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "3bdc28c4",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "print(f\"\\n=> Imported Qrisp circuit of name {circuit_name.capitalize()}:\\n\")\n",
-    "print(qc) if qc else print(\"ERROR. Failed to load a circuit\")"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "ea56a685",
-   "metadata": {},
-   "source": [
[…]
