### 2025-10-10 96cffc7
pre-commit on topologiq interop files
