+++ pull/720/_sources/gallery/index.rst.txt
@@ -1,0 +1,16 @@
+.. _gallery-reference-label:
+
+Computation Gallery
+===================
+
+This section contains a collection of constructed logical computations and their simulation results.
+
+.. nbgallery::
+
+   memory.ipynb
+   cnot.ipynb
+   move_rotation.ipynb
+   three_cnots.ipynb
+   steane_encoding.ipynb
+   qiskit_integration.ipynb
+   qrisp_integration.ipynb
