### 2025-10-10 2fed55c
docs: Add comprehensive validation and testing framework
- CONTRIBUTION_VALIDATION.md: Self-review checklist and defensive arguments
- TESTING_RESULTS_MAC.md: Complete Mac phase testing results
- CONTRIBUTOR_SPRINT_OCT10-15.md: 5-day strategic plan

All contributions validated against:
- Project requirements (Issue #718, #262)
- Code quality standards (PR #726 benchmark)
- Project mission alignment
- No-conflict verification

Ready for Threadripper comprehensive testing phase.
