+++ CONTRIBUTOR_SPRINT_OCT10-15.md
@@ -1,0 +1,408 @@
+# 5-Day Sprint: Becoming a TQEC Contributor
+**Goal**: Establish myself as a trusted contributor by October 15, 2025  
+**Author**: Sean Collins (SMC17)  
+**Start**: October 10, 2025
+
+---
+
+## Maintainer Feedback Analysis
+
+**From @jbolns on PR #726:**
+> "This is outstanding! Will merge into `framework_integration` branch as given. Any errors that may arise we'll fix once in the branch because **your implementation is better than what there is currently in place**."
+
+### Key Takeaways:
+✅ **Quality matters more than speed** - Comprehensive tests won approval  
+✅ **Clear documentation wins** - Detailed PR description was appreciated  
+✅ **Don't be afraid to improve existing code** - They want better solutions  
+✅ **Thorough testing = trust** - 17 tests showed I care about correctness  
+
+---
+
+## Current Status (Oct 10, 9am)
+
+### ✅ Completed
+- **PR #726 MERGED**: Fix topologiq coordinate transformation
+  - 17 comprehensive tests
+  - "Outstanding" feedback from maintainer
+  - Merged into `framework_integration`
+
+### 🔄 In Progress (On Fork)
+- **feature/improve-error-messages**: Ready, tested, pushed to fork
+- **feature/docs-contributing-guide**: Ready, tested, pushed to fork
+
+---
+
+## Strategic Analysis: What The Project Needs
+
+### High-Value, Low-Conflict Opportunities
+
+#### 1. **Issue #718** - Documentation (PRIORITY 1) ⭐
+- **Status**: Unassigned, good first issue
+- **Labels**: documentation, good first issue, non-quantum
+- **What you have**: Complete solution ready on branch
+- **Risk**: LOW - pure documentation, non-controversial
+- **Impact**: HIGH - helps future contributors
+- **Action**: Comment + open PR
+
+#### 2. **Issue #723** - Topologiq Integration (PRIORITY 2) ⭐⭐⭐
+- **Status**: Unassigned, HIGH PRIORITY, help wanted
+- **Labels**: enhancement, help wanted, priority: high
+- **What you have**: Your merged PR fixes part of this!
+- **Risk**: MEDIUM - already demonstrated capability
+- **Impact**: VERY HIGH - directly tied to framework integration
+- **Action**: Comment offering to help with remaining work
+
+#### 3. **Issue #690** - Opacity Control (PRIORITY 3)
+- **Status**: Unassigned, good first issue
+- **Labels**: enhancement, good first issue, non-quantum
+- **Risk**: LOW - visualization enhancement
+- **Impact**: MEDIUM - nice quality-of-life improvement
+- **Action**: Implement if time permits
+
+#### 4. **Issue #673** - Dark Mode Text (PRIORITY 4)
+- **Status**: Unassigned, good first issue
+- **Labels**: good first issue, non-quantum
+- **Risk**: LOW - UI improvement
+- **Impact**: MEDIUM - improves accessibility
+- **Action**: Quick win if time permits
+
+### ❌ AVOID (Stepping on Toes)
+- **Issue #262**: Exception refactoring - @purva-thakre assigned
+- **Issue #696**: Performance - @Zhaoyilunnn assigned  
+- **Issue #681**: Mac/Linux DB - @KabirDubey assigned
+- **Issue #637**: Correlation surfaces doc - 2 people assigned
+- **PR #720**: Framework integration - @jbolns's main work (draft)
+- **PR #721**: Temporal height - @SchmidtMoritz active work
+
+---
+
+## 5-Day Execution Plan
+
[…]
