+++ CONTRIBUTION_VALIDATION.md
@@ -1,0 +1,432 @@
+# Contribution Validation & Self-Review
+**Date**: October 10, 2025  
+**Purpose**: Ensure all contributions meet TQEC standards and actually solve problems
+
+---
+
+## Philosophy
+
+**We don't just write code - we solve problems correctly.**
+
+Before submitting ANY PR, we must prove:
+1. ✅ The problem exists and is worth solving
+2. ✅ Our solution is correct and complete
+3. ✅ We follow project conventions
+4. ✅ We don't duplicate or conflict with existing work
+5. ✅ Tests prove correctness
+6. ✅ Documentation explains the solution
+
+---
+
+## Branch 1: feature/docs-contributing-guide
+
+### Problem Statement (Issue #718)
+
+**Requested by**: @purva-thakre  
+**Issue**: Contributors don't know how to add documentation  
+
+**Specific Requirements from Issue:**
+- [ ] How to add a page in the user guide?
+- [ ] How to add an example to the docs gallery?
+- [ ] Must be added to `docs/contributor_guide.rst`
+- [ ] Explain `.rst` format
+- [ ] Explain `jupyter-sphinx` for executable code
+- [ ] Explain `sphinxcontrib-bibtex` for references
+- [ ] Note: Gallery notebooks must have outputs cleared
+
+### Our Solution Analysis
+
+**What we added:**
+1. ✅ Section: "Contributing to Documentation"
+2. ✅ Subsection: "How to add a page to the user guide"
+3. ✅ Subsection: "How to add an example to the docs gallery"
+4. ✅ Explains RST format with examples
+5. ✅ Explains jupyter-sphinx usage
+6. ✅ Explains sphinxcontrib-bibtex with examples
+7. ✅ Emphasizes clearing notebook outputs
+8. ✅ Includes best practices and common pitfalls
+
+**Location**: Added to `docs/contributor_guide.rst` line 161-396 (237 new lines)
+
+### Validation Checklist
+
+#### 1. Does it solve the stated problem?
+- [ ] Addresses "How to add a page in the user guide?"
+- [ ] Addresses "How to add an example to the docs gallery?"
+- [ ] Covers all points mentioned in issue description
+- [ ] **Action**: Re-read issue #718 and map each requirement to our content
+
+#### 2. Does it follow project conventions?
+- [ ] Check existing RST structure in `docs/contributor_guide.rst`
+- [ ] Match heading hierarchy (= for main, - for sub, ~ for subsub)
+- [ ] Use same code-block style as existing docs
+- [ ] Follow link format used elsewhere
+- [ ] **Action**: Compare our style to lines 1-160 of contributor_guide.rst
+
+#### 3. Does it duplicate existing content?
+- [ ] Check if docs/README has similar content
+- [ ] Check if any user guide pages explain this
+- [ ] Search for "jupyter-sphinx" mentions elsewhere
+- [ ] Search for "gallery" mentions in docs
+- [ ] **Action**: `grep -r "jupyter-sphinx\|gallery\|footcite" docs/`
+
+#### 4. Is it technically accurate?
+- [ ] Test: Create a dummy user guide page following our instructions
+- [ ] Test: Create a dummy gallery notebook following our instructions
+- [ ] Verify: All paths mentioned are correct (`docs/user_guide/`, `docs/gallery/`)
+- [ ] Verify: All commands work (`make html`, `jupyter nbconvert`)
+- [ ] **Action**: Actually execute the workflow we documented
+
+#### 5. Does it integrate well?
+- [ ] Placement in file makes sense (after "Merge the PR" section)
+- [ ] References to other sections are valid
+- [ ] Links to external docs are current
[…]

[…]
