+++ pull/720/_sources/gallery/qrisp_integration.ipynb.txt
@@ -409,17 +409,17 @@
     "    filled_block_graphs = block_graph.fill_ports_for_minimal_simulation()\n",
     "    \n",
     "    # Show all block_graphs in minimal set\n",
-    "    for i, block_graph in enumerate(filled_block_graphs):\n",
+    "    for i, filled_block_graph in enumerate(filled_block_graphs):\n",
     "\n",
-    "        for j, correlation_surface in enumerate(block_graph.observables):\n",
-    "            block_graph.graph.view_as_html(\n",
+    "        for j, correlation_surface in enumerate(filled_block_graph.observables):\n",
+    "            filled_block_graph.graph.view_as_html(\n",
     "                pop_faces_at_directions=(\"-Y\", \"+X\"),\n",
-    "                show_correlation_surface=block_graph.observables[j],\n",
+    "                show_correlation_surface=filled_block_graph.observables[j],\n",
     "            )\n",
     "\n",
-    "            html = block_graph.graph.view_as_html(\n",
+    "            html = filled_block_graph.graph.view_as_html(\n",
     "                pop_faces_at_directions=(\"-Y\", \"+X\"),\n",
-    "                show_correlation_surface=block_graph.observables[j],\n",
+    "                show_correlation_surface=filled_block_graph.observables[j],\n",
     "            )\n",
     "\n",
     "            display(html)\n",
