### 2025-10-11 08ff44b
chore: only include src for ty checking
