### 2025-10-11 b6be094
feat(visualization): Add opacity controls for BlockGraph visualizations
- Add node_opacity, edge_opacity, hadamard_opacity parameters to PyZX plotting functions
- Update DrawerConfiguration with opacity control attributes
- Maintain backward compatibility with default opacity=1.0
- Update matplotlib scatter() and plot() calls with alpha parameter
- Comprehensive parameter validation and documentation

Addresses issue #690 for BlockGraph visualization opacity control
