+++ TESTING_RESULTS_MAC.md
@@ -3,8 +3,8 @@
 ## Branch: feature/improve-error-messages
 
 ### ✅ Test 1: Unit Tests
-**Command**: `pytest src/tqec/interop/pyzx/positioned_test.py -v`  
-**Result**: **PASS** - All 6 tests passed in 1.32s  
+**Command**: `pytest src/tqec/interop/pyzx/positioned_test.py -v`
+**Result**: **PASS** - All 6 tests passed in 1.32s
 
 **Tests executed:**
 - `test_positions_not_specified` ✅
@@ -14,7 +14,7 @@
 - `test_y_connect_in_space` ✅
 - `test_3d_corner` ✅
 
-**Findings**: 
+**Findings**:
 - All test regex patterns correctly match new error messages
 - Tests validate behavior, not just message content
 - No regressions detected
@@ -22,8 +22,8 @@
 ---
 
 ### ✅ Test 2: Error Message Quality
-**Command**: Manual error triggering with assertions  
-**Result**: **PASS** - All error messages include detailed context  
+**Command**: Manual error triggering with assertions
+**Result**: **PASS** - All error messages include detailed context
 
 **Validated Messages:**
 1. **Vertex ID Mismatch**:
@@ -43,8 +43,8 @@
 ---
 
 ### ✅ Test 3: Pre-commit Hooks
-**Command**: `pre-commit run --files <modified files>`  
-**Result**: **PASS** - All hooks passed  
+**Command**: `pre-commit run --files <modified files>`
+**Result**: **PASS** - All hooks passed
 
 **Hooks validated:**
 - `check yaml` ✅ Skipped (no YAML files)
@@ -56,7 +56,7 @@
 - `ruff check` ✅ Passed (line length, style)
 - `ruff format` ✅ Passed (formatting)
 
-**Findings**: 
+**Findings**:
 - All files conform to project style
 - Line length < 100 characters maintained
 - No formatting issues
@@ -64,20 +64,20 @@
 ---
 
 ### ⏳ Test 4: Full Integration Tests
-**Status**: Ready to run on Threadripper  
-**Command**: `pytest src/tqec/interop/ -v --tb=short`  
+**Status**: Ready to run on Threadripper
+**Command**: `pytest src/tqec/interop/ -v --tb=short`
 
 **Expected scope**:
 - All positioned.py tests
-- All collada read_write tests  
+- All collada read_write tests
 - Integration with rest of interop module
 - ~50-100 tests estimated
 
 ---
 
 ### ⏳ Test 5: Coverage Analysis
-**Status**: Ready to run on Threadripper  
-**Command**: `pytest src/tqec/interop/ --cov=src/tqec/interop --cov-report=html`  
+**Status**: Ready to run on Threadripper
+**Command**: `pytest src/tqec/interop/ --cov=src/tqec/interop --cov-report=html`
 
 **Goal**: Verify our changes don't reduce coverage
 
@@ -86,8 +86,8 @@
 ## Branch: feature/docs-contributing-guide
 
 ### ✅ Test 1: RST Syntax Validation
-**Method**: docutils parsing  
-**Result**: **PASS** - 773 nodes parsed successfully  
+**Method**: docutils parsing
+**Result**: **PASS** - 773 nodes parsed successfully
 
 **Previous test output**:
 ```
@@ -105,15 +105,15 @@ Document has 773 nodes
 ---
 
 ### ⏳ Test 2: Full Docs Build
-**Status**: Ready to run on Threadripper (takes 10-30 min)  
-**Command**: `cd docs && make clean && make html`  
+**Status**: Ready to run on Threadripper (takes 10-30 min)
+**Command**: `cd docs && make clean && make html`
 
[…]

[…]
