### 2025-10-12 be574b9
Set background color to white in HTML viewer
