+++ pull/720/_sources/gallery/qiskit_integration.ipynb.txt
@@ -1,605 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "695f75d2",
-   "metadata": {},
-   "source": [
-    "# Importing a Qiskit circuit and running simulations with it"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "5986b007",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "import random\n",
-    "import numpy as np\n",
-    "import matplotlib.pyplot as plt\n",
-    "\n",
-    "from IPython.display import display\n",
-    "from pathlib import Path\n",
-    "\n",
-    "ASSETS_FOLDER = Path(\"../../assets/\").resolve()\n",
-    "\n",
-    "# To display interactive 3D visualisations, uncomment the following line.\n",
-    "# %matplotlib widget"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "27f0f36b",
-   "metadata": {},
-   "source": [
-    "This notebook illustrates the process of importing into TQEC circuits originally created with Qiskit.\n",
-    "\n",
-    "The notebook assumes you have already designed a Qiskit circuit and saved it using Qiskit's `.qpy` format. \n",
-    "\n",
-    "> **NB!** Run `uv add --only-group integration` to install the libraries used in this notebook (ignore if you installed all dependency groups when setting up TQEC)."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "2be0a675",
-   "metadata": {},
-   "source": [
-    "## Load Qiskit circuit"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "2f541923",
-   "metadata": {},
-   "source": [
-    "The first step is, naturally, to load the Qiskit circuit."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "66009227",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "from qiskit import qasm2, qpy\n",
-    "\n",
-    "# NAME OF CIRCUIT\n",
-    "circuit_name = \"steane\"\n",
-    "\n",
-    "# LOAD CIRCUIT\n",
-    "path_to_c = ASSETS_FOLDER / f\"{circuit_name}_qiskit.qpy\"\n",
-    "with open(path_to_c, \"rb\") as f:\n",
-    "    qc = qpy.load(f)[0]\n",
-    "\n",
-    "# CONVERT to QASM\n",
-    "qasm_str = qasm2.dumps(qc)"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "3fac4bac",
-   "metadata": {},
-   "source": [
-    "We can print the original circuit for inspection."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "3bdc28c4",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "print(f\"\\n=> Imported Qiskit circuit of name {circuit_name.capitalize()}:\\n\")\n",
-    "print(qc) if qc else print(\"ERROR. Failed to load a circuit\")"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "ea56a685",
-   "metadata": {},
-   "source": [
-    "And we should also inspect the QASM string resulting from the circuit."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "bee986d0",
-   "metadata": {},
-   "outputs": [],
-   "source": [
[…]
