### 2025-10-13 349fd16
documenting my understanding
