+++ src/tqec/compile/blocks/block.py
@@ -30,6 +30,24 @@ class Block(ABC):
 
     A block represents a unit of quantum computation with a defined spatial
     and temporal footprint that scales with the scaling factor ``k``.
+
+    This abstract base class defines the interface that all block types must
+    implement. Concrete implementations include:
+
+    - :class:`LayeredBlock`: Blocks represented as sequences of layers, suitable
+      for standard memory blocks, pipes, and other layer-synchronous operations.
+    - :class:`InjectedBlock`: Blocks represented as raw circuits with flow interfaces,
+      used for operations like Y-basis measurements that cannot be decomposed into
+      simple layer sequences.
+
+    The abstraction allows the compilation framework to handle blocks uniformly
+    while supporting both template-based layer composition and direct circuit
+    injection with flow-based detector annotation.
+
+    See Also:
+        - :class:`LayeredBlock` for layer-based block implementation
+        - :class:`InjectedBlock` for injection-based block implementation
+        - :mod:`tqec.compile.blocks` module documentation for usage guidance
     """
 
     @property
@@ -221,10 +239,37 @@ def merge_parallel_block_layers(
 class CircuitWithInterface:
     """A quantum circuit with its expected flow interface.
 
-    Attributes:
-        circuit: the quantum circuit.
-        interface: the expected flow interface of the circuit.
+    This dataclass pairs a Stim circuit with a flow interface from Gidney's gen
+    library, enabling :class:`InjectedBlock` instances to specify both the quantum
+    operations and their stabilizer flow contracts.
 
+    The interface describes which stabilizers (Pauli products) enter and exit the
+    circuit, allowing the compilation framework to properly connect injected blocks
+    with tree-generated circuits while maintaining correct detector annotations.
+
+    Attributes:
+        circuit: The quantum circuit implementing the block's operations.
+        interface: The flow interface specifying input/output stabilizers as
+            :class:`gen.ChunkInterface` with ports (stabilizer Pauli products)
+            and discards (rejected flows).
+
+    Example:
+        For a Y-basis measurement block::
+
+            circuit = stim.Circuit()
+            # ... build Y-measurement circuit with gate teleportation ...
+
+            interface = gen.ChunkInterface(ports=[
+                gen.PauliMap({(0,0): 'X', (1,0): 'X'}),  # X stabilizer
+                gen.PauliMap({(0,0): 'Z', (1,0): 'Z'}),  # Z stabilizer
+            ])
+
+            circuit_with_interface = CircuitWithInterface(circuit, interface)
+
+    See Also:
+        - :class:`InjectedBlock` for usage in block representation
+        - :class:`InjectionFactory` protocol for circuit generation
+        - :mod:`tqec.compile.tree.injection` for injection mechanics
     """
 
     circuit: stim.Circuit
@@ -253,6 +298,39 @@ class CircuitWithInterface:
 
 
 class InjectionFactory(Protocol):
+    """Protocol for callables that generate scalable injected circuits.
+
+    An InjectionFactory is a callable that produces a :class:`CircuitWithInterface`
+    from a scaling factor and optional observable annotations. This protocol enables
+    :class:`InjectedBlock` to defer circuit generation until compilation time when
+    the actual scaling factor is known.
+
+    Implementations of this protocol typically encapsulate the logic for generating
+    circuits that don't fit the layer model, such as Y-basis measurements via gate
+    teleportation or other protocols with complex temporal dependencies.
+
+    The factory pattern allows the same block specification to be used at different
+    scaling factors without regenerating the factory itself.
+
+    Example:
+        A Y-basis measurement factory might be implemented as::
+
+            def make_y_half_cube(
+                k: int,
+                annotate_observables: list[int] | None
+            ) -> CircuitWithInterface:
+                # Build circuit for Y-basis measurement at scaling factor k
[…]
