### 2025-10-13 cbc44a1
restoring layer doc info
