### 2025-10-14 fb0e26e
add z_coordinate property
