### 2025-10-14 f1694da
documenting my understanding of mids
