### 2025-10-14 52ae61c
data class structure for transition step geometry
