+++ src/tqec/compile/specs/library/generators/y_basis.py
@@ -6,6 +6,8 @@ with modification to support different boundary conditions and conventions.
 
 import functools
 from collections.abc import Callable, Iterable, Set
+from dataclasses import dataclass
+from enum import Enum
 from typing import Any, Literal, cast
 
 import gen
@@ -18,6 +20,189 @@ from tqec.utils.scale import LinearFunction, PhysicalQubitScalable2D
 
 DIRS: list[complex] = [(0.5 + 0.5j) * 1j**d for d in range(4)]
 DR, DL, UL, UR = DIRS
+
+class DiagonalType(Enum):
+    """Type of diagonal in the surface code patch.
+
+    MAIN: Top-left to bottom-right diagonal where real + imag = constant
+    ANTI: Bottom-left to top-right diagonal where real - imag = constant
+    """
+    MAIN = "main"
+    ANTI = "anti"
+
+
+@dataclass(frozen=True)
+class TwistLine:
+    """Represents a twist/domain wall line in the Y-basis transition.
+
+    A twist line is where the code patch undergoes a topological twist during
+    the Y-basis transition, changing from a normal surface code patch to a
+    degenerate boundary patch.
+
+    Attributes:
+        diagonal: Whether this is a main diagonal or anti-diagonal
+        position: Which diagonal line (as an offset from origin)
+        orientation: The stabilizer basis orientation along the twist (X or Z)
+    """
+    diagonal: DiagonalType
+    position: int
+    orientation: Basis
+
+    def contains(self, m: complex) -> bool:
+        """Check if measurement qubit m lies on this twist line.
+
+        Args:
+            m: Complex number representing qubit position
+
+        Returns:
+            True if m is on the twist line
+        """
+        if self.diagonal == DiagonalType.MAIN:
+            return int(m.real + m.imag) == self.position
+        else:
+            return int(m.real - m.imag) == self.position
+
+    def is_before(self, m: complex) -> bool:
+        """Check if measurement qubit m is before (above/left of) the twist line."""
+        if self.diagonal == DiagonalType.MAIN:
+            return m.real + m.imag < self.position
+        else:
+            return m.real - m.imag < self.position
+
+    def is_after(self, m: complex) -> bool:
+        """Check if measurement qubit m is after (below/right of) the twist line."""
+        if self.diagonal == DiagonalType.MAIN:
+            return m.real + m.imag > self.position
+        else:
+            return m.real - m.imag > self.position
+
+
+@dataclass(frozen=True)
+class BoundaryRegion:
+    """Represents boundary regions in the surface code patch.
+
+    Attributes:
+        top_row: Set of qubits on the top boundary (imag = -0.5)
+        bottom_row: Set of qubits on the bottom boundary
+        left_col: Set of qubits on the left boundary (real = -0.5)
+        right_col: Set of qubits on the right boundary
+    """
+    top_row: frozenset[complex]
+    bottom_row: frozenset[complex]
+    left_col: frozenset[complex]
+    right_col: frozenset[complex]
+
+    def is_boundary(self, m: complex) -> bool:
+        """Check if qubit m is on any boundary."""
+        return m in self.top_row or m in self.bottom_row or m in self.left_col or m in self.right_col
+
+
+@dataclass(frozen=True)
+class TransitionGeometry:
+    """Geometric description of a Y-basis transition circuit.
+
+    This class encapsulates all the geometric information needed to understand
+    and generate a Y-basis transition circuit, making the implicit geometric
+    constraints explicit.
+
+    Attributes:
+        distance: Code distance
+        top_boundary_basis: Basis of the top boundary 
+        convention: Boundary convention 
[…]
