+++ docs/integration/index.rst
@@ -1,0 +1,13 @@
+.. _integration-reference-label:
+
+Integration Gallery
+===================
+
+This section contains examples of how to use (NISQ) circuits designed with external frameworks including, but not limited to, Qiskit and Cirq.
+
+.. nbintegration::
+
+   cirq.ipynb
+   pytket.ipynb
+   qiskit.ipynb
+   qrisp.ipynb
