+++ docs/gallery/index.rst
@@ -12,5 +12,3 @@ This section contains a collection of constructed logical computations and their
    move_rotation.ipynb
    three_cnots.ipynb
    steane_encoding.ipynb
-   qiskit_integration.ipynb
-   qrisp_integration.ipynb
