### 2025-10-14 37b4d18
eliminated unecessary flattening, clarifying comments, used id() flow deduplication
