### 2025-10-14 f4f9c96
docstrings, z-slice pregrouping for scheduling along tree
