### 2025-10-14 b2b3a51
z coordinate
