### 2025-10-15 1b1fe5e
Add `pymatching>=2.3.1` to docs dependency group and bump pymatching version in dev group
