### 2025-10-15 90f7f1d
Do not bump major version of detector database
