### 2025-10-15 4170bb0
Fix trailing whitespace with pre-commit
