### 2025-10-15 dd04ef2
remove empty moments after building circuit
