### 2025-10-15 ec96ff2
fix(detector-database): some pickle load errors not handled (#734)
As
[described](https://github.com/tqec/tqec/issues/724#issuecomment-3397118181)
by @KabirDubey, there's an issue in loading faulty/old-version detector
database.

Previously Adrien has added some exception handling, when failing to
load a db, we'll raise warining, move the db to a `faulty_xx` file, and
return an empty database.


https://github.com/tqec/tqec/blob/99e0b66d1b8359038604867c8a02833f2eadf350/src/tqec/compile/detectors/database.py#L223-L238

However, it only handles `PickleError`, according to [the
documentation](htt
[…]
