+++ pull/719/_sources/_autosummary/tqec.utils.ExtendedBasis.rst.txt
@@ -1,0 +1,27 @@
+ExtendedBasis
+=============
+
+.. currentmodule:: tqec.utils
+
+.. autoclass:: ExtendedBasis
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~ExtendedBasis.X
+      ~ExtendedBasis.Y
+      ~ExtendedBasis.Z
+      ~ExtendedBasis.H
+   
+   
+
+   
+   
+   
