+++ pull/719/_sources/_autosummary/tqec.utils.PauliBasis.rst.txt
@@ -1,0 +1,37 @@
+PauliBasis
+==========
+
+.. currentmodule:: tqec.utils
+
+.. autoclass:: PauliBasis
+
+   
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~PauliBasis.to_extended_basis
+   
+   
+
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~PauliBasis.X
+      ~PauliBasis.Y
+      ~PauliBasis.Z
+   
+   
+
+   
+   
+   .. rubric:: Detailed methods
+
+   
+   .. automethod:: to_extended_basis
+   
+   
