+++ pull/719/_sources/_autosummary/tqec.gallery.rst.txt
@@ -1,0 +1,36 @@
+﻿tqec.gallery
+============
+
+.. currentmodule:: tqec.gallery
+
+.. automodule:: tqec.gallery
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+      :toctree:
+   
+      cnot
+      cz
+      memory
+      move_rotation
+      s_gate_teleportation
+      stability
+      steane_encoding
+      three_cnots
+   
+   
+
+   
+   
+   
+
+   
+   
+   
