+++ pull/719/_sources/gallery/s_gate.ipynb.txt
@@ -1,0 +1,306 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Logical S Gate\n",
+    "\n",
+    "This notebook demonstrates the logical S gate with inplace Y basis measurement and gate teleportation [<cite data-footcite-t=\"Gidney_inplace_access_2024\"></cite>]."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Construction\n",
+    "\n",
+    "S gate can be implemented by gate teleportation:\n",
+    "\n",
+    "- $M_{ZZ}$ measurement between the target qubit and an ancilla qubit initialized in the $|+\\rangle$ state.\n",
+    "- Y basis measurement of the ancilla qubit.\n",
+    "\n",
+    "`tqec` provides builtin functions `tqec.gallery.s_gate_teleportation` to construct it."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "from tqec import PauliBasis\n",
+    "from tqec.gallery import s_gate_teleportation\n",
+    "\n",
+    "graph = s_gate_teleportation(PauliBasis.Y)\n",
+    "graph.view_as_html()"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "S gate realizes $Y \\rightarrow X$ and $Z \\rightarrow Z$ stabilizer flows, ignoring the signs. The following correlation surface shows that the computation maps the $Y$ operator at the bottom input to the $X$ operator at the top output."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "correlation_surfaces = graph.find_correlation_surfaces()\n",
+    "graph.view_as_html(\n",
+    "    pop_faces_at_directions=(\"-Y\",),\n",
+    "    show_correlation_surface=correlation_surfaces[0],\n",
+    ")"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Circuit\n",
+    "\n",
[…]
