### 2025-10-15 377b180
Better management of correlated dependency groups
