+++ pyproject.toml
@@ -51,6 +51,7 @@ dependencies = [
     "stim<=1.15.0",
     "svg-py<=1.8.0",
     "tqecd<=0.1.2",
+    "ty>=0.0.1a22",
     "typing-extensions<=4.15.0",
 ]
 
@@ -69,6 +70,12 @@ test = [
   "scipy-stubs<=1.16.0.0",
 ]
 
+integration = [
+  "topologiq@git+https://github.com/tqec/topologiq.git",
+  "qiskit>=2.1.2",
+  "qrisp>=0.7.9"
+]
+
 docs = [
   "sphinx<=8.1.3",
   "myst-parser<=4.0.1",
@@ -81,8 +88,7 @@ docs = [
   "sphinx-copybutton<=0.5.2",
   "sphinxcontrib-bibtex<=2.64",
   "jupyter-sphinx<=0.5.3",
-  "qiskit>=2.1.1",
-  "topologiq@git+https://github.com/tqec/topologiq.git"
+  {include-group = "integration"}
 ]
 
 bench = ["pyinstrument"]
@@ -106,12 +112,6 @@ dev = [
     "sphinx<=8.1.3",
     "sphinxcontrib-mermaid<=1.0.0",
     "types-networkx<=3.5.0.20250610",
-]
-
-integration = [
-  "topologiq@git+https://github.com/tqec/topologiq.git",
-  "qiskit>=2.1.2",
-  "qrisp>=0.7.9"
 ]
 
 all = [
