### 2025-10-15 7206ddf
Add integration gallery to main docs index & update integration gallery index (testing options for structuring docs)
