### 2025-10-15 46927d2
Clean-up & improve readability of integration notebooks
