+++ pull/720/_sources/index.rst.txt
@@ -1,0 +1,17 @@
+.. See https://stackoverflow.com/a/69134918 and
+.. https://myst-parser.readthedocs.io/en/latest/faq/index.html#include-markdown-files-into-an-rst-file
+.. include:: ../README.md
+   :parser: myst_parser.sphinx_
+
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   Introduction <self>
+   User Guide <user_guide/index>
+   Gallery <gallery/index>
+   Integration <integration/index>
+   Contributor Guide <contributor_guide>
+   API Reference <api>
+   Ethical Notice <ethical_notice>
