+++ pull/720/_sources/integration/index.rst.txt
@@ -1,0 +1,11 @@
+.. _integration-reference-label:
+
+Integration Gallery
+===================
+
+Examples of how to use NISQ circuits designed with external frameworks (e.g., Qiskit, Qrisp, etc.).
+
+.. nbgallery::
+
+   qiskit.ipynb
+   qrisp.ipynb
