### 2025-10-15 708b802
Update src/tqec/compile/specs/library/generators/y_basis.py
Co-authored-by: Yiming Zhang <61700160+inmzhang@users.noreply.github.com>
