### 2025-10-15 e1addeb
deleted unecessary attribute check
