### 2025-10-15 a5bae1a
region split twist line class function
