+++ src/tqec/compile/blocks/__init__.py
@@ -17,15 +17,17 @@ abstract base class:
 1. **LayeredBlock** - Represents blocks as a sequence of layers that can be instances
    of :class:`~tqec.compile.blocks.layers.atomic.base.BaseLayer` or
    :class:`~tqec.compile.blocks.layers.composed.base.BaseComposedLayer`. This is the
-   standard representation for blocks that follow a layer-synchronous schedule. 
-   
-   Note that different blocks may have different schedules. For example, the regular memory block follows the schedule:
+   standard representation for blocks that follow a layer-synchronous schedule.
+
+   Note that different blocks may have different schedules. For example, the
+   regular memory block follows the schedule:
 
     1. initialisation layer,
     2. repeat [memory layer],
     3. measurement layer,
 
-whereas a spatial pipe in the ``Y`` axis needs to alternate plaquettes in its repeated layer, leading to a schedule that is::
+whereas a spatial pipe in the ``Y`` axis needs to alternate plaquettes in its
+   repeated layer, leading to a schedule that is::
 
     1. initialisation layer,
     2. repeat [memory layer 1 alternated with memory layer 2],
@@ -34,8 +36,8 @@ whereas a spatial pipe in the ``Y`` axis needs to alternate plaquettes in its re
 Both of these can be represented as LayeredBlock instances.
 
 2. **InjectedBlock** - Represents blocks via direct circuit representation using
-   :class:`~tqec.compile.blocks.block.CircuitWithInterface`. 
-   
+   :class:`~tqec.compile.blocks.block.CircuitWithInterface`.
+
 Block Compilation
 ==================
 
@@ -45,9 +47,10 @@ are encoded in the same moment (i.e., between the same two ``TICK`` instructions
 in the resulting ``.stim`` file.
 
 For LayeredBlock instances, this is handled by the layer merging system. For
-InjectedBlock instances, this is handled by the :class:`~tqec.compile.tree.injection.InjectionBuilder`
-which weaves injected circuits into layer-generated circuits while maintaining
-correct detector annotations via flow interfaces from Gidney's gen library.
+InjectedBlock instances, this is handled by the
+:class:`~tqec.compile.tree.injection.InjectionBuilder` which weaves injected
+circuits into layer-generated circuits while maintaining correct detector
+annotations via flow interfaces from Gidney's gen library.
 
 See Also
 ========
