### 2025-10-16 28f6e8d
Y half cube updates (#736)
- Documented explanations of `InjectedBlock`
- Replaced  `additional_metadata` with a `z` coordinate for linking
- Added a class data structure for organizing the Y basis transition
round circuit generation
- Eliminated unnecessary flattening by keeping REPEAT loops intact

---------

Co-authored-by: Yiming Zhang <61700160+inmzhang@users.noreply.github.com>
