+++ src/tqec/compile/specs/library/generators/y_basis.py
@@ -6,6 +6,8 @@ with modification to support different boundary conditions and conventions.
 
 import functools
 from collections.abc import Callable, Iterable, Set
+from dataclasses import dataclass
+from enum import Enum
 from typing import Any, Literal, cast
 
 import gen
@@ -18,6 +20,239 @@ from tqec.utils.scale import LinearFunction, PhysicalQubitScalable2D
 
 DIRS: list[complex] = [(0.5 + 0.5j) * 1j**d for d in range(4)]
 DR, DL, UL, UR = DIRS
+
+
+class DiagonalType(Enum):
+    """Type of diagonal in the surface code patch.
+
+    MAIN: Top-left to bottom-right diagonal where real + imag = constant
+    ANTI: Bottom-left to top-right diagonal where real - imag = constant
+    """
+
+    MAIN = "main"
+    ANTI = "anti"
+
+
+@dataclass(frozen=True)
+class TwistLine:
+    """Represents a twist line in the Y-basis transition.
+
+    Attributes:
+        diagonal: Whether this is a main diagonal or anti-diagonal
+        position: Which diagonal line (as an offset from origin)
+        orientation: The stabilizer basis orientation along the twist (X or Z)
+
+    """
+
+    diagonal: DiagonalType
+    position: int
+    orientation: Basis
+
+    def contains(self, m: complex) -> bool:
+        """Check if measurement qubit m lies on this twist line.
+
+        Args:
+            m: Complex number representing qubit position
+
+        Returns:
+            True if m is on the twist line
+
+        """
+        if self.diagonal == DiagonalType.MAIN:
+            return int(m.real + m.imag) == self.position
+        else:
+            return int(m.real - m.imag) == self.position
+
+    def is_before(self, m: complex) -> bool:
+        """Check if measurement qubit m is before (above/left of) the twist line."""
+        if self.diagonal == DiagonalType.MAIN:
+            return m.real + m.imag < self.position
+        else:
+            return m.real - m.imag < self.position
+
+    def is_after(self, m: complex) -> bool:
+        """Check if measurement qubit m is after (below/right of) the twist line."""
+        if self.diagonal == DiagonalType.MAIN:
+            return m.real + m.imag > self.position
+        else:
+            return m.real - m.imag > self.position
+
+    def split_qubits_by_region(
+        self, qubits: set[complex]
+    ) -> tuple[set[complex], set[complex], set[complex]]:
+        """Split qubits into upper-left (before), middle (on/near), down-right (after) regions.
+
+        The middle region includes qubits on the twist line (position) and
+        qubits one position before it (position-1), which form the transition zone.
+
+        Returns:
+            (upper_left, middle, down_right) sets of qubits
+
+        """
+        upper_left: set[complex] = set()
+        middle: set[complex] = set()
+        down_right: set[complex] = set()
+
+        for q in qubits:
+            if self.diagonal == DiagonalType.MAIN:
+                coord_sum = q.real + q.imag
+                # Middle includes twist line (position) and one before (position-1)
+                if coord_sum < self.position - 1:
+                    upper_left.add(q)
+                elif coord_sum > self.position:
+                    down_right.add(q)
+                else:
+                    middle.add(q)
+            else:  # ANTI diagonal
+                coord_diff = q.real - q.imag
+                if coord_diff < self.position - 1:
+                    upper_left.add(q)
+                elif coord_diff > self.position:
+                    down_right.add(q)
[…]
