### 2025-10-16 fb6526b
Improve readability of notebooks and integration gallery index description
