+++ docs/integration/index.rst
@@ -3,7 +3,7 @@
 Integration Gallery
 ===================
 
-Examples of how to use NISQ circuits designed with external frameworks (e.g., Qiskit, Qrisp, etc.).
+Examples of how to use circuits designed with external frameworks (e.g., Qiskit, Qrisp, etc.).
 
 .. nbgallery::
 
