### 2025-10-16 b333649
Separate computation & integration galleries to single gallery with two sections
