+++ docs/integration/index.rst
@@ -1,11 +1,0 @@
-.. _integration-reference-label:
-
-Integration Gallery
-===================
-
-Examples of how to use circuits designed with external frameworks (e.g., Qiskit, Qrisp, etc.).
-
-.. nbgallery::
-
-   qiskit.ipynb
-   qrisp.ipynb
