### 2025-10-16 08acc94
Revert the changes of class abstraction
