### 2025-10-17 b8ef1e5
Do not early return when the Y-port check passes and revert changes made in #704
