### 2025-10-17 020a151
improve test coverage
