+++ src/tqec/compile/observables/abstract_observable_test.py
@@ -1,10 +1,17 @@
+from typing import Literal
+
+import pytest
+
 from tqec.compile.observables.abstract_observable import (
     AbstractObservable,
     CubeWithArms,
     PipeWithArms,
+    _check_correlation_surface_validity,
     compile_correlation_surface_to_abstract_observable,
 )
+from tqec.compile.specs.enums import SpatialArms
 from tqec.computation.block_graph import BlockGraph
+from tqec.computation.correlation import CorrelationSurface, ZXEdge, ZXNode
 from tqec.computation.cube import Cube, ZXCube
 from tqec.computation.pipe import Pipe
 from tqec.gallery.cnot import cnot
@@ -13,6 +20,7 @@ from tqec.gallery.move_rotation import move_rotation
 from tqec.gallery.stability import stability
 from tqec.gallery.three_cnots import three_cnots
 from tqec.utils.enums import Basis, PauliBasis
+from tqec.utils.exceptions import TQECError
 from tqec.utils.position import Position3D
 
 
@@ -171,3 +179,92 @@ def test_abstract_observable_for_temporal_hadamard() -> None:
     )
     assert len(observable.top_readout_cubes) == 1
     assert len(observable.temporal_hadamard_pipes) == 1
+
+
+def test_x_correlation_surface_on_x_node_four_arms() -> None:
+    g = BlockGraph()
+    n1 = g.add_cube(Position3D(0, 0, 0), "ZZX")
+    n2 = g.add_cube(Position3D(1, 0, 0), "XZX")
+    n3 = g.add_cube(Position3D(-1, 0, 0), "XZX")
+    n4 = g.add_cube(Position3D(0, 1, 0), "ZXX")
+    n5 = g.add_cube(Position3D(0, -1, 0), "ZXX")
+    g.add_pipe(n1, n2)
+    g.add_pipe(n1, n3)
+    g.add_pipe(n1, n4)
+    g.add_pipe(n1, n5)
+    zx = g.to_zx_graph()
+    center = zx.p2v[Position3D(0, 0, 0)]
+    corrlation_surface = CorrelationSurface(
+        frozenset(
+            {
+                ZXEdge(ZXNode(center, Basis.X), (ZXNode(n, Basis.X)))
+                for n in zx.g.vertex_set()
+                if n != center
+            }
+        )
+    )
+    obs = compile_correlation_surface_to_abstract_observable(g, corrlation_surface)
+    assert len(obs.top_readout_cubes) == 6
+    assert (
+        CubeWithArms(
+            Cube(Position3D(0, 0, 0), ZXCube.from_str("ZZX")), SpatialArms.LEFT | SpatialArms.DOWN
+        )
+        in obs.top_readout_cubes
+    )
+    assert (
+        CubeWithArms(
+            Cube(Position3D(0, 0, 0), ZXCube.from_str("ZZX")), SpatialArms.UP | SpatialArms.RIGHT
+        )
+        in obs.top_readout_cubes
+    )
+
+
+def _make_correlation_surface(edges: list[tuple[int, Literal["Z", "X"]]]) -> CorrelationSurface:
+    return CorrelationSurface(
+        frozenset(
+            {
+                ZXEdge(ZXNode(u[0], Basis(u[1])), ZXNode(v[0], Basis(v[1])))
+                for u, v in zip(edges[::2], edges[1::2])
+            }
+        )
+    )
+
+
+@pytest.mark.parametrize("cross_center", ("ZZX", "XXZ"))
+def test_correlation_surface_validity_raise(cross_center: str) -> None:
+    g = BlockGraph()
[…]

[…]
