### 2025-10-17 74902ac
correlation surface finding: reduce generator set redundancy; switch from DFS to BFS
