### 2025-10-18 af9b355
correlation surface finding: reduce redundancy in generating new correlation surfaces
