### 2025-10-18 2bb8687
correlation surface finding: avoid the extremely slow dict deepcopy
