### 2025-10-18 8cad845
correlation surface finding: adopt a more natural representation of Pauli and avoid using Enum
