+++ src/tqec/interop/pyzx/correlation_experimental.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import operator
 from collections.abc import Iterable
 from copy import copy
 from fractions import Fraction
@@ -15,7 +16,7 @@ from pyzx.utils import FractionLike, VertexType
 
 from tqec.computation.correlation import CorrelationSurface, ZXEdge, ZXNode
 from tqec.interop.pyzx.utils import is_boundary, is_hardmard, is_s, is_x_no_phase, is_z_no_phase
-from tqec.utils.enums import Basis, Pauli
+from tqec.utils.enums import Basis
 from tqec.utils.exceptions import TQECError
 
 
@@ -121,52 +122,94 @@ def find_correlation_surfaces(
     return sorted(correlation_surfaces, key=lambda x: sorted(x.span))
 
 
+class PauliGraph(dict[int, dict[int, tuple[bool, bool]]]):
+    """Correlation surface represented as X and Z supports on half-edges."""
+
+    def add_support(
+        self, zx_graph: GraphS, edge: tuple[int, int], xz_support: tuple[bool, bool]
+    ) -> PauliGraph:
+        """Add X and Z supports on the edge (u, v)."""
+        for (u, v), support in zip(
+            (edge, edge[::-1]),
+            (
+                xz_support,
+                xz_support[::-1] if is_hardmard(zx_graph, edge) else xz_support,
+            ),
+        ):
+            if u not in self:
+                self[u] = {}
+            self[u][v] = support
+        return self
+
+    def support_on_nodes(self, nodes: list[int]) -> tuple[tuple[bool], tuple[bool]]:
+        """Get the X and Z supports on the given nodes."""
+        return tuple(zip(*chain.from_iterable(self[n].values() for n in nodes)))
+
+    def __mul__(self, other: PauliGraph) -> PauliGraph:
+        new_pauli_graph = PauliGraph()
+        for (v, neighbors_a), neighbors_b in zip(self.items(), other.values()):
+            new_pauli_graph[v] = {}
+            for (n, support_a), support_b in zip(neighbors_a.items(), neighbors_b.values()):
+                new_pauli_graph[v][n] = tuple(map(operator.xor, support_a, support_b))
+        return new_pauli_graph
+
+
 def _find_correlation_surfaces_from_leaf(
     g: GraphS,
     leaf: int,
 ) -> list[CorrelationSurface]:
     """Find the correlation surfaces starting from a leaf node in the graph."""
-    bases_at_leaves: dict[int, Pauli] = {}
-    for closed_leaf in filter(
-        lambda v: g.vertex_degree(v) == 1 and not is_boundary(g, v), g.vertices()
-    ):
-        if is_s(g, closed_leaf):
-            bases_at_leaves[closed_leaf] = Pauli.Y
-        elif is_z_no_phase(g, closed_leaf):
-            bases_at_leaves[closed_leaf] = Pauli.X
-        else:
-            bases_at_leaves[closed_leaf] = Pauli.Z
-    neighbor = next(iter(g.neighbors(leaf)))
+    # bases_at_leaves: dict[int, Pauli] = {}
+    # for closed_leaf in filter(
+    #     lambda v: g.vertex_degree(v) == 1 and not is_boundary(g, v), g.vertices()
+    # ):
+    #     if is_s(g, closed_leaf):
+    #         bases_at_leaves[closed_leaf] = Pauli.Y
+    #     elif is_z_no_phase(g, closed_leaf):
+    #         bases_at_leaves[closed_leaf] = Pauli.X
+    #     else:
+    #         bases_at_leaves[closed_leaf] = Pauli.Z
[…]
