### 2025-10-18 8763a3e
correlation surface finding: avoid unnecessary copying
