### 2025-10-18 c48e6c0
correlation surface finding: multiply pauli graphs in batch
