### 2025-10-19 0c0bd94
correlation surface finding: reduce gaussian elimination overhead
