+++ src/tqec/interop/pyzx/correlation_experimental.py
@@ -145,16 +145,16 @@ class PauliGraph(dict[int, dict[int, tuple[bool, bool]]]):
         """Get the X and Z supports on the given nodes."""
         return tuple(zip(*chain.from_iterable(self[n].values() for n in nodes)))
 
-    def to_correlation_surface(self, g: GraphS) -> CorrelationSurface:
+    def to_correlation_surface(self, zx_graph: GraphS) -> CorrelationSurface:
         """Convert a Pauli graph to a correlation surface."""
         span = []
         bases = (Basis.X, Basis.Z)
-        for u, v in g.edges():
+        for u, v in zx_graph.edges():
             pauli_u = self[u][v]
             pauli_v = self[v][u]
             for basis_u, basis_v in product(range(2), repeat=2):
                 if (
-                    (is_hardmard(g, (u, v)) ^ (basis_u == basis_v))
+                    (is_hardmard(zx_graph, (u, v)) ^ (basis_u == basis_v))
                     and pauli_u[basis_u]
                     and pauli_v[basis_v]
                 ):
@@ -172,55 +172,20 @@ def _multiply_pauli_graphs(pauli_graphs: list[PauliGraph]) -> PauliGraph:
 
 
 def _find_correlation_surfaces_from_leaf(
-    g: GraphS,
+    zx_graph: GraphS,
     leaf: int,
 ) -> list[CorrelationSurface]:
     """Find the correlation surfaces starting from a leaf node in the graph."""
-    # bases_at_leaves: dict[int, Pauli] = {}
-    # for closed_leaf in filter(
-    #     lambda v: g.vertex_degree(v) == 1 and not is_boundary(g, v), g.vertices()
-    # ):
-    #     if is_s(g, closed_leaf):
-    #         bases_at_leaves[closed_leaf] = Pauli.Y
-    #     elif is_z_no_phase(g, closed_leaf):
-    #         bases_at_leaves[closed_leaf] = Pauli.X
-    #     else:
-    #         bases_at_leaves[closed_leaf] = Pauli.Z
-    # neighbor = next(iter(g.neighbors(leaf)))
-
-    # pauli_graphs = []
-    # for basis in (
-    #     [bases_at_leaves[leaf]]
-    #     if leaf in bases_at_leaves
-    #     else [Pauli.X, Pauli.Z, Pauli.Y]
-    # ) + [Pauli.I]:
-    #     graph = {}
-    #     graph[leaf] = {}
-    #     graph[leaf][neighbor] = basis
-    #     graph[neighbor] = {}
-    #     graph[neighbor][leaf] = _flip_pauli_based_on_edge_type(
-    #         g, (leaf, neighbor), basis
-    #     )
-    #     pauli_graphs.append(graph)
-
-    neighbor: int = next(iter(g.neighbors(leaf)))
+    neighbor: int = next(iter(zx_graph.neighbors(leaf)))
     pauli_graphs: list[PauliGraph] = [
-        PauliGraph().add_support(g, (leaf, neighbor), support)
+        PauliGraph().add_support(zx_graph, (leaf, neighbor), support)
         for support in [(True, False), (False, True)]
     ]
 
-    if g.vertex_degree(neighbor) == 1:  # make sure no leaf node will be in the frontier
-        frontier: list[int] = []
-        # if neighbor in bases_at_leaves:
-        #     pauli_graphs = list(
-        #         filter(
-        #             lambda pg: pg[neighbor][leaf] in (bases_at_leaves[neighbor], Pauli.I),
-        #             pauli_graphs,
-        #         )
-        #     )
-    else:
-        frontier: list[int] = [neighbor]
-        explored_leaves: list[int] = [leaf]
+    frontier: list[int] = []
[…]
