### 2025-10-19 4aa814a
correlation surface finding: align with port basis constraints
