### 2025-10-20 1c9afb5
correlation surface finding: reduce pauli graph multiplication overhead
