### 2025-10-20 a301ecb
correlation surface finding: reduce stabilizer length when the current node doesn't lead to new edges
