+++ src/tqec/interop/pyzx/correlation_experimental.py
@@ -15,7 +15,7 @@ from pyzx.pauliweb import PauliWeb, multiply_paulis
 from pyzx.utils import FractionLike, VertexType
 
 from tqec.computation.correlation import CorrelationSurface, ZXEdge, ZXNode
-from tqec.interop.pyzx.utils import is_boundary, is_hardmard, is_s, is_x_no_phase, is_z_no_phase
+from tqec.interop.pyzx.utils import is_boundary, is_hardmard, is_s, is_z_no_phase
 from tqec.utils.enums import Basis
 from tqec.utils.exceptions import TQECError
 
@@ -56,7 +56,7 @@ def pauli_web_to_correlation_surface(
         (u, v), pauli_u = half_edges.popitem()
         pauli_v = half_edges.pop((v, u))
         if pauli_u == "Y":  # pragma: no cover
-            span.add(ZXEdge(ZXNode(u, Basis.X), ZXNode(v, Basis.Z)))
+            span.add(ZXEdge(ZXNode(u, Basis.X), ZXNode(v, Basis.X)))
             span.add(ZXEdge(ZXNode(u, Basis.Z), ZXNode(v, Basis.Z)))
             continue
         span.add(ZXEdge(ZXNode(u, Basis(pauli_u)), ZXNode(v, Basis(pauli_v))))
@@ -127,9 +127,8 @@ class PauliGraph(dict[int, dict[int, tuple[bool, bool]]]):
                 xz_support[::-1] if is_hardmard(zx_graph, edge) else xz_support,
             ),
         ):
-            if u not in self:
-                self[u] = {}
-            self[u][v] = support
+            edges = self.setdefault(u, {})
+            edges[v] = support
         return self
 
     def support_on_nodes(
@@ -183,9 +182,11 @@ def _find_correlation_surfaces_from_leaf(zx_graph: GraphS, leaf: int) -> list[Co
             z_leaves.append(closed_leaf)
         else:
             x_leaves.append(closed_leaf)
+
     pauli_graphs = _find_pauli_graph_generator_set_from_leaf(zx_graph, leaf)
     if sum(len(leaves) for leaves in (x_leaves, y_leaves, z_leaves)) == 0:
         basis_pauli_graphs = pauli_graphs
+
     stabilizer_basis, basis_pauli_graphs, valid_pauli_graphs = {}, [], []
     for pauli_graph in pauli_graphs:
         stabilizer = _bits_to_int(
@@ -209,10 +210,11 @@ def _find_correlation_surfaces_from_leaf(zx_graph: GraphS, leaf: int) -> list[Co
 
 def _find_pauli_graph_generator_set_from_leaf(zx_graph: GraphS, leaf: int) -> list[PauliGraph]:
     """Find the correlation surfaces starting from a leaf node in the graph."""
+    paulis_xz = ((False, True), (True, False))
+    pauli_i = (False, False)
     neighbor: int = next(iter(zx_graph.neighbors(leaf)))
     pauli_graphs: list[PauliGraph] = [
-        PauliGraph().add_support(zx_graph, (leaf, neighbor), support)
-        for support in [(True, False), (False, True)]
+        PauliGraph().add_support(zx_graph, (leaf, neighbor), support) for support in paulis_xz
     ]
 
     frontier: list[int] = []
@@ -222,8 +224,6 @@ def _find_pauli_graph_generator_set_from_leaf(zx_graph: GraphS, leaf: int) -> li
         explored_nodes = {leaf}
 
     while frontier:
-        stabilizer_nodes = explored_leaves + frontier
-        stabilizer_length = sum(len(pauli_graphs[0][n]) for n in stabilizer_nodes)
         current_node = frontier.pop(0)
[…]
