### 2025-10-21 739d29b
correlation surface finding: refactor code to use IntFlag as Pauli; improve readability
