+++ src/tqec/interop/pyzx/correlation_experimental.py
@@ -3,11 +3,12 @@
 from __future__ import annotations
 
 import operator
-from collections.abc import Iterable
+from collections.abc import Callable, Generator, Iterable
 from copy import copy
+from enum import IntFlag
 from fractions import Fraction
-from functools import reduce
-from itertools import chain, pairwise, product
+from functools import partial, reduce
+from itertools import accumulate, chain, pairwise, product, repeat, starmap
 
 import stim
 from pyzx.graph.graph_s import GraphS
@@ -15,7 +16,8 @@ from pyzx.pauliweb import PauliWeb, multiply_paulis
 from pyzx.utils import FractionLike, VertexType
 
 from tqec.computation.correlation import CorrelationSurface, ZXEdge, ZXNode
-from tqec.interop.pyzx.utils import is_boundary, is_hardmard, is_s, is_z_no_phase
+from tqec.interop.pyzx.utils import is_boundary, is_s, is_z_no_phase
+from tqec.interop.pyzx.utils import is_hardmard as is_hadamard
 from tqec.utils.enums import Basis
 from tqec.utils.exceptions import TQECError
 
@@ -113,45 +115,97 @@ def find_correlation_surfaces(
     return sorted(correlation_surfaces, key=lambda x: sorted(x.span))
 
 
-class PauliGraph(dict[int, dict[int, tuple[bool, bool]]]):
+class Pauli(IntFlag):
+    """Pauli operators as bit flags of X and Z supports."""
+
+    I = 0  # noqa: E741
+    X = 1
+    Z = 2
+    Y = X | Z
+
+    def flipped(self) -> Pauli:
+        """Return the Pauli operator with X and Z supports flipped."""
+        return Pauli((self >> 1) | ((self % 2) << 1))
+
+
+PAULIS_XZ = (Pauli.X, Pauli.Z)
+PAULIS_XYZ = (Pauli.X, Pauli.Y, Pauli.Z)
+
+
+class PauliGraph(dict[int, dict[int, Pauli]]):
     """Correlation surface represented as X and Z supports on half-edges."""
 
-    def add_support(
-        self, zx_graph: GraphS, edge: tuple[int, int], xz_support: tuple[bool, bool]
+    def add_pauli_to_edge(
+        self, edge: tuple[int, int], pauli: Pauli, edge_is_hadamard: bool
     ) -> PauliGraph:
-        """Add X and Z supports on the edge (u, v)."""
-        for (u, v), support in zip(
+        """Add Pauli operators to both ends of the given edge."""
+        for (u, v), p in zip(
             (edge, edge[::-1]),
             (
-                xz_support,
-                xz_support[::-1] if is_hardmard(zx_graph, edge) else xz_support,
+                pauli,
+                pauli.flipped() if edge_is_hadamard else pauli,
             ),
         ):
             edges = self.setdefault(u, {})
-            edges[v] = support
+            edges[v] = p
         return self
 
-    def support_on_nodes(
-        self, nodes: list[int], transpose: bool = True
-    ) -> tuple[tuple[bool], tuple[bool]] | tuple[tuple[bool, bool]]:
-        """Get the X and Z supports on the given nodes."""
-        supports = chain.from_iterable(self[n].values() for n in nodes)
-        return tuple(zip(*supports)) if transpose else tuple(supports)
+    def paulis_at_nodes(self, nodes: list[int]) -> list[Pauli]:
+        """Get the Pauli operators at the given nodes."""
+        return list(chain.from_iterable(self[n].values() for n in nodes))
+
+    def signature_at_nodes(
+        self,
+        nodes: list[int],
+        func: Callable[[Pauli], int] | None = None,
+        bit_length: int = 2,
+    ) -> int:
[…]
