### 2025-10-25 2092421
add to workflow and ignore certain domains
