### 2025-10-25 8b24fd0
add working directory to job
