### 2025-10-27 210384a
Workflow to verify broken links in docs (#740)
This PR adds a workflow that verifies if there are broken links in the
docs.

Fixes #739 
Fixes #738
