+++ pull/720/_sources/gallery/index.rst.txt
@@ -1,29 +1,0 @@
-.. _gallery-reference-label:
-
-Examples Gallery
-===================
-
-This section contains a collection of constructed logical computations and their simulation results.
-
-
-Computation Gallery
-===================
-
-.. nbgallery::
-   :caption: Example using blockgraphs created in TQEC.
-
-   memory.ipynb
-   cnot.ipynb
-   move_rotation.ipynb
-   three_cnots.ipynb
-   steane_encoding.ipynb
-
-
-Integration Gallery
-===================
-
-.. nbgallery::
-   :caption: Examples using circuits designed with external frameworks (e.g., Qiskit, Qrisp, etc.).
-
-   qiskit.ipynb
-   qrisp.ipynb
