+++ pull/719/_sources/_autosummary/tqec.utils.rst.txt
@@ -1,40 +1,0 @@
-﻿tqec.utils
-==========
-
-.. currentmodule:: tqec
-
-.. automodule:: tqec.utils
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-      :toctree:
-   
-      Basis
-      ExtendedBasis
-      NoiseModel
-      Orientation
-      PauliBasis
-   
-   
-
-   
-   
-   .. rubric:: Exceptions
-
-   .. autosummary::
-      :toctree:
-   
-      TQECError
-   
-   
