### 2025-10-28 4d89160
correlation surface finding: traverse the graph in the node order (pop the smallest node from the frontier); this is close to the circuit time order if the graph is from pyzx's circuit to graph parser.
