### 2025-10-29 b835657
chore: add pymatching dependency
