+++ pyproject.toml
@@ -43,6 +43,7 @@ dependencies = [
     "pycollada<=0.9.1",
     "pycryptosat<=5.11.21",
     "python-sat[cryptosat]<=1.8.dev24",
+    "pymatching<=2.3.0",
     "pyzx<=0.9.0",
     "scipy<=1.15.3",
     "semver<=3.0.4",
@@ -92,7 +93,6 @@ dev = [
     "pandoc<=2.4",
     "pre-commit<=4.3.0",
     "pydata-sphinx-theme<=0.16.1",
-    "pymatching<=2.3.0",
     "pytest<=8.4.1",
     "pytest-cov<=6.2.1",
     # Ruff version fixed to ensure replicability of results. Ruff being part of pre-commit, we
