### 2025-10-29 059f3ae
Update pyproject.toml
Co-authored-by: Yiming Zhang <61700160+inmzhang@users.noreply.github.com>
