+++ pyproject.toml
@@ -43,7 +43,7 @@ dependencies = [
     "pycollada<=0.9.1",
     "pycryptosat<=5.11.21",
     "python-sat[cryptosat]<=1.8.dev24",
-    "pymatching<=2.3.0",
+    "pymatching<=2.3.1",
     "pyzx<=0.9.0",
     "scipy<=1.15.3",
     "semver<=3.0.4",
