### 2025-10-29 7dc01f0
Reduce CLI CNOT example overhead (#746)
Closes #744. It now finish under 10min on colab when I tested locally.
