### 2025-10-30 af3f20c
correlation surface finding: caching local valid Pauli generation
