+++ src/tqec/interop/pyzx/correlation_test.py
@@ -4,10 +4,9 @@ import pytest
 from pyzx.graph.graph_s import GraphS
 from pyzx.utils import EdgeType, VertexType
 
+from tqec.compile.observables.abstract_observable import _check_correlation_surface_validity
 from tqec.computation.correlation import CorrelationSurface, ZXEdge, ZXNode
-from tqec.interop.pyzx.correlation import (
-    find_correlation_surfaces,
-)
+from tqec.interop.pyzx.correlation_experimental import find_correlation_surfaces
 from tqec.utils.enums import Basis
 
 
@@ -28,6 +27,8 @@ def test_correlation_two_xz_nodes(ty: VertexType) -> None:
     g.add_vertex(ty)
     g.add_edge((0, 1))
     surfaces = find_correlation_surfaces(g)
+    for i, surface in enumerate(surfaces):
+        _check_correlation_surface_validity(surface, g)
     assert len(surfaces) == 1
     b = Basis.Z if ty == VertexType.X else Basis.X
     assert surfaces[0].span == frozenset([ZXEdge(ZXNode(0, b), ZXNode(1, b))])
@@ -47,6 +48,8 @@ def test_correlation_hadamard() -> None:
     g.add_vertex(VertexType.Z)
     g.add_edge((0, 1), EdgeType.HADAMARD)
     surfaces = find_correlation_surfaces(g)
+    for i, surface in enumerate(surfaces):
+        _check_correlation_surface_validity(surface, g)
     assert len(surfaces) == 1
     assert surfaces[0].span == frozenset(
         [
@@ -64,6 +67,8 @@ def test_correlation_y_node() -> None:
     g.add_vertex()
     g.add_edge((0, 1))
     surfaces = find_correlation_surfaces(g)
+    for i, surface in enumerate(surfaces):
+        _check_correlation_surface_validity(surface, g)
     assert len(surfaces) == 1
     assert surfaces[0].span == frozenset(
         [
@@ -86,6 +91,8 @@ def test_correlation_port_passthrough() -> None:
     g.set_type(1, VertexType.X)
 
     surfaces = find_correlation_surfaces(g)
+    for i, surface in enumerate(surfaces):
+        _check_correlation_surface_validity(surface, g)
     assert surfaces == [
         CorrelationSurface(
             frozenset(
@@ -115,30 +122,32 @@ def test_correlation_logical_s_via_gate_teleportation() -> None:
     g.add_vertex(VertexType.Z, phase=Fraction(1, 2))
     g.add_edges([(0, 1), (1, 2), (1, 3), (3, 4)])
     surfaces = set(find_correlation_surfaces(g, reduce_to_minimal_generators=False))
-    assert len(surfaces) == 3
-    assert {
-        CorrelationSurface(
-            frozenset(
-                {
-                    ZXEdge(ZXNode(0, Basis.X), ZXNode(1, Basis.X)),
-                    ZXEdge(ZXNode(0, Basis.Z), ZXNode(1, Basis.Z)),
-                    ZXEdge(ZXNode(1, Basis.X), ZXNode(2, Basis.X)),
-                    ZXEdge(ZXNode(1, Basis.X), ZXNode(3, Basis.X)),
-                    ZXEdge(ZXNode(1, Basis.Z), ZXNode(3, Basis.Z)),
-                    ZXEdge(ZXNode(3, Basis.X), ZXNode(4, Basis.X)),
-                    ZXEdge(ZXNode(3, Basis.Z), ZXNode(4, Basis.Z)),
-                }
-            )
-        ),
-        CorrelationSurface(
-            frozenset(
-                {
[…]
