### 2025-10-31 c100dad
correlation surface finding: lazy generation of pauli graphs
