+++ pyproject.toml
@@ -94,7 +94,7 @@ dev = [
     "pre-commit<=4.3.0",
     "pydata-sphinx-theme<=0.16.1",
     "pytest<=8.4.1",
-    "pytest-cov<=6.2.1",
+    "pytest-cov<=7.0.0",
     # Ruff version fixed to ensure replicability of results. Ruff being part of pre-commit, we
     # cannot afford someone installing a different version of ruff and being forbidden to commit
     # anything because the different version sees some errors in the code.
