### 2025-11-04 92476ed
xor cube top observable qubits together
