### 2025-11-04 8a0d515
correlation surface finding: attempt at solving and piecing together subgraphs from vertex partitions and disconnected components
