+++ src/tqec/interop/pyzx/correlation_experimental.py
@@ -2,8 +2,17 @@
 
 from __future__ import annotations
 
+import multiprocessing
 import operator
-from collections.abc import Callable, Generator, Iterable
+from collections import ChainMap
+from collections.abc import (
+    Callable,
+    Generator,
+    Iterable,
+    Iterator,
+    MutableMapping,
+    Sequence,
+)
 from copy import copy
 from enum import IntFlag
 from fractions import Fraction
@@ -65,8 +74,11 @@ def pauli_web_to_correlation_surface(
     return CorrelationSurface(frozenset(span))
 
 
-def find_correlation_surfaces(
-    g: GraphS, reduce_to_minimal_generators: bool = True
+def find_correlation_surfaces(  # noqa: D417
+    g: GraphS,
+    reduce_to_minimal_generators: bool = True,
+    vertex_ordering: Sequence[set[int]] | None = None,
+    parallel: bool = False,
 ) -> list[CorrelationSurface]:
     """Find the correlation surfaces in a ZX graph.
 
@@ -106,7 +118,10 @@ def find_correlation_surfaces(
         raise TQECError(
             "The graph must contain at least one leaf node to find correlation surfaces."
         )
-    correlation_surfaces = _find_correlation_surfaces_from_leaf(g, min(leaves))
+    correlation_surfaces = [
+        cs.to_correlation_surface(g)
+        for cs in _find_pauli_graphs_with_vertex_ordering(g, vertex_ordering, parallel)
+    ]
 
     # if CorrelationSurface(frozenset()) in correlation_surfaces:
     #     correlation_surfaces.remove(CorrelationSurface(frozenset()))
@@ -132,12 +147,12 @@ PAULIS_XZ = (Pauli.X, Pauli.Z)
 PAULIS_XYZ = (Pauli.X, Pauli.Y, Pauli.Z)
 
 
-class PauliGraph(dict[int, dict[int, Pauli]]):
+class PauliGraphBase(MutableMapping[int, dict[int, Pauli]]):
     """Correlation surface represented as X and Z supports on half-edges."""
 
     def add_pauli_to_edge(
         self, edge: tuple[int, int], pauli: Pauli, edge_is_hadamard: bool
-    ) -> PauliGraph:
+    ) -> PauliGraphBase:
         """Add Pauli operators to both ends of the given edge."""
         for (u, v), p in zip(
             (edge, edge[::-1]),
@@ -150,13 +165,13 @@ class PauliGraph(dict[int, dict[int, Pauli]]):
             edges[v] = p
         return self
 
-    def paulis_at_nodes(self, nodes: list[int]) -> Iterable[Pauli]:
+    def paulis_at_nodes(self, nodes: Iterable[int]) -> Iterable[Pauli]:
         """Get the Pauli operators at the given nodes."""
         return chain.from_iterable(self[n].values() for n in nodes)
 
     def signature_at_nodes(
         self,
-        nodes: list[int],
+        nodes: Iterable[int],
         func: Callable[[Pauli], int] | None = None,
         bit_length: int = 2,
     ) -> int:
@@ -209,7 +224,27 @@ class PauliGraph(dict[int, dict[int, Pauli]]):
         return CorrelationSurface(frozenset(span))
 
 
-def _multiply_pauli_graphs(pauli_graphs: list[PauliGraph]) -> PauliGraph:
+class PauliGraph(dict[int, dict[int, Pauli]], PauliGraphBase):
+    pass
+
+
+class PauliGraphView(ChainMap[int, dict[int, Pauli]], PauliGraphBase):
+    def __setitem__(self, key, value):
+        for mapping in self.maps:
+            if key in mapping:
+                mapping[key] = value
+                return
+        self.maps[0][key] = value
+
[…]
