### 2025-11-07 58df0cb
🩹 never upper cap Python
From Scientific Python:
You should never upper cap your Python requirement. This is rarely correct, and tools like pip do not handle this properly even if it is correct. This field is used to back-solve. If you want to specify versions you've tested on, use classifiers. If you want to add a custom error message, add a build-type and/or runtime assert.
