### 2025-11-07 b052976
🔥 remove unused setuptools configuration
