### 2025-11-07 7166e7a
🔧 ruff infers its target-version from requires-python
