### 2025-11-07 17ad4e5
🔥 no need to run `pyupgrade` with ruff present and `UP` rules enabled
