+++ .pre-commit-config.yaml
@@ -5,10 +5,6 @@ repos:
       - id: check-yaml
       - id: end-of-file-fixer
       - id: trailing-whitespace
-  - repo: https://github.com/asottile/pyupgrade
-    rev: v3.19.0
-    hooks:
-      - id: pyupgrade
   - repo: https://github.com/crate-ci/typos
     rev: v1.26.1
     hooks:
