### 2025-11-07 2af9799
🔥 remove pyright and mypy configuration
neither pyright nor mypy are run in the repository
