### 2025-11-07 8f8108f
📄 use PEP639 for specifying license
