### 2025-11-07 tqec/tqec#751 @burgholzer
Proper Python dependency configuration
## Is your feature request related to a problem? Please describe.

The current dependency setup in the project seems to be less than ideal as it places upper bounds on all dependencies.
Please consider reading https://iscinumpy.dev/post/bound-version-constraints/ in that regard. In almost all instances, upper caps are considered a bad practice.

To make matters a little more extreme, none of the packages have lower bounds specified. As such, people might be running fully outdated versions of dependencies that might, by any chance, install and yield a completely broken configuration.

## Describe the solution you'd like

I assume the main goal here is to get reproducible environments.
However, in that case, the proper tool for that would be a lock file.
As you are using `uv` already, adopting that is as straight-forward as it gets.
Simply run `uv lock` and commit the respective lock file.
Commands like `uv sync` or `uv pip install` will then use the known-good versions from the lockfile.
Services like dependabot or renovate can regularly update the lockfile for you.

As for the lower bound problem: `uv` also offers great tooling support for this through its `--resolution-strategy=lowest` flag, which automatically sets up an environment with the minimum versions it can resolve according to the project dependencies.

If you want some inspiration for how this could look like (also in CI), have a look at https://github.com/munich-quantum-toolkit/core/blob/98c31eb50cc6e84b676b0545237aee4ce21a6339/noxfile.py#L120-L130 and https://github.com/munich-quantum-toolkit/workflows/blob/main/.github/workflows/reusable-python-tests.yml

## Describe alternatives you've considered

N/A

## Additional context

This is part of https://github.com/openjournals/joss-reviews/issues/9142
