### 2025-11-07 tqec/tqec#752 @burgholzer
CD workflow and release on PyPI
## Is your feature request related to a problem? Please describe.

The repository is currently missing a CD workflow for automatic releases and has no official release available from PyPI.
The fact that no official release is available and the version number is still at `v0.0.1` makes it hard to communicate changes. Even though the project might not be ready for a 1.0 release, it would certainly deserve a `0.1.0` release with subsequent updates when new developments roll in.

## Describe the solution you'd like

A GitHub workflow triggered on GitHub releases that builds the source distribution as well as the wheel and uploads them to PyPI (preferably using trusted publishing).
Some information on trusted publishing https://docs.pypi.org/trusted-publishers/

You could take https://github.com/munich-quantum-toolkit/bench/blob/main/.github/workflows/cd.yml as an inspiration, which internally uses
- https://github.com/munich-quantum-toolkit/workflows/blob/main/.github/workflows/reusable-python-packaging-sdist.yml
- https://github.com/munich-quantum-toolkit/workflows/blob/main/.github/workflows/reusable-python-packaging-wheel-build.yml

Afterwards, an official release should be published.

## Describe alternatives you've considered

Obviously, manual releases may be an option here, but I'd personally recommend going with the automatic workflow which also provides automatic build attestations assuring the integrity of the distributed assets.

## Additional context

This is part of https://github.com/openjournals/joss-reviews/issues/9142
