### 2025-11-07 tqec/tqec#753 @burgholzer
Move tests out of source tree
## Is your feature request related to a problem? Please describe.

Currently, all tests are incorporated into the `src` tree of the Python package.
This is generally not considered good practice as users of the library do not need the test code, which is also why test code is typically not shipped as part of the source distribution our binary wheels. It just bloats these artefacts.

See https://learn.scientific-python.org/development/tutorials/test/ and https://learn.scientific-python.org/development/guides/pytest/ for more context on this topic.

## Describe the solution you'd like

Move the tests to a dedicated `tests` directory.
The respective directory structure may mirror the `src` structure.
https://learn.scientific-python.org/development/guides/pytest/#configuring-pytest might be a good reference for the respective `pyproject.toml` configuration.

## Describe alternatives you've considered

N/A

## Additional context

This is part of https://github.com/openjournals/joss-reviews/issues/9142
