### 2025-11-07 #9142 @burgholzer
Hey @BSchelpe 👋🏼
sorry that it took a while to get to reviewing this submission.
I have been following the TQEC activities over the last three years from the sidelines.
Great to finally get the chance to check it out in more detail.

@lockwo
I am mostly done with my review. Opened a small PR and three issues with things that I would like to see addressed before recommending this for publication.

One last thing that I noticed is that the requirements for a JOSS paper explicitly ask for
> State of the field: Do the authors describe how this software compares to other commonly-used packages?

This is not covered in the paper at the moment and would probably deserve a paragraph or two.

That being said; great stuff! Enjoyed checking out the repository.
