### 2025-11-08 91e05ac
JOSS review fixes (#750)
Hey everyone 👋🏼 

While doing my review as part of the ongoing JOSS submission
https://github.com/openjournals/joss-reviews/issues/9142 I accumulated
some of the smaller things to fix in small, digestible commits.

They mostly declutter existing configuration files and adopt one or two
best practices from
https://learn.scientific-python.org/development/guides/repo-review/

I'll post a couple of follow-up issues separately.
