+++ pyproject.toml
@@ -10,9 +10,10 @@ authors = [
 ]
 description = "A design automation framework for Topological Quantum Error Correction."
 readme = "README.md"
-license = { file = "LICENSE" }
+license = "Apache-2.0"
+license-files = ["LICENSE"]
 keywords = ["topological quantum error correction", "qec"]
-requires-python = ">= 3.10, <3.14"
+requires-python = ">=3.10"
 classifiers = [
   "Development Status :: 4 - Beta",
   "Intended Audience :: Developers",
@@ -116,78 +117,6 @@ all = [
 [project.scripts]
 tqec = "tqec._cli.tqec:main"
 
-[tool.setuptools]
-include-package-data = true
-
-[tool.setuptools.packages.find]
-where = ["src"]
-exclude = ["*test.py"]
-
-[tool.setuptools.package-data]
-tqec = ["py.typed"]
-# See https://setuptools.pypa.io/en/latest/userguide/datafiles.html#subdirectory-for-data-files
-"tqec.gallery.dae" = ["*.dae"]
-
-[tool.setuptools.dynamic]
-dependencies = { file = "requirements.txt" }
-
-
-[tool.mypy]
-# See https://numpy.org/devdocs/reference/typing.html
-plugins = "numpy.typing.mypy_plugin"
-# Start off with these
-warn_unused_configs = true
-warn_redundant_casts = true
-# Explicitly disable that for the moment, will eventually be turned on again.
-warn_unused_ignores = false
-warn_unreachable = true
-
-# Getting these passing should be easy
-strict_equality = true
-extra_checks = true
-
-# Strongly recommend enabling this one as soon as you can
-check_untyped_defs = true
-
-# These shouldn't be too much additional work, but may be tricky to
-# get passing if you use a lot of untyped libraries
-disallow_subclassing_any = true
-disallow_untyped_decorators = true
-disallow_any_generics = true
-
-# These next few are various gradations of forcing use of type annotations
-disallow_untyped_calls = true
-disallow_incomplete_defs = true
-disallow_untyped_defs = true
-
-# This one isn't too hard to get passing, but return on investment is lower
-# Disabled due to issues with cirq not reexporting explicitly.
-no_implicit_reexport = true
-
-# This one can be tricky to get passing if you use a lot of untyped libraries
-warn_return_any = true
-
-
-# See https://mypy.readthedocs.io/en/stable/config_file.html#using-a-pyproject-toml
-[[tool.mypy.overrides]]
-module = [
-  "stim",
-  "sinter",
-  "pysat",
-  "pysat.solvers",
-  "collada",
-  "collada.source",
-  "mpl_toolkits.mplot3d.axes3d",
-]
-ignore_missing_imports = true
-
-[[tool.mypy.overrides]]
-module = "pyzx"
-no_implicit_reexport = false
-
-[[tool.mypy.overrides]]
-module = ["tqec.interop.pyzx.*"]
-disallow_untyped_calls = false
 
 [tool.coverage.run]
 omit = ["*/*_test.py", "*/simulation/*", "*/visualisation/*"]
@@ -209,53 +138,9 @@ exclude_also = [
 ]
 
 
-[tool.pyright]
-# Paths of directories or files that should be considered part of the project.
-# If no paths are specified, pyright defaults to the directory that contains the
-# config file. Paths may contain wildcard characters ** (a directory or multiple
-# levels of directories), * (a sequence of zero or more characters), or ? (a
-# single character). If no include paths are specified, the root path for the
-# workspace is assumed.
-include = ["src"]
-
-# Paths of directories or files that should not be considered part of the project.
-# These override the directories and files that include matched, allowing specific
-# subdirectories to be excluded. Note that files in the exclude paths may still be
[…]
