### 2025-11-08 tqec/tqec#752 @purva-thakre
> The repository is currently missing a CD workflow for automatic releases and has no official release available from PyPI.

If I recall older discussions correctly, we plan on releasing an official version sometime next year. We were procrastinating adding this workflow until then. 😅
