### 2025-11-08 tqec/tqec#752 @burgholzer
> > The repository is currently missing a CD workflow for automatic releases and has no official release available from PyPI.
>
> If I recall older discussions correctly, we plan on releasing an official version sometime next year. We were procrastinating adding this workflow until then. 😅

Is there any compelling reason for not starting to create releases. I totally get procrastinating a 1.0 release, but something like development releases would certainly rather help than hurt in my opinion.
Especially since no available releases make it unnecessarily hard for other projects to build on the TQEC library. From experience, people have a hard time adding git dependencies to their Python projects.
