### 2025-11-08 bb3202e
Move tests out of source tree into dedicated tests/ directory (#754)
Tests were embedded in `src/tqec/` alongside production code, causing
them to be included in distribution artifacts. This bloats packages and
violates scientific Python conventions.

## Changes

- **Moved 66 test files** from `src/tqec/` to `tests/` with mirrored
directory structure
- **Updated `pyproject.toml`**:
  - Added `testpaths = ["tests"]` to pytest configuration
  - Changed coverage omit pattern from `*/*_test.py` to `tests/*`
- **Fixed cross-test import** in `tests/computation/block_graph_test.py`
to reference helpe
[…]
