+++ tests/computation/block_graph_test.py
@@ -3,11 +3,11 @@ import tempfile
 
 import pytest
 
+from tests.interop.collada.read_write_test import rotated_cnot
 from tqec.computation.block_graph import BlockGraph
 from tqec.computation.cube import ZXCube
 from tqec.computation.pipe import PipeKind
 from tqec.gallery import cnot, memory
-from tqec.interop.collada.read_write_test import rotated_cnot
 from tqec.utils.enums import Basis
 from tqec.utils.exceptions import TQECError
 from tqec.utils.position import Direction3D, Position3D
