### 2025-11-09 tqec/tqec#752 @purva-thakre
> something like development releases would certainly rather help than hurt

To be honest, we didn't think about a dev release.  But I agree with you. This workflow should be added.
