### 2025-11-09 bc034bb
Add test script
