### 2025-11-10 c092f9a
Add publish workflow
