### 2025-11-10 6483ac7
Add CI for testing minimum dependency versions
