### 2025-11-10 5179077
Revert "feat: reduce branches for passthrough node when finding corre… (#756)
…lation surfaces (#704)"

This reverts commit 9b1294ac32def8bdb637423e495c3ff298c09a90.


--- 

As identified by @inmzhang , this change is incorrect. Let's revert it
and wait for #707 to be merged
