### 2025-11-10 73c3375
Fix dependency version configuration: remove most upper bound and set lower bounds
