### 2025-11-10 f8fa8d9
Add CD workflow publishing to PYPI (#755)
Related to #752. Once #751 is resolved, we can hopefully release a
v0.1.0 version of tqec.
