### 2025-11-10 958a3f8
Fix publish workflow: should only trigger when pushing tags
