+++ .github/workflows/publish.yml
@@ -7,8 +7,6 @@ name: Publish Python distribution to PyPI
 
 on:
   push:
-    branches:
-      - main
     tags:
       - "v*"
 
