+++ pyproject.toml
@@ -65,7 +65,7 @@ Issues = "https://github.com/tqec/tqec/issues"
 
 [dependency-groups]
 test = [
-  "pytest<=8.4.1",
+  "pytest<=9.0.0",
   "pytest-cov<=7.0.0",
   "types-networkx<=3.5.0.20250610",
   "scipy-stubs<=1.16.0.0",
@@ -94,7 +94,7 @@ dev = [
     "pandoc<=2.4",
     "pre-commit<=4.3.0",
     "pydata-sphinx-theme<=0.16.1",
-    "pytest<=8.4.1",
+    "pytest<=9.0.0",
     "pytest-cov<=7.0.0",
     # Ruff version fixed to ensure replicability of results. Ruff being part of pre-commit, we
     # cannot afford someone installing a different version of ruff and being forbidden to commit
