### 2025-11-11 4f5642a
Update .github/workflows/test-min-versions.yml
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>
