+++ .github/workflows/test-min-versions.yml
@@ -42,7 +42,6 @@ jobs:
 
       - name: Install with minimum dependency versions
         run: |
-          uv venv
           uv sync --group test --resolution=lowest-direct
 
       - name: Run tests with minimum versions
