### 2025-11-11 f1631f5
Skip python setup in workflow
