### 2025-11-11 f62e533
Simplify Python setup in workflows
