+++ .github/workflows/gh-pages.yml
@@ -24,16 +24,11 @@ jobs:
         uses: astral-sh/setup-uv@v6
         with:
           enable-cache: true
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: "3.13"
-      - name: Install dependencies
+      - name: Setup Python development environment
         run: |
-          uv venv
           sudo apt-get update
           sudo apt-get install -y --no-install-recommends pandoc
-          uv sync --group all
+          uv sync --group all --python "3.13"
       - uses: actions/cache@v4
         id: cache-docs
         with:
