### 2025-11-11 c10f272
Python dependency configuration (#757)
Closes #751.

This PR removes most upper bounds on dependency versions and adds
reasonable lower bounds. It also introduces a CI workflow to verify that
the minimum dependency versions are satisfied when `pyproject.toml`
changes. I don’t have a systematic way to determine the exact minimum
versions yet, but I confirmed that the current configuration works
across all supported Python versions (3.10–3.13).

@burgholzer Hi Lukas, does this change make sense to you?

---------

Co-authored-by: Yilun Zhao <40353317+Zhaoyilunnn@users.noreply.github.com>
Co-auth
[…]
