+++ .github/workflows/exotic-oses.yml
@@ -21,17 +21,12 @@ jobs:
         uses: astral-sh/setup-uv@v6
         with:
           enable-cache: true
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
+      - name: Setup Python development environment
+        run: |
           # Use smallest supported Python version as the common denominator.
           # In theory, if everything succeed in 3.10, everything should succeed in
           # 3.11, 3.12, 3.13 and follow-up version.
-          python-version: "3.10"
-      - name: Install dependencies
-        run: |
-          uv venv
-          uv sync --group all
+          uv sync --group all --python "3.10"
       # Perform type checking before testing because that is quicker.
       - name: Type checking
         run: uv run ty check
