### 2025-11-11 0997ac8
chore: Configure Renovate (#761)
Welcome to [Renovate](https://redirect.github.com/renovatebot/renovate)!
This is an onboarding PR to help you understand and configure settings
before regular Pull Requests begin.

🚦 To activate Renovate, merge this Pull Request. To disable Renovate,
simply close this Pull Request unmerged.



---
### Detected Package Files

 * `.github/workflows/ci.yml` (github-actions)
 * `.github/workflows/exotic-oses.yml` (github-actions)
 * `.github/workflows/gh-pages.yml` (github-actions)
 * `.github/workflows/pr-close.yml` (github-actions)
 * `.github/workflows/publish.y
[…]
