### 2025-11-11 7a60d73
chore(deps): update dependency ruff to v0.14.4
Signed-off-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
