### 2025-11-11 673fe22
fix(deps): update dependency numpy to >=2.3,<3
Signed-off-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
