+++ pyproject.toml
@@ -39,7 +39,7 @@ classifiers = [
 dependencies = [
   "matplotlib>=3.9",
   "networkx>=3.0",
-  "numpy>=1.22,<2.3",                                  # Upper bound: https://github.com/tqec/tqec/pull/659
+  "numpy>=2.3,<3",                                  # Upper bound: https://github.com/tqec/tqec/pull/659
   "numpy>=1.24; python_version >= '3.11'",
   "numpy>=1.26; python_version >= '3.12'",
   "numpy>=2.1; python_version >= '3.13'",
