### 2025-11-12 tqec/tqec#752 @burgholzer
> Hi Lukas [@burgholzer](https://github.com/burgholzer), we’ve published the `v0.1.0` release on PyPI. Do you think we can close this issue now?

Yes, definitely! 👍🏼
