### 2025-11-24 #9142 @lockwo
> All of the github issues @burgholzer raised have been resolved. There is just the outstanding change to the paper (adding comparisons), which we are currently working on.

Great, thank you. Feel free to ping me when you think these are complete.
