### 2026-01-07 #9142 @BSchelpe
Hi @lockwo, I'm afraid in terms of paper changes we are still where we were on the 19th Nov, ie working on the last change @burgholzer requested (adding a paragraph to the paper of comparisons). I will try to get an ETA on that and come back to you.
