### 2026-02-20 #9142 @BSchelpe
Hi @lockwo, the last of @burgholzer's comments has been addressed now: the comparison to other packages has been added to paper.md in the repo (under 'state of the field' heading). Thank you very much for your patience!
