### 2026-02-23 #9142 @lockwo
Just a note, but you can get the usual parenthetical citations with [@ cite] (like \cite{} in latex) and the @ cite in markdown will become like the \citeauthor{} in latex.
