### 2026-02-25 #9142 @lockwo
A few notes on the paper text:
- United States vs USA in affiliations (either is fine, consistency is what matters)
- "the result is a `Stim` circuit" (can cite Stim here)
- "Building the logical Stim circuits is a hassle" I might suggest something more along the lines of "Building the logical `Stim` circuits can be a complex [or time consuming] process"
- "difficult to perform exactly, @gidney:2025" extra comma here?
- The first paragraph is a bit unclear, in the summary it seems as though tqec is a framework to go from logical designs -> stim circuits, but "the full scale simulations performed by tqec" (contrasting it with the MC of stim) makes the simulation seem an important feature too
- Extra space before period in "@kan:2025 ."
- The Densely packed quantum memory citation looks like it needs this DOI (10.1109/qce65121.2025.10483).
- "clifford + t lattice surgery" Clifford and T should be capitalized (with {C}lifford etc)
- linking this from above https://github.com/openjournals/joss-reviews/issues/9142#issuecomment-3942891937

Following these point could you:
