### 2026-02-25 #9142 @HaoTy
> * Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

Would it be possible to add authors (with nontrivial contributions) with the revision?
