### 2026-02-26 #9142 @lockwo
If things need to be adjusted now they can, it's just important that the JOSS paper and the archive data match (so the authors of both are the same). Once the archive/paper are released, they are then static (although of course, future releases might have new authors).
