### 2026-02-27 #9142 @HaoTy
> If things need to be adjusted now they can, it's just important that the JOSS paper and the archive data match (so the authors of both are the same). Once the archive/paper are released, they are then static (although of course, future releases might have new authors).

Got it, thanks!
