### 2026-03-04 #9142 @BSchelpe
Hi @lockwo, we've made the paper text changes you suggested now. There was just one point we weren't sure about: "clifford + t lattice surgery" Clifford and T should be capitalized (with {C}lifford etc). We've capitalised to Clifford + T lattice surgery, but we haven't put brackets around the C, because we weren't quite sure why you wanted them - I'm not sure why that capital letter should be treated differently from other capital letters. But we wanted to check we weren't missing something?

We are ready to do the tagged release of the software onwards steps, but we thought best to hold off until @alexandrupaler has done his review in case he suggests changes to the code.
