### 2026-03-05 #9142 @lockwo
> I'm not sure why that capital letter should be treated differently from other capital letters

Ah no, it's just a force of habit (many bib compilers, e.g. for LaTeX, don't respect capitalization unless you do {}), but as long as the output is capitalized that's what matters.

> We are ready to do the tagged release of the software onwards steps, but we thought best to hold off until @alexandrupaler has done his review in case he suggests changes to the code.

I can reach out to them, I wasn't too concerned since for JOSS we require 2 reviewers (and 3 is good, because if one drops, then we still have the required 2) so we aren't technically bottlenecked by them (but if you want their input we can also wait that's not a problem).
