### 2026-03-31 #9142 @purva-thakre
Hi @lockwo, I am one of the authors of this paper.

Just wanted to let you know that we created a tagged release as well as completed the Zenodo step as stated in https://github.com/openjournals/joss-reviews/issues/9142#issuecomment-3947743934.

- The DOI for the Zenodo archive upload is https://doi.org/10.5281/zenodo.18927232
- The version number for the new tagged release of `tqec` is `0.2.0`: https://github.com/tqec/tqec/releases/tag/v0.2.0
