### 2026-04-01 #9142 @lockwo
@BSchelpe - please let me know if you see any issues with the proof that this generates
