### 2026-04-01 #9142 @BSchelpe
> [@BSchelpe](https://github.com/BSchelpe) - please let me know if you see any issues with the proof that this generates

All looks good to me, thank you!
