### 2026-04-01 #9142 @danielskatz
👋 @BSchelpe - as track editor, I've proofread the paper, and suggest the changes in https://github.com/tqec/tqec/pull/890 - please merge this, or let me know what you disagree with
