### 2026-04-01 #9142 @danielskatz
Congratulations to @BSchelpe (Arabella Schelpe) and co-authors on your JOSS publication!!

JOSS fully depends on volunteers - we couldn't do this without you
